distrDoc-package(distrDoc)	R Documentation

distrDoc – documentation for packages distr distrEx distrSim distrTEst

Description

distrDoc provides documentation in form of a common vignette to packages distr, distrEx, distrSim, distrTEst.

Details

Package: 	 distrDoc

Version: 	 1.9

Date: 	 2007-07-29

Depends: 	 R(>= 2.2.0), tools, distr(>= 1.8.0), startupmsg

Suggests: 	 distrEx(>= 0.4-4), distrSim(>= 1.8.0), distrTEst(>= 1.8.0), Biobase

SaveImage: 	 no

LazyLoad: 	 yes

License: 	 GPL (version 2 or later)

URL: 	 http://www.uni-bayreuth.de/departments/math/org/mathe7/DISTR/

Acknowledgement

The idea to split off a common vignette for a set of packages in a particular documentation package is due to Seth Falcon.

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Matthias Kohl Matthias.Kohl@stamats.de

Maintainer: Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

See Also

see mails "[Rd] package vignettes, suggests and R CMD check" on r-devel by Seth Falcon,
https://stat.ethz.ch/pipermail/r-devel/2006-November/043702.html,
https://stat.ethz.ch/pipermail/r-devel/2006-November/043705.html,
and by the author,
https://stat.ethz.ch/pipermail/r-devel/2006-November/043700.html,
https://stat.ethz.ch/pipermail/r-devel/2006-November/043703.html

Examples


## Not run: 
## Open the distr vignette
     vignette("distr")
## End(Not run)

[Package distrDoc version 1.9 Index]