distrEllipse-package(distrEllipse)	R Documentation

distrEllipse – S4 classes for elliptically contoured distributions

Description

distrEllipse provides infrastructure / (S4-)classes for elliptically contoured distributions (based on package distr).

Details

Package: 	 distrEllipse

Version: 	 0.2

Date: 	 2009-09-04

Depends: 	 R(>= 2.8.0), methods, graphics, mvtnorm, setRNG(>= 2006.2-1), distr(>= 2.2), distrEx(>= 2.2), distrSim(>= 2.2), startupmsg

LazyLoad: 	 yes

License: 	 LGPL-3

URL: 	 http://distr.r-forge.r-project.org/

Classes


###################################
Distribution Classes
###################################
[*]: there is a generating function with the same name


"Distribution" (from distr)
|>"MultivariateDistribution" (from distrEx)
|>|>"MultivarMixingDistribution" [*]
|>|>"SphericalDistribution" [*]
|>|>|>|>"EllipticalDistribution" [*]
|>|>|>|>"MVNormDistribution" [*]

"DistrList" (from distr)
|>"MultivarDistrList" [*/class union of "MVDistrList", "UnivarDistrList"]
|>|>"MVDistrList"
|>"UnivarDistrList" (from distr) [*]

Methods


plot-methods            Methods for Function plot
                       (for SphericalDistribution)
show-methods            Methods for Function show
                       (for Simulation/Contsimulation)

Functions


distrEllipseoptions            Functions to change the global variables of the
                        package 'distrEllipse'

Slot accessors / -replacement functions

All slots are inspected / modified by corresponding accessors / -replacement functions.

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

Package versions

Note: The first two numbers of package versions do not necessarily reflect package-individual development, but rather are chosen for the distrXXX family as a whole in order to ease updating "depends" information.

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Note

Global options controlling the plots and summaries of Dataclass and Simulation/Contsimulation objects may be inspected / set by () and ().

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Maintainer: Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

P. Ruckdeschel, M. Kohl, T. Stabla, F. Camphausen (2006): S4 Classes for Distributions, R News, 6(2), 2-6. http://CRAN.R-project.org/doc/Rnews/Rnews_2006-2.pdf

A vignette for packages distr, distrSim, distrTEst, distrEx, distrTeach, distrMod, and distrEllipse is included into the mere documentation package distrDoc and may be called by require("distrDoc");vignette("distr").

A homepage to this package is available under
http://distr.r-forge.r-project.org/.

[Package distrEllipse version 2.2 Index]