distrEx-package(distrEx)	R Documentation

distrEx – Extensions of package distr

Description

distrEx proides some extensions of package distr:

extreme value distribution classes,
expectations in the form

E(X) for the expectation of a distribution object X
E(X,f) for the expectation of f(X) where X is some distribution object and f some function in X

further functionals: var, sd, IQR, mad, median, skewness, kurtosis
truncated moments,
distances between distributions (Hellinger, Cramer von Mises, Kolmogorov, total variation, "convex contamination")
lists of distributions,
conditional distributions in factorized form
conditional expectations in factorized form

Details

Package: 	 distrEx

Version: 	 2.1

Date: 	 2008-11-27

Depends: 	 R(>= 2.6.0), methods, distr(>= 2.0), evd, startupmsg

LazyLoad: 	 yes

License: 	 LGPL-3

URL: 	 http://distr.r-forge.r-project.org/

Classes



Distribution Classes

"Distribution" (from distr)
|>"UnivariateDistribution" (from distr)
|>|>"AbscontDistribution" (from distr)
|>|>|>"Gumbel"
|>|>|>"Pareto"
|>"MultivariateDistribution"
|>|>"DiscreteMVDistribution-class"
|>"UnivariateCondDistribution"
|>|>"AbscontCondDistribution"
|>|>|>"PrognCondDistribution"
|>|>"DiscreteCondDistribution"


Condition Classes

"Condition"
|>"EuclCondition"
|>"PrognCondition"

Parameter Classes

"OptionalParameter" (from distr)
|>"Parameter" (from distr)
|>|>"LMParameter"
|>|>"GumbelParameter"
|>|>"ParetoParameter"

Functions



Integration:
GLIntegrate             Gauss-Legendre quadrature
distrExIntegrate        Integration of one-dimensional functions

Options:
distrExOptions          Function to change the global variables of the
                        package 'distrEx'
Standardization:
make01                  Centering and standardization of univariate
                        distributions


Generating Functions



Distribution Classes
ConvexContamination     Generic function for generating convex
                        contaminations
DiscreteMVDistribution
                        Generating function for
                        DiscreteMVDistribution-class
Gumbel                  Generating function for Gumbel-class
LMCondDistribution      Generating function for the conditional
                        distribution of a linear regression model.

Condition Classes
EuclCondition           Generating function for EuclCondition-class

Parameter Classes
LMParameter             Generating function for LMParameter-class


Methods


Distances:
ContaminationSize       Generic function for the computation of the
                        convex contamination (Pseudo-)distance of two
                        distributions
HellingerDist           Generic function for the computation of the
                        Hellinger distance of two distributions
KolmogorovDist          Generic function for the computation of the
                        Kolmogorov distance of two distributions
TotalVarDist            Generic function for the computation of the
                        total variation distance of two distributions
AsymTotalVarDist        Generic function for the computation of the
                        asymmetric total variation distance of two distributions
                        (for given ratio rho of negative to positive part of deviation)
OAsymTotalVarDist       Generic function for the computation of the minimal (in rho)
                        asymmetric total variation distance of two distributions
vonMisesDist            Generic function for the computation of the
                        von Mises distance of two distributions

liesInSupport           Generic function for testing the support of a
                        distribution


Functionals:
E                       Generic function for the computation of
                        (conditional) expectations
var                     Generic functions for the computation of
                        functionals
IQR                     Generic functions for the computation of
                        functionals
sd                      Generic functions for the computation of
                        functionals
mad                     Generic functions for the computation of
                        functionals
median                  Generic functions for the computation of
                        functionals
skewness                Generic functions for the computation of
                        functionals
kurtosis                Generic functions for the computation of
                        Functionals


truncated Moments:
m1df                    Generic function for the computation of clipped
                        first moments
m2df                    Generic function for the computation of clipped
                        second moments


Demos

Demos are available — see demo(package="distrEx").

Acknowledgement

G. Jay Kerns, gkerns@ysu.edu, has provided a major contribution, in particular the functionals skewness and kurtosis are due to him. Natalyia Horbenko, natalyia.horbenko@itwm.fraunhofer.de has ported the actuar code for the Pareto distribution to this setup.

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Note

Some functions of package stats have intentionally been masked, but completely retain their functionality — see distrExMASK().

If any of the packages e1071, moments, fBasics is to be used together with distrEx the latter must be attached after any of the first mentioned. Otherwise kurtosis() and skewness() defined as methods in distrEx may get masked.
To re-mask, you may use kurtosis <- distrEx::kurtosis; skewness <- distrEx::skewness. See also distrExMASK()

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de and
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,

Maintainer: Matthias Kohl Matthias.Kohl@stamats.de

References

P. Ruckdeschel, M. Kohl, T. Stabla, F. Camphausen (2006): S4 Classes for Distributions, R News, 6(2), 2-6. http://CRAN.R-project.org/doc/Rnews/Rnews_2006-2.pdf

a vignette for packages distr, distrSim, distrTEst, and distrEx is included into the mere documentation package distrDoc and may be called by require("distrDoc");vignette("distr")

a homepage to this package is available under
http://distr.r-forge.r-project.org/

M. Kohl (2005): Numerical Contributions to the Asymptotic Theory of Robustness. PhD Thesis. Bayreuth. Available as http://www.stamats.de/ThesisMKohl.pdf

See Also

distr

[Package distrEx version 2.1 Index]