distrMod-package(distrMod)	R Documentation

distrMod – probability models

Description

Based on the packages distr and distrEx package distrMod provides a flexible framework which allows computation of estimators like maximum likelihood or minimum distance estimators for probability models.

Details

Package: 	 distrMod

Version: 	 2.0.3

Date: 	 2008-11-24

Depends: 	 R(>= 2.6.0), methods, startupmsg, distr(>= 2.0), distrEx(>= 2.0), RandVar(>= 0.6.3), MASS, stats4

LazyLoad: 	 yes

License: 	 LGPL-3

URL: 	 http://distr.r-forge.r-project.org/

Classes


[*]: there is a generating function with the same name

##########################
ProbFamily classes
##########################
slots: [<name>(<class>)] 
name(character), distribution(Distribution),
distrSym(DistributionSymmetry), props(character)

"ProbFamily"
|>"ParamFamily"     [*] 
additional slots:
param(ParamFamParameter), modifyParam(function),
startPar(function), makeOKPar(function), fam.call(call)
|>|>"L2ParamFamily" [*]
additional slots:
L2deriv(EuclRandVarList), L2deriv.fct(function), 
L2derivSymm(FunSymmList), L2derivDistr(DistrList),
L2derivDistrSymm(DistrSymmList), FisherInfo(PosSemDefSymmMatrix),
FisherInfo.fct(function)
|>|>|>"BinomFamily" [*]
|>|>|>"PoisFamily"  [*]
|>|>|>"GammaFamily" [*]
|>|>|>"BetaFamily"  [*]
|>|>|>"L2GroupParamFamily"
additional slots:
LogDeriv(function)
|>|>|>|>"L2LocationScaleUnion"  /VIRTUAL/
additional slots:
locscalename(character)
|>|>|>|>|>"L2LocationFamily"            [*]
|>|>|>|>|>|>"NormLocationFamily"        [*] 
|>|>|>|>|>|>"GumbelLocationFamily"      [*] 
|>|>|>|>|>"L2ScaleFamily"               [*]
|>|>|>|>|>|>"NormScaleFamily"           [*] 
|>|>|>|>|>|>"ExpScaleFamily"            [*] 
|>|>|>|>|>|>"LnormScaleFamily"          [*] 
|>|>|>|>|>"L2LocationScaleFamily"       [*]
|>|>|>|>|>|>"NormLocationScaleFamily"   [*] 
|>|>|>|>|>|>"CauchyLocationScaleFamily" [*] 

##########################
ParamFamParameter
##########################
"ParamFamParameter" [*] is subclass of class "Parameter" of package "distr".
Additional slots:
main(numeric), nuisance(OptionalNumeric), fixed(OptionalNumeric),
trafo(MatrixorFunction)

##########################
Class unions
##########################
"OptionalNumeric" = union("numeric", "NULL")
"OptionalMatrix" = union("matrix", "NULL")
"MatrixorFunction" = union("matrix", "OptionalFunction")
"PrintDetails" = union("Estimate", "Confint", 
                   "PosSemDefSymmMatrix",
                   "ParamFamParameter", "ParamFamily")


##########################
Symmetry classes
##########################
slots:
type(character), SymmCenter(ANY)

"Symmetry"
|>"NoSymmetry"          [*]
|>"EllipticalSymmetry"  [*] 
|>|>"SphericalSymmetry" [*]
|>"DistributionSymmetry"
|>"FunctionSymmetry"
|>|>"NonSymmetric"      [*]
|>|>"EvenSymmetric"     [*]
|>|>"OddSymmetric"      [*]

list thereof
"DistrSymmList"         [*] 
"FunSymmList"           [*]


##########################
Matrix classes
##########################
slots:
none
"PosSemDefSymmMatrix" [*] is subclass of class "matrix" of package "base".
|>"PosDefSymmMatrix"  [*]


##########################
Norm Classes
##########################
slots:
name(character), fct(function)

"NormType"        [*]
|>"QFNorm"        [*]
Additional slots:
QuadForm(PosSemDefSymmMatrix)
|>|>"InfoNorm"    [*]
|>|>"SelfNorm"    [*]


##########################
Bias Classes
##########################
slots:
name(character)

"BiasType"
|>"symmetricBias"   [*]
|>"onesidedBias"
Additional slots:
sign(numeric)
|>"asymmetricBias"  [*]
Additional slots:
nu(numeric)


##########################
Risk Classes
##########################
slots:
type(character)

"RiskType"
|>"asRisk"
|>|>"asCov"       [*]
|>|>"trAsCov"     [*]
|>"fiRisk"
|>|>"fiCov"       [*]
|>|>"trfiCov"     [*]
|>|>"fiHampel"    [*]
Additional slots:
bound(numeric)
|>|>"fiMSE"       [*]
|>|>"fiBias"      [*]
|>|>"fiUnOvShoot" [*]
Additional slots:
width(numeric)

Risk with Bias:
"asRiskwithBias"
slots: biastype(BiasType), normtype(NormType),
|>"asHampel"      [*]
Additional slots:
bound(numeric)
|>"asBias"        [*]
|>"asGRisk"
|>|>"asMSE"       [*]
|>|>"asUnOvShoot" [*]
Additional slots:
width(numeric)
|>|>"asSemivar"   [*]


##########################
Estimate Classes
##########################
slots:
name(character), estimate(ANY),
samplesize(numeric), asvar(OptionalMatrix),
Infos(matrix), nuis.idx(OptionalNumeric)
fixed.estimate(OptionalNumeric), 
estimate.call(call), trafo(list[of function, matrix]),
untransformed.estimate(ANY), 
untransformed.asvar(OptionalMatrix)
criterion.fct(function), method(character),

"Estimate"
|>"MCEstimate", 
Additional slots:
criterion(numeric)

##########################
Confidence interval class
##########################
slots:
type(character), confint(array),
estimate.call(call), name.estimate(character),
trafo.estimate(list[of function, matrix]), 
nuisance.estimate(OptionalNumeric)
"Confint"


Methods

besides accessor and replacement functions, we have methods

solve, sqrt for matrices checkL2deriv, existsPIC for class L2ParamFamily LogDeriv for class L2GroupParamFamily

validParameter for classes ParamFamily, L2ScaleFamily, L2LocationFamily, and L2LocationScaleFamily

modifyModel for the pairs of classes L2ParamFamily and ParamFamParameter, L2LocationFamily and ParamFamParameter, L2ScaleFamily and ParamFamParameter, L2LocationScaleFamily and ParamFamParameter, GammaFamily and ParamFamParameter, and ExpScaleFamily and ParamFamParameter

mceCalc for the pair of classes numeric and ParamFamily

mleCalc for the pairs of classes numeric and ParamFamily, numeric and BinomFamily, numeric and PoisFamily, numeric and NormLocationFamily, numeric and NormScaleFamily, and numeric and NormLocationScaleFamily

coerce from class MCEstimate to class mle

confint for class Estimate profile for class MCEstimate

Functions


Management of global options:
"distrModOptions", "distrModoptions", "getdistrModOption",

check for ker of matrix: "isKerAinKerB"

particular norms: "EuclideanNorm", "QuadFormNorm"
onesided bias: "positiveBias", "negativeBias", 

Estimators:
"Estimator", "MCEstimator", "MLEstimator", "MDEstimator"

special location/scale models:
"L2LocationUnknownScaleFamily", "L2ScaleUnknownLocationFamily"

some special normal models:
"NormScaleUnknownLocationFamily", "NormLocationUnknownScaleFamily",

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Demos

Demos are available — see demo(package="distrMod").

Scripts

Example scripts are available — see folder ‘scripts’ in the package folder to package distrMod in your library.

Note

Some functions of packages stats, base have intentionally been masked, but completely retain their functionality — see distrModMASK().

If any of the packages stats4, fBasics is to be used together with distrMod, the latter must be attached after any of the first mentioned. Otherwise confint() defined as method in distrMod may get masked.
To re-mask, you may use confint <- distrMod::confint. See also distrModMASK()

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

Maintainer: Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

P. Ruckdeschel, M. Kohl, T. Stabla, F. Camphausen (2006): S4 Classes for Distributions, {em R News}, {em 6}(2), 2-6. http://CRAN.R-project.org/doc/Rnews/Rnews_2006-2.pdf

a vignette for packages distr, distrSim, distrTEst, and distrEx is included into the mere documentation package distrDoc and may be called by require("distrDoc");vignette("distr")

[Package distrMod version 2.0.3 Index]