distrSim-package(distrSim)	R Documentation

distrSim – (S4-) classes for simulations based on package distr

Description

distrSim is to provide a conceptual treatment of simulations by means of S4 classes. The package is based on our package distr, hence uses distribution classes as introduced there to describe the distributions under which simulations are performed.

Details

Package: 	 distrSim

Version: 	 2.2

Date: 	 2009-06-29

Depends: 	 R(>= 2.6.0), methods, graphics, setRNG(>= 2006.2-1), distr(>= 2.0), startupmsg

LazyLoad: 	 yes

License: 	 LGPL-3

URL: 	 http://distr.r-forge.r-project.org/

Classes


"SeqDataFrame"
      slots: data(list[of dataframes with common column structure])
"Dataclass"
      slots: [<name>(<class>)] 
      filename(vectororNULL), 
      name(character), 
      Data(ArrayorNULLorVectororDataframeorSeqDataFrames), 
      runs(numeric), 
      samplesize(numeric),
      obsDim(numeric)      
|>"Simulation"
      (additional) slots: [<name>(<class>)] 
      seed(list), distribution(Distribution)
|>"Contsimulation"
      (additional) slots: [<name>(<class>)] 
      seed(list), ind(vectororNULL), rate(numeric),
      distribution.id(Distribution),
      distribution.c(Distribution),
      Data.id(vectororNULL),
      Data.c(vectororNULL)

Methods


savedata                Methods to save the data slot 
                       (for Simulation/Contsimulation)
simulate                Methods to fill the data slot 
                       (for Simulation/Contsimulation)
plot-methods            Methods for Function plot
                       (for Simulation/Contsimulation)
print-, show-methods    Methods for Function print/show 
                       (for Simulation/Contsimulation)
summary-methods         Methods for Function summary 
                       (for Simulation/Contsimulation)

Functions


cload                   loads just the comment slot 
                       (for Simulation/Contsimulation)

Slot accessors / -replacement functions

All slots are inspected / modified by corresponding accessors / -replacement functions, e.g. rate(X) for an object of class "Contsimulation".

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

Far-reaching Change in Design

From version 1.8 of this package on, we have changed the format how data / simulations are stored: In order to be able to cope with multivariate distributions, regression distributions and (later) time series distributions, we have switched to the common array format: samplesize x obsDim x runs; you may check the version under which an object was generated by getVersion; for saved objects from earlier versions, we provide the functions isOldVersion, and conv2NewVersion-methods to check whether the object was generated by an older version of this package and to convert such an object to the new format, respectively.

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Note

Global options controlling the plots and summaries of Dataclass and Simulation/Contsimulation objects may be inspected / set by distrSimoptions() and getdistrSimOption().

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de
Maintainer: Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

P. Ruckdeschel, M. Kohl, T. Stabla, F. Camphausen (2006): S4 Classes for Distributions, R News, 6(2), 2-6. http://CRAN.R-project.org/doc/Rnews/Rnews_2006-2.pdf

A vignette for packages distr, distrSim, distrTEst, and distrEx is included into the mere documentation package distrDoc and may be called by require("distrDoc");vignette("distr").

A homepage to this package is available under
http://distr.r-forge.r-project.org/

See Also

distr setRNG

[Package distrSim version 2.2 Index]