distrTEst-package(distrTEst)	R Documentation

dsitrTEst – Estimation and Testing classes based on package distr

Description

distrTest provides (S4-)classes for evaluating procedures (estimators/tests) at data/simulation in a unified way based on package distr. This is achieved by means of the S4 class Evaluation. The package is based on our packages distr and distrSim, hence uses distribution classes and simulation classes as introduced there to capture the situation from which the simulations stem.

Details

Package: 	 distrTEst

Version: 	 1.9

Date: 	 2007-07-27

Depends: 	 R(>= 2.2.0), methods, graphics, setRNG(>= 2006.2-1), startupmsg, distr(>=1.8), distrSim(>= 1.8)

Suggests: 	 distrEx(>= 0.4-4)

Imports: 	 stats

SaveImage: 	 no

LazyLoad: 	 yes

License: 	 GPL (version 2 or later)

URL: 	 http://www.uni-bayreuth.de/departments/math/org/mathe7/DISTR/

Classes


"Evaluation"
      slots: [<name>(<class>)] 
      name(character), filename(character), call.ev(call),
      Data(Dataclass), result(DataframeorNULL), 
      estimator(OptionalFunction)
"EvaluationList"
      slots: [<name>(<class>)] 
      name(character), Elist(list)

Objects of class "Evaluation" are generated by a call to evaluate.

Methods


plot            plot method for "Evaluation" and for "EvaluationList"
print,show      print/show method for "Evaluation" and for "EvaluationList"
summary         summary method for "Evaluation" and for "EvaluationList"
Data            accessor method for "Evaluation", and, for "EvaluationList" 
                returns common Data

Slot accessors / -replacement functions

All slots are inspected / modified by corresponding accessors / -replacement functions, e.g. call.ev(X) or filename(X)<-"myevaluation.sav" for an object of class "Evaluation".

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Note

Global options controlling the plots and summaries of Evaluationlist objects may be inspected / set by distrTEstoptions() and getdistrTEstOption().

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Matthias Kohl Matthias.Kohl@stamats.de

Maintainer: Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

A vignette for packages distr, distrSim, distrTEst, and distrEx is included into the mere documentation package distrDoc and may be called by require("distrDoc");vignette("distr").

A homepage to this package is available under
http://www.uni-bayreuth.de/departments/math/org/mathe7/DISTR/distrTEst.html.

See Also

distr, distrSim, setRNG

[Package distrTEst version 1.9 Index]