distrTeach-package(distrTeach)	R Documentation

distrTeach – Teaching Extensions of package distr

Description

distrTeach provides some illustrations based on package distr for teaching Stochastics / Statistics in secondary school; so far the following has been implemented

illustrateLLT: function for the generation of LLN - visualizations
illustrateCLT: function for the generation of CLT - visualizations
plotCLT: Generic function for the plotting of CLT-approximations

as well as a Tcl/Tk based demo forillustrateCLT

Details

Package: 	 distrTeach

Version: 	 2.0

Date: 	 2008-01-17

Depends: 	 R(>= 2.2.0), methods, distr(>= 1.8), evd, startupmsg

Suggests: 	 tcltk

SaveImage: 	 no

LazyLoad: 	 yes

License: 	 GPL (version 2 or later)

URL: 	 http://www.uni-bayreuth.de/departments/math/org/mathe7/DISTR/

Classes



Teaching Classes


Functions




Generating Functions





Methods



illustration:
illustrateLLT           function for the generation of LLN - visualizations
illustrateCLT           function for the generation of CLT - visualizations
plotCLT                 Generic function for the plotting of CLT-approximations


Demos

Demos are available — see demo(package="distrTeach").

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de and
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Eleonara Feist eleonoragerber@gmx.de, and,
Anja Hueller anja_h86@web.de

Maintainer: Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

P. Ruckdeschel, M. Kohl, T. Stabla, F. Camphausen (2006): S4 Classes for Distributions, {em R News}, {em 6}(2), 2-6. http://CRAN.R-project.org/doc/Rnews/Rnews_2006-2.pdf

a vignette for packages distr, distrSim, distrTEst, and distrTeach is included into the mere documentation package distrDoc and may be called by require("distrDoc");vignette("distr")

a homepage to this package is available under
http://distr.r-forge.r-project.org/ and the pages ...

M. Kohl (2005): Numerical Contributions to the Asymptotic Theory of Robustness. PhD Thesis. Bayreuth. Available as http://www.stamats.de/ThesisMKohl.pdf

See Also

distr distrEx

[Package distrTeach version 2.0 Index]