AbscontCondDistribution-class(distrEx)	R Documentation

Absolutely continuous conditional distribution

Description

The class of absolutely continuous conditional univariate distributions.

Objects from the Class

Objects can be created by calls of the form new("AbscontCondDistribution", ...).

Slots

cond:

Object of class "Condition": condition

img:

Object of class "rSpace": the image space.

param:

Object of class "OptionalParameter": an optional parameter.

r:

Object of class "function": generates random numbers.

d:

Object of class "OptionalFunction": optional conditional density function.

p:

Object of class "OptionalFunction": optional conditional cumulative distribution function.

q:

Object of class "OptionalFunction": optional conditional quantile function.

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

Extends

Class "UnivariateCondDistribution", directly.
Class "Distribution", by class "UnivariateCondDistribution".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

UnivariateCondDistribution-class, Distribution-class

Examples

new("AbscontCondDistribution")

[Package distrEx version 2.0 Index]