Arcsine-class(distr)	R Documentation

Class 'Arcsine'

Description

The Arcsine distribution has density

1/sqrt(1+x^2)/pi

for -1 <= x <= 1.

Objects from the Class

Objects can be created by calls of the form Arcsine(). This object is an Arcsine distribution.

Slots

img:

Object of class "Reals": The space of the image of this distribution has got dimension 1 and the name "Real Space".

r:

Object of class "function": generates random numbers (calls function rArcsine)

d:

Object of class "function": density function (calls function dArcsine)

p:

Object of class "function": cumulative function (calls function pArcsine)

q:

Object of class "function": inverse of the cumulative function (calls function qArcsine)

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "AbscontDistribution".

Methods

initialize

signature(.Object = "Arcsine"): initialize method

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

AbscontDistribution-class Reals-class

Examples


A <- Arcsine()
# A is a Arcsine distribution with shape1 = 1 and shape2 = 1.
r(A)(3) # three random number generated from this distribution, e.g. 0.6979795
d(A)(c(-2,-1,-0.2,0,0.2,1,2)) # Density at x=c(-1,-0.2,0,0.2,1).
p(A)(c(-2,-1,-0.2,0,0.2,1,2)) # cdf at q=c(-1,-0.2,0,0.2,1).
q(A)(c(0,0.2,1,2)) # quantile function at at x=c(0,0.2,1).

[Package distr version 2.1 Index]