CauchyParameter-class(distr)	R Documentation

Class 'CauchyParameter'

Description

The parameter of a Cauchy distribution, used by Cauchy-class

Objects from the Class

Objects can be created by calls of the form new("CauchyParameter", location, scale). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Cauchy is instantiated.

Slots

location:

Object of class "numeric": the location of a Cauchy distribution

scale:

Object of class "numeric": the scale of a Cauchy distribution

name:

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "CauchyParameter"): initialize method

scale

signature(object = "CauchyParameter"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "CauchyParameter"): modifies the slot scale of the parameter of the distribution

location

signature(object = "CauchyParameter"): returns the slot location of the parameter of the distribution

location<-

signature(object = "CauchyParameter"): modifies the slot location of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Cauchy-class Parameter-class

Examples


W <- new("CauchyParameter",location=1,scale=1)
location(W) # location of this distribution is 1.
location(W) <- 2 # location of this distribution is now 2.

[Package distr version 2.1 Index]