Confint-class(distrMod)	R Documentation

Confint-class

Description

Return value S4 classes for method ``confint''.

Objects from the Class

Objects could in principle be created by calls of the form new("Confint", ...). The preferred form is to have them created via a call to confint.

Slots

type:

Object of class "character": type of the confidence interval (asymptotic, bootstrap,...).

confint:

Object of class "array": the confidence interval(s).

call.estimate:

Object of class "call": the estimate(s) for which the confidence intervals are produced.

name.estimate:

Object of class "character": the name of the estimate(s) for which the confidence intervals are produced.

samplesize.estimate:

Object of class "numeric": the sample size of the estimate(s) for which the confidence intervals are produced.

trafo.estimate:

Object of class "matrix": the trafo/derivative matrix of the estimate(s) for which the confidence intervals are produced.

nuisance.estimate:

Object of class "OptionalNumeric": the nuisance parameter (if any) at which the confidence intervals are produced.

Methods

type

signature(object = "Confint"): accessor function for slot type.

confint

signature(object = "Confint", method = "missing"): accessor function for slot type.

call.estimate

signature(object = "Confint"): accessor function for slot call.estimate.

name.estimate

signature(object = "Confint"): accessor function for slot name.estimate.

trafo.estimate

signature(object = "Confint"): accessor function for slot trafo.estimate.

samplesize.estimate

signature(object = "Confint"): accessor function for slot samplesize.estimate.

nuisance.estimate

signature(object = "Confint"): accessor function for slot nuisance.estimate.

show

signature(object = "Confint"): shows a detailed view of the object; slot nuisance.estimate is only shown if non-null, and slot trafo.estimate only if different from a unit matrix.

print

signature(object = "Confint"): just as show, but with additional arguments digits.

Details for methods this-is-escaped-codenormal-bracket97bracket-normal, this-is-escaped-codenormal-bracket98bracket-normal

Detailedness of output by methods show, print is controlled by the global option show.details to be set by distrModoptions.

As method show is used when inspecting an object by typing the object's name into the console, show comes without extra arguments and hence detailedness must be controlled by global options.

Method print may be called with a (partially matched) argument show.details, and then the global option is temporarily set to this value.

More specifically, when show.detail is matched to "minimal" you will be shown only the type of the confidence interval(s) and its/their values. When show.detail is matched to "medium", you will in addition see the type of the estimator(s) for which it is produced, the corresponding call of the estimater, its sample size, and, if present, the value of the corresponding nuisance parameter. Finally, when show.detail is matched to "maximal", additionally you will be shown the transformation of the estimator (if non-trivial, i.e. the identity) in form of its function code respectively of its derivative matrix.

Note

The pretty-printing code for methods show and print has been borrowed from confint.default in package stats.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Estimator, confint, Estimate-class, trafo-methods

Examples


## some transformation
mtrafo <- function(x){
     nms0 <- c("scale","shape")
     nms <- c("shape","rate")
     fval0 <- c(x[2], 1/x[1])
     names(fval0) <- nms
     mat0 <- matrix( c(0, -1/x[1]^2, 1, 0), nrow = 2, ncol = 2,
                     dimnames = list(nms,nms0))                          
     list(fval = fval0, mat = mat0)}

x <- rgamma(50, scale = 0.5, shape = 3)

## parametric family of probability measures
G <- GammaFamily(scale = 1, shape = 2, trafo = mtrafo)
## MLE
res <- MLEstimator(x = x, ParamFamily = G)
ci <- confint(res)
print(ci, digits = 4, show.details="maximal")
print(ci, digits = 4, show.details="medium")
print(ci, digits = 4, show.details="minimal")

[Package distrMod version 2.0 Index]