AbscontDistribution(distr)	R Documentation

Generating function 'AbscontDistribution'

Description

Generates an object of class "AbscontDistribution"

Usage


  AbscontDistribution(r)
  AbscontDistribution(r = NULL, d)
  AbscontDistribution(r = NULL, d = NULL, p)
  AbscontDistribution(r = NULL, d = NULL, p = NULL, d)
  AbscontDistribution(r, d, p, q)
  AbscontDistribution(r = NULL, d = NULL, p = NULL, q = NULL,
                   gaps = NULL, param = NULL, img = new("Reals"),
                   .withSim = FALSE, .withArith = FALSE,
                   low1 = NULL, up1 = NULL, low = -Inf, up =Inf,
                   ngrid = getdistrOption("DefaultNrGridPoints"),
                   ep = getdistrOption("TruncQuantile"),
                   e = getdistrOption("RtoDPQ.e"),
                   withgaps = getdistrOption("withgaps"))

Arguments

r 	 slot r to be filled

d 	 slot d to be filled

p 	 slot p to be filled

q 	 slot q to be filled

gaps 	 slot gaps (of class "matrix" with two columns) to be filled (i.e. t(gaps) must be ordered if read as vector)

param 	 parameter (of class "OptionalParameter")

img 	 image range of the distribution (of class "rSpace")

low1 	 lower bound (to be the lower TruncQuantile-quantile of the distribution)

up1 	 upper bound (to be the upper TruncQuantile-quantile of the distribution)

low 	 lower bound (to be the 100-percent-quantile of the distribution)

up 	 upper bound (to be the 100-percent-quantile of the distribution)

ngrid 	 number of gridpoints

ep 	 tolerance epsilon

e 	 exponent to base 10 to be used for simulations

withgaps 	 logical; shall gaps be reconstructed empirically?

.withArith 	 normally not set by the user, but if determining the entries supp, prob distributional arithmetics was involved, you may set this to TRUE.

.withSim 	 normally not set by the user, but if determining the entries supp, prob simulations were involved, you may set this to TRUE.

Details

Minimally, only one of the slots r, d, p or q needs to be given as argument. The other non-given slots are then reconstructed according to the following scheme:

r 	 d 	 p 	 q 	 proceding

- 	 - 	 - 	 - 	 excluded

- 	 + 	 - 	 - 	 p by .D2P, q by .P2Q, r by q(runif(n))

- 	 - 	 + 	 - 	 d by .P2D, q by .P2Q, r by q(runif(n))

- 	 + 	 + 	 - 	 q by .P2Q, r by q(runif(n))

- 	 - 	 - 	 + 	 p by .Q2P, d by .P2D, r by q(runif(n))

- 	 + 	 - 	 + 	 p by .Q2P, r by q(runif(n))

- 	 - 	 + 	 + 	 d by .P2D, r by q(runif(n))

- 	 + 	 + 	 + 	 r by q(runif(n))

+ 	 - 	 - 	 - 	 call to RtoDPQ

+ 	 + 	 - 	 - 	 p by .D2P, q by .P2Q

+ 	 - 	 + 	 - 	 d by .P2D, q by .P2Q

+ 	 + 	 + 	 - 	 q by .P2Q

+ 	 - 	 - 	 + 	 p by .Q2P, d by .P2D

+ 	 + 	 - 	 + 	 p by .Q2P

+ 	 - 	 + 	 + 	 d by .P2D

+ 	 + 	 + 	 + 	 nothing

For this purpose, one may alternatively give arguments low1 and up1 (NULL each by default, and determined through slot q, resp. p, resp. d, resp. r in this order according to availability), for the (finite) range of values in the support of this distribution, as well as the possibly infinite theoretical range given by arguments low and up with default values -Inf, Inf, respectively. Of course all other slots may be specified as arguments.

Value

Object of class "AbscontDistribution"

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

AbscontDistribution-class, DiscreteDistribution-class, RtoDPQ

Examples


plot(Norm())
plot(AbscontDistribution(r = rnorm))
plot(AbscontDistribution(d = dnorm))
plot(AbscontDistribution(p = pnorm))
plot(AbscontDistribution(q = qnorm))

[Package distr version 2.0 Index]