ContaminationSize(distrEx)	R Documentation

Generic Function for the Computation of the Convex Contamination (Pseudo-)Distance of Two Distributions

Description

Generic function for the computation of convex contamination (pseudo-)distance of two probability distributions P and Q. That is, the minimal size 0 <= epsilon <= 1 is computed such that there exists some probability distribution R with

Q = (1 - epsilon)P + epsilon R

Usage


ContaminationSize(e1, e2, ...)
## S4 method for signature 'AbscontDistribution,
##   AbscontDistribution':
ContaminationSize(e1,e2)
## S4 method for signature 'DiscreteDistribution,
##   DiscreteDistribution':
ContaminationSize(e1,e2)
## S4 method for signature 'AcDcLcDistribution,
##   AcDcLcDistribution':
ContaminationSize(e1,e2)

Arguments

e1 	 object of class "Distribution"

e2 	 object of class "Distribution"

... 	 further arguments to be used in particular methods (not in package distrEx)

Details

Computes the distance from e1 to e2 respectively P to Q. This is not really a distance as it is not symmetric!

Value

A list containing the following components:

e1 	 object of class "Distribution"; ideal distribution

e2 	 object of class "Distribution"; 'contaminated' distribution

size.of.contamination 	 size of contamination

Methods

e1 = "AbscontDistribution", e2 = "AbscontDistribution":

convex contamination (pseudo-)distance of two absolutely continuous univariate distributions.

e1 = "DiscreteDistribution", e2 = "DiscreteDistribution":

convex contamination (pseudo-)distance of two discrete univariate distributions.

e1 = "AcDcLcDistribution", e2 = "AcDcLcDistribution":

convex contamination (pseudo-)distance of two discrete univariate distributions.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Huber, P.J. (1981) Robust Statistics. New York: Wiley.

See Also

KolmogorovDist, TotalVarDist, HellingerDist, Distribution-class

Examples


ContaminationSize(Norm(), Norm(mean=0.1))
ContaminationSize(Pois(), Pois(1.5))

[Package distrEx version 2.0 Index]