DExp-class(distr)	R Documentation

Class "DExp"

Description

The double exponential or Laplace distribution with rate \lambda has density

f(x) = 1/2 lambda e^(- lambda |x|)

C.f. , rexp

Objects from the Class

Objects can be created by calls of the form DExp(rate). This object is a double exponential (or Laplace) distribution.

Slots

img

Object of class "Reals": The space of the image of this distribution has got dimension 1 and the name "Real Space".

param

Object of class "ExpParameter": the parameter of this distribution (rate), declared at its instantiation

r

Object of class "function": generates random numbers (calls function rexp)

d

Object of class "function": density function (calls function dexp)

p

Object of class "function": cumulative function (calls function pexp)

q

Object of class "function": inverse of the cumulative function (calls function qexp)

.withArith

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim

logical: used internally to issue warnings as to accuracy

.logExact

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Symmetry

object of class "DistributionSymmetry"; used internally to avoid unnecessary calculations.

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution". Class "Distribution", by class "AbscontDistribution".

Methods

initialize

signature(.Object = "DExp"): initialize method

rate

signature(object = "DExp"): returns the slot rate of the parameter of the distribution

rate<-

signature(object = "DExp"): modifies the slot rate of the parameter of the distribution

*

signature(e1 = "DExp", e2 = "numeric"): For the Laplace distribution we use its closedness under scaling transformations.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

rexp

Examples


D <- DExp(rate = 1) # D is a Laplace distribution with rate = 1.
r(D)(1) # one random number generated from this distribution, e.g. 0.4190765
d(D)(1) # Density of this distribution is 0.1839397 for x = 1.
p(D)(1) # Probability that x < 1 is 0.8160603.
q(D)(.1) # Probability that x < -1.609438 is 0.1.
rate(D) # rate of this distribution is 1.
rate(D) <- 2 # rate of this distribution is now 2.
3*D ###  still a DExp -distribution

[Package distr version 2.2 Index]