Dataclass-class(distrSim)	R Documentation

Class 'Dataclass'

Description

In an object of type "Dataclass" data can be saved containing any number of runs in any dimension. All information about the data is stored in a unified way.

Objects from the Class

Objects can be created by calls to the generating function, method Dataclass(Data, filename = "Data-set", name = "Data-Set"). A Dataclass-object includes, aside from the actual data, a filename and the size of the sample, the observation dimension, and the number of runs, which give the number of rows and columns (and, if more than one run, slices) of the data array.

Slots

filename:

Object of class "character": the filename the data shall be saved

name:

Object of class "character": a name for the Data

Data:

Object of class "ArrayorNULLorVectororDataframeorSeqDataFrames": the actual data, either of type "NULL" (means no data) or "vector" or "array" or "Dataframe" or "SeqDataFrames"

obsDim:

Object of class "numeric": the observation dimension of the data

runs:

Object of class "numeric": the number of runs of the data

samplesize:

Object of class "numeric": the size of the sample

version:

Object of class "character": the package version under which this object was generated

Methods

Data

signature(object = "Dataclass"): returns the actual data

Data<-

signature(object = "Dataclass"): changes the data

evaluate

signature(object = "Dataclass", estimator = "function"):
creates an object of type "Evaluation", see there for further information

filename

signature(object = "Dataclass"): returns the the filename

filename<-

signature(object = "Dataclass"): changes the the filename

name

signature(object = "Dataclass"): returns the the name

name<-

signature(object = "Dataclass"): changes the the name

initialize

signature(.Object = "Dataclass"): initialize method

obsDim

signature(object = "Dataclass"): returns the dimension of the observations

runs

signature(object = "Dataclass"): returns the number of runs

samplesize

signature(object = "Dataclass"): returns the size of the sample

getVersion

signature(object = "Dataclass"): returns the version slot of this object

savedata

signature(object = "Dataclass"): saves the object in the directory of R and also a copy without data

plot

signature(x = "Dataclass"): produces a plot of the data matrix; ; for details confer plot-methods

print

signature(x = "Dataclass"): returns filename, the observation dimension, the number of runs and the size of the sample, and, if from a version > 1.8, also the package version under which the object was generated

summary

signature(object = "Dataclass"): returns the same information as print, moreover a statistical summary for each run

Note

The saved Dataclass can be loaded with the usual load-command, the saved comment with the function cload.

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@web.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Simulation-class Contsimulation-class Evaluation-class plot-methods
print-methods summary-methods load cload savedata-methods getVersion-methods

Examples


D66 <- Dataclass(filename="N74", Data = matrix(1:36,6))
D66
#
D <- Dataclass(Data = array(c(1,2,3,4,5,6),c(samplesize=2,obsdim=3,Runs=1)),
               filename = "xyz.sav")
# A new object of type "Dataclass" is created.
#
isOldVersion(D) ##NO!
#
savedata(D)
# creates a file called "xyz.sav" where the information is saved and a
# copy "xyz.sav.comment" without data
Data(D) <- array(c(11,12,13,14,15,16),c(samplesize=2,obsdim=3,Runs=1)) # changes the data of D
cload("xyz.sav") # loads the object without data - it is called "D.comment"
D.comment
load("xyz.sav") # loads the original object "D"
Data(D) # the original data: matrix(c(1,2,3,4,5,6),2)
#if you have distrTEst available:
#evaluate(object = D, estimator = mean) # returns the mean of each variable

[Package distrSim version 2.0 Index]