DiscreteMVDistribution(distrEx)	R Documentation

Generating function for DiscreteMVDistribution-class

Description

Generates an object of class "DiscreteMVDistribution".

Usage


  DiscreteMVDistribution(supp, prob)
  DiscreteMVDistribution(supp)

Arguments

supp 	 numeric matrix whose rows form the support of the discrete multivariate distribution.

prob 	 vector of probability weights for the elements of supp.

Details

Identical rows are collapsed to unique support values. If prob is missing, all elements in supp are equally weighted.

Value

Object of class "DiscreteMVDistribution"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

DiscreteMVDistribution-class

Examples


# Dirac-measure at (0,0,0)
D1 <- DiscreteMVDistribution(supp = c(0,0,0))
support(D1)

# simple discrete distribution
D2 <- DiscreteMVDistribution(supp = matrix(c(0,1,0,2,2,1,1,0), ncol=2), 
                prob = c(0.3, 0.2, 0.2, 0.3))
support(D2)
r(D2)(10)

[Package distrEx version 2.1 Index]