DiscreteMVDistribution-class(distrEx)	R Documentation

Discrete Multivariate Distributions

Description

The class of discrete multivariate distributions.

Objects from the Class

Objects can be created by calls of the form new("DiscreteMVDistribution", ...). More frequently they are created via the generating function DiscreteMVDistribution.

Slots

img:

Object of class "rSpace". Image space of the distribution. Usually an object of class "EuclideanSpace".

param:

Object of class "OptionalParameter". Optional parameter of the multivariate distribution.

r:

Object of class "function": generates (pseudo-)random numbers

d:

Object of class "OptionalFunction": optional density function

p:

Object of class "OptionalFunction": optional cumulative distribution function

q:

Object of class "OptionalFunction": optional quantile function

support:

numeric matrix whose rows form the support of the distribution

Extends

Class "MultivariateDistribution", directly.
Class "Distribution", by class "MultivariateDistribution".

Methods

support

signature(object = "DiscreteMVDistribution"): accessor function for slot support.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Distribution-class, MultivariateDistribution-class, DiscreteMVDistribution, E-methods

Examples


(D1 <- new("MultivariateDistribution")) # Dirac measure in (0,0)
r(D1)(5)

(D2 <- DiscreteMVDistribution(supp = matrix(c(1:5, rep(3, 5)), ncol=2, byrow=TRUE)))
support(D2)
r(D2)(10)
d(D2)(support(D2))
p(D2)(lower = c(1,1), upper = c(3,3))
q(D2)
param(D2)
img(D2)

e1 <- E(D2) # expectation

[Package distrEx version 2.0 Index]