DistrList(distr)	R Documentation

Generating function for DistrList-class

Description

Generates an object of class "DistrList".

Usage

DistrList(..., Dlist)

Arguments

... 	 Objects of class "Distribution" (or subclasses)

Dlist 	 an optional list or object of class "DistrList"; if not missing it is appended to argument ...; this way DistrList may also be called with a list (or "DistrList"-object) as argument as suggested in an e-mail by Krunoslav Sever (thank you!)

Value

Object of class "DistrList"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

DistrList-class, UnivarDistrList-class, UnivarDistrList

Examples


(DL <- DistrList(Norm(), Exp(), Pois()))
plot(DL)
as(Norm(), "DistrList")

## The function is currently defined as
function(...){ 
    new("DistrList", list(...)) 
}

[Package distr version 2.2 Index]