DistrList-class(distr)	R Documentation

List of distributions

Description

Create a list of distributions

Objects from the Class

Objects can be created by calls of the form new("DistrList", ...). More frequently they are created via the generating function DistrList.

Slots

.Data:

Object of class "list". A list of distributions.

Extends

Class "list", from data part.
Class "vector", by class "list".

Methods

show

signature(object = "DistrList")

plot

signature(object = "DistrList")

coerce

signature(from = "Distribution", to = "DistrList"): create a "DistrList" object from a "Distribution" object

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

DistrList, Distribution-class

Examples


(DL <- new("DistrList", list(Norm(), Exp())))
plot(DL)
as(Norm(), "DistrList")

[Package distr version 2.2 Index]