Distribution-class(distr)	R Documentation

Class 'Distribution'

Description

The Distribution-class is the mother-class of the class UnivariateDistribution.

Objects from the Class

Objects can be created by calls of the form new("Distribution").

Slots

img:

Object of class "Parameter": the space of the image

param:

Object of class "function": the parameter

r:

Object of class "function": generates random numbers

d:

Object of class "OptionalFunction": density function

p:

Object of class "OptionalFunction": cumulative distribution function

q:

Object of class "OptionalFunction": quantile function

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

Methods

img

signature(object = "Distribution"): returns the space of the image

param

signature(object = "Distribution"): returns the parameter

r

signature(object = "Distribution"): returns the random number generator

d

signature(object = "Distribution"): returns the density function

p

signature(object = "Distribution"): returns the cumulative distribution function

q

signature(object = "Distribution"): returns the quantile function

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

UnivariateDistribution-class Parameter-class

[Package distr version 1.9 Index]