EllipticalDistribution(distrEllipse)	R Documentation

Generating function for EllipticalDistribution-class

Description

Generates an object of class "EllipticalDistribution".

Usage

EllipticalDistribution(radDistr = sqrt(Chisq(df = length(loc))),
                            loc = c(0,0), scale = diag(length(loc)), p = NULL, q = NULL)

Arguments

radDistr 	 an object of class UnivariateDistribution with positive support, i.e. p(radDistr)(0)==0; the radial distribution.

loc 	 real number: location / center of the elliptical distribution.

scale 	 a square matrix (with nrow(scale)==ncol(scale)==length(loc)) of full rank: the / a scale matrix of the elliptical distribution — unique only upto scale%*%t(scale), i.e. if A1 and A2 are two square matrices of full rank such that A1%*%t(A1)==A2%*%t(A2), then we obtain the same elliptical distribution for scale = A1 and for scale = A2.

p 	 optional: p-slot of the corresponding distribution;

q 	 optional: q-slot of the corresponding distribution;

Value

Object of class "EllipticalDistribution"

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

EllipticalDistribution-class

Examples


E0 <- EllipticalDistribution()
plot(E0)
E1 <- diag(1,2)%*%E0+c(1,2)
plot(E1)
E(E1)
var(E1)

[Package distrEllipse version 0.1 Index]