EllipticalParameter-class(distrEllipse)	R Documentation

Paramter of an Elliptical distributions

Description

The class of the parameter of Elliptical distributions.

Objects from the Class

Objects can be created by calls of the form new("EllipticalParameter", ...).

Slots

loc:

numeric; center / location of the distribution.

scale:

matrix; the scale matrix; the number of rows of this matrix must be the same as the length of location.

name:

default name is “parameter of a Elliptical distribution”.

Extends

Class "Parameter", directly.
Class "OptionalParameter", by class "Parameter".

Methods

location

signature(object = "EllipticalParameter"): access method for slot location.

scale

signature(x = "EllipticalParameter"): access method for slot scale.

location<-

signature(object = "EllipticalParameter"): replace method for slot location.

scale<-

signature(object = "EllipticalParameter"): replace method for slot scale.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

, Parameter-class

Examples


new("EllipticalParameter")

[Package distrEllipse version 0.2 Index]