Estimate-class(distrMod)	R Documentation

Estimate-class.

Description

Class of estimates.

Objects from the Class

Objects can be created by calls of the form new("Estimate", ...). More frequently they are created via the generating function Estimator.

Slots

name:

Object of class "character": name of the estimator.

estimate:

Object of class "ANY": estimate.

estimate.call:

Object of class "call": call by which estimate was produced.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

asvar:

object of class "OptionalMatrix" which may contain the asymptotic (co)variance of the estimator.

samplesize:

object of class "numeric" — the samplesize at which the estimate was evaluated.

nuis.idx:

object of class "OptionalNumeric": indices of estimate belonging to the nuisance part

Methods

name

signature(object = "Estimate"): accessor function for slot name.

name<-

signature(object = "Estimate"): replacement function for slot name.

estimate

signature(object = "Estimate"): accessor function for slot estimate.

estimate.call

signature(object = "Estimate"): accessor function for slot estimate.call.

samplesize

signature(object = "Estimate"): accessor function for slot samplesize.

asvar

signature(object = "Estimate"): accessor function for slot asvar.

asvar<-

signature(object = "Estimate"): replacement function for slot asvar.

nuisance

signature(object = "Estimate"): accessor function for nuisance part of slot estimate.

main

signature(object = "Estimate"): accessor function for main part of slot estimate.

Infos

signature(object = "Estimate"): accessor function for slot Infos.

Infos<-

signature(object = "Estimate"): replacement function for slot Infos.

addInfo<-

signature(object = "Estimate"): function to add an information to slot Infos.

show

signature(object = "Estimate")

print

signature(object = "Estimate"): just as show, but with additional arguments digits.

Note

Parts of the code for methods show and print has been borrowed from print.fitdistr in package MASS by B.D. Ripley.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Estimator

Examples


x <- rnorm(100)
Estimator(x, estimator = mean, name = "mean")

[Package distrMod version 2.0 Index]