Estimator(distrMod)	R Documentation

Function to compute estimates

Description

The function Estimator provides a general way to compute estimates.

Usage


Estimator(x, estimator, name, Infos, asvar = NULL, nuis.idx,
          trafo = NULL, fixed = NULL, asvar.fct, ...)

Arguments

x 	 (empirical) data

estimator 	 function: estimator to be evaluated on x.

name 	 optional name for estimator.

Infos 	 character: optional informations about estimator

asvar 	 optionally the asymptotic (co)variance of the estimator

nuis.idx 	 optionally the indices of the estimate belonging to nuisance parameter

fixed 	 optionally (numeric) the fixed part of the parameter

trafo 	 an object of class MatrixorFunction – a transformation for the main parameter

asvar.fct 	 optionally: a function to determine the corresponding asymptotic variance; if given, asvar.fct takes arguments L2Fam((the parametric model as object of class L2ParamFamily)) and param (the parameter value as object of class ParamFamParameter); arguments are called by name; asvar.fct may also process further arguments passed through the ... argument

... 	 further arguments to estimator.

Details

The argument criterion has to be a function with arguments the empirical data as well as an object of class "Distribution" and possibly ....

Value

An object of S4-class "Estimate".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Estimate-class

Examples


x <- rnorm(100)
Estimator(x, estimator = mean, name = "mean")

X <- matrix(rnorm(1000), nrow = 10)
Estimator(X, estimator = rowMeans, name = "mean")

[Package distrMod version 2.2 Index]