EuclCondition-class(distrEx)	R Documentation

Conditioning by an Euclidean space.

Description

Conditioning by an Euclidean space.

Objects from the Class

Objects can be created by calls of the form new("EuclCondition", ...). More frequently they are created via the generating function EuclCondition.

Slots

Range:

Object of class "EuclideanSpace".

name:

Object of class "character": name of condition.

Extends

Class "Condition", directly.

Methods

Range

signature(object = "EuclCondition") accessor function for slot Range.

show

signature(object = "EuclCondition")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Condition-class, EuclCondition

Examples


  new("EuclCondition")

[Package distrEx version 1.9 Index]