EvaluationList-class(distrTEst)	R Documentation

Class 'EvaluationList'

Description

Several objects of class "Evaluation" may be gathered in a list of class "EvaluationList", if they all have the same result-format and also share the same data-set.

Objects from the Class

Objects may be created by the generating function EvaluationList, i.e.; EvaluationList(..., name0 = "a list of \"Evaluation\" objects"), where all arguments passed through ... have to be objects of class "Evaluation", the corresponding result-slots have to contain data.frames of identical dimension; the corresponding calls have to have identical object-arguments (for the data set), and the corresponding Data-slots have to be identical.

Slots

name:

Object of class "character": the name of the EvaluationList object

Elist:

Object of class "list": the list of Evaluation objects

Accesor/Replacement methods

Elist

signature(object = "EvaluationList"): returns the list with the Evaluation objects

name

signature(object = "EvaluationList"): returns/modifies the name of the EvaluationList object

Methods

Data

signature(object = "EvaluationList"): returns the common Data-slot of one of the Evaluation objects

plot

signature(object = "EvaluationList"): returns grouped boxplots of the results

print

signature(object = "EvaluationList"): for each list element returns the name of the data object, its filename, the estimator used and the result

show

signature(object = "EvaluationList"): as print

summary

signature(object = "EvaluationList"): returns the name of the data object, its filename, the estimator used and a statistical summary of the result

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Dataclass-class Simulation-class Contsimulation-class Evaluation-class print-methods plot-methods simulate-methods summary-methods

Examples


N <- Norm() # N is a standard normal distribution.
C <- Cauchy() # C is a Cauchy distribution
cs <- Contsimulation(filename = "csim",
                     runs = 15,
                     samplesize=500,
                     seed=setRNG(),
                     distribution.id = N,
                     distribution.c = C,
                     rate = 0.1)
simulate(cs)
# Each of the 25000 random numbers is ideal (N-distributed) with
# probability 0.9 and contaminated (C-distributed) with probability = 0.1
summary(cs)
ev1 <- evaluate(cs, mean) # estimates the data with mean
ev1 # bad results
ev2 <- evaluate(cs,median) # estimates the data with median
ev2 # better results because median is robust
savedata(ev1)
# saves the EvaluationList with result as "csim.mean" and without result as
# "csim.mean.comment" in the working directory # of R - "csim" is the
# filename of the Contsimulation object, mean the name of the estimator
rm(ev1)
cload("csim.mean")
# loads the EvaluationList without result - the object is called ev1.comment
ev1.comment
load("csim.mean") # loads the EvaluationList with result
ev1
ElistObj <- EvaluationList(ev1,ev2,name0="myEvalList")
plot(ElistObj,ylim=matrix(c(-0.5,0.5,0.5,4),nrow=2),main=c("location","scale"))
plot(ElistObj,ylim=c(-0.5,0.5),main=c("location"),runs0=3:12,dims0=1,evals0=2)
ElistObj
summary(ElistObj)

[Package distrTEst version 2.2 Index]