ExpParameter-class(distr)	R Documentation

Class "ExpParameter"

Description

The parameter of an exponential distribution, used by Exp-class and DExp-class

Objects from the Class

Objects can be created by calls of the form new("ExpParameter", rate). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Exp is instantiated.

Slots

rate

Object of class "numeric": the rate of an exponential distribution

name

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "ExpParameter"): initialize method

rate

signature(object = "ExpParameter"): returns the slot rate of the parameter of the distribution

rate<-

signature(object = "ExpParameter"): modifies the slot rate of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Examples


W <- new("ExpParameter", rate = 1)
rate(W) # rate of this distribution is 1.
rate(W) <- 2 # rate of this distribution is now 2.

[Package distr version 2.2 Index]