FParameter-class(distr)	R Documentation

Class 'FParameter'

Description

The parameter of a F distribution, used by Fd-class

Objects from the Class

Objects can be created by calls of the form new("FParameter", df1, df2, ncp). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Fd is instantiated.

Slots

df1:

Object of class "numeric": the degrees of freedom of the nominator of an F distribution

df2:

Object of class "numeric": the degrees of freedom of the denominator of an F distribution

ncp:

Object of class "numeric": the noncentrality parameter of an F distribution

name:

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "FParameter"): initialize method

df1

signature(object = "FParameter"): returns the slot df1 of the parameter of the distribution

df1<-

signature(object = "FParameter"): modifies the slot df1 of the parameter of the distribution

df2

signature(object = "FParameter"): returns the slot df2 of the parameter of the distribution

df2<-

signature(object = "FParameter"): modifies the slot df2 of the parameter of the distribution

ncp

signature(object = "FParameter"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "FParameter"): modifies the slot ncp of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Fd-class Parameter-class

Examples


W <- new("FParameter", df1 = 1, df2 = 1, ncp = 0)
df2(W) # df2 of this distribution is 1.
df2(W) <- 2 # df2 of this distribution is now 2.

[Package distr version 2.2 Index]