Fd-class(distr)	R Documentation

Class 'Fd'

Description

The F distribution with df1 = n1, by default = 1, and df2 = n2, by default = 1, degrees of freedom has density

d(x) = Gamma((n1 + n2)/2) / (Gamma(n1/2) Gamma(n2/2)) (n1/n2)^(n1/2) x^(n1/2 - 1) (1 + (n1/n2) x)^-(n1 + n2)/2

for x > 0.

C.f. rf

Objects from the Class

Objects can be created by calls of the form Fd(df1, df2). This object is a F distribution.

Slots

img:

Object of class "Reals": The space of the image of this distribution has got dimension 1 and the name "Real Space".

param:

Object of class "FParameter": the parameter of this distribution (df1 and df2), declared at its instantiation

r:

Object of class "function": generates random numbers (calls function rf)

d:

Object of class "function": density function (calls function df)

p:

Object of class "function": cumulative function (calls function pf)

q:

Object of class "function": inverse of the cumulative function (calls function qf)

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "AbscontDistribution".

Methods

initialize

signature(.Object = "Fd"): initialize method

df1

signature(object = "Fd"): returns the slot df1 of the parameter of the distribution

df1<-

signature(object = "Fd"): modifies the slot df1 of the parameter of the distribution

df2

signature(object = "Fd"): returns the slot df2 of the parameter of the distribution

df2<-

signature(object = "Fd"): modifies the slot df2 of the parameter of the distribution

Ad hoc methods

An ad hoc method is provided for slot d if ncp!=0.
For R Version <2.3.0 ad hoc methods are provided for slots q, r if ncp!=0; for R Version >=2.3.0 the methods from package stats are used.

Note

It is the distribution of the ratio of the mean squares of n1 and n2 independent standard normals, and hence of the ratio of two independent chi-squared variates each divided by its degrees of freedom. Since the ratio of a normal and the root mean-square of m independent normals has a Student's t_m distribution, the square of a t_m variate has a F distribution on 1 and m degrees of freedom.

The non-central F distribution is again the ratio of mean squares of independent normals of unit variance, but those in the numerator are allowed to have non-zero means and ncp is the sum of squares of the means.

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

FParameter-class AbscontDistribution-class Reals-class rf

Examples


F <- Fd(df1 = 1, df2 = 1) # F is a F distribution with df=1 and df2=1.
r(F)(1) # one random number generated from this distribution, e.g. 29.37863
d(F)(1) # Density of this distribution is 0.1591549 for x=1 .
p(F)(1) # Probability that x<1 is 0.5.
q(F)(.1) # Probability that x<0.02508563 is 0.1.
df1(F) # df1 of this distribution is 1.
df1(F) <- 2 # df1 of this distribution is now 2.
Fn <- Fd(df1 = 1, df2 = 1, ncp = 0.5) 
  # Fn is a F distribution with df=1, df2=1 and ncp =0.5.
d(Fn)(1) ## from R 2.3.0 on ncp no longer ignored...

[Package distr version 2.1 Index]