GParetoParameter-class(distrEx)	R Documentation

Parameter of generalized Pareto distributions

Description

The class of the parameter of generalized Pareto distribution.

Objects from the Class

Objects can be created by calls of the form new("GParetoParameter", ...).

Slots

loc:

real number: location parameter of a generalized Pareto distribution.

scale:

real number: scale parameter of a generalized Pareto distribution.

shape:

real number: shape parameter of a generalized Pareto distribution.

name:

default name is “parameter of a GPareto distribution”.

Extends

Class "Parameter", directly.
Class "OptionalParameter", by class "Parameter".

Methods

loc

signature(object = "GParetoParameter"): access method for slot loc.

location

signature(object = "GParetoParameter"): alias to loc, to support argument naming of package VGAM.

scale

signature(object = "GParetoParameter"): access method for slot scale.

shape

signature(object = "GParetoParameter"): access method for slot shape.

loc<-

signature(object = "GParetoParameter"): replace method for slot loc.

location<-

signature(object = "GParetoParameter"): alias to loc<-, to support argument naming of package VGAM.

shape<-

signature(object = "GParetoParameter"): replace method for slot shape.

shape<-

signature(object = "GParetoParameter"): replace method for slot shape.

Author(s)

Nataliya Horbenko Nataliya.Horbenko@itwm.fraunhofer.de

See Also

GPareto-class, Parameter-class

Examples


new("GParetoParameter")

[Package distrEx version 2.2 Index]