Gammad-class(distr)	R Documentation

Class 'Gammad'

Description

The Gammad distribution with parameters shape = a, by default = 1, and scale = s, by default = 1, has density

d(x)= 1/(s^a Gamma(a)) x^(a-1) e^-(x/s)

for x > 0, a > 0 and s > 0. The mean and variance are E(X) = a*s and Var(X) = a*s^2. C.f. rgamma

Objects from the Class

Objects can be created by calls of the form Gammad(scale, shape). This object is a gamma distribution.

Slots

img:

Object of class "Reals": The space of the image of this distribution has got dimension 1 and the name "Real Space".

param:

Object of class "GammaParameter": the parameter of this distribution (scale and shape), declared at its instantiation

r:

Object of class "function": generates random numbers (calls function rgamma)

d:

Object of class "function": density function (calls function dgamma)

p:

Object of class "function": cumulative function (calls function pgamma)

q:

Object of class "function": inverse of the cumulative function (calls function qgamma)

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Extends

Class "ExpOrGammaOrChisq", directly.
Class "AbscontDistribution", by class "ExpOrGammaOrChisq".
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "UnivariateDistribution".

Methods

initialize

signature(.Object = "Gammad"): initialize method

scale

signature(object = "Gammad"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "Gammad"): modifies the slot scale of the parameter of the distribution

shape

signature(object = "Gammad"): returns the slot shape of the parameter of the distribution

shape<-

signature(object = "Gammad"): modifies the slot shape of the parameter of the distribution

+

signature(e1 = "Gammad", e2 = "Gammad"): For the Gamma distribution we use its closedness under convolutions.

*

signature(e1 = "Gammad", e2 = "numeric"): For the Gamma distribution we use its closedness under positive scaling transformations.

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

GammaParameter-class AbscontDistribution-class Reals-class rgamma

Examples


G <- Gammad(scale=1,shape=1) # G is a gamma distribution with scale=1 and shape=1.
r(G)(1) # one random number generated from this distribution, e.g. 0.1304441
d(G)(1) # Density of this distribution is 0.3678794 for x=1.
p(G)(1) # Probability that x<1 is 0.6321206.
q(G)(.1) # Probability that x<0.1053605 is 0.1.
scale(G) # scale of this distribution is 1.
scale(G) <- 2 # scale of this distribution is now 2.

[Package distr version 2.1 Index]