Geom-class(distr)	R Documentation

Class 'Geom'

Description

The geometric distribution with prob = p has density

p(x) = p (1-p)^x

for x = 0, 1, 2, ...

C.f. rgeom

Objects from the Class

Objects can be created by calls of the form Geom(prob). This object is a geometric distribution.

Slots

img:

Object of class "Naturals": The space of the image of this distribution has got dimension 1 and the name "Natural Space".

param:

Object of class "NbinomParameter": the parameter of this distribution (prob), declared at its instantiation (size=1)

r:

Object of class "function": generates random numbers (calls function rgeom)

d:

Object of class "function": density function (calls function dgeom)

p:

Object of class "function": cumulative function (calls function pgeom)

q:

Object of class "function": inverse of the cumulative function (calls function qgeom). The quantile is defined as the smallest value x such that F(x) >= p, where F is the distribution function.

support:

Object of class "numeric": a (sorted) vector containing the support of the discrete density function

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Extends

Class "DiscreteDistribution", directly.
Class "Nbinom", directly.
Class "UnivariateDistribution", by class "DiscreteDistribution".
Class "Distribution", by class "DiscreteDistribution".

Contains-Relations

By means of a contains argument in the class declaration, R ``knows'' that a distribution object obj of class "Geom" also is a negative Binomial distribution with parameters size = 1, prob = prob(obj)

Methods

initialize

signature(.Object = "Geom"): initialize method

prob

signature(object = "Geom"): returns the slot prob of the parameter of the distribution

prob<-

signature(object = "Geom"): modifies the slot prob of the parameter of the distribution

Note

Working with a computer, we use a finite interval as support which carries at least mass 1-getdistrOption("TruncQuantile").

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Nbinom-class GeomParameter-class DiscreteDistribution-class Naturals-class rgeom

Examples


G <- Geom(prob = 0.5) # G is a geometric distribution with prob = 0.5.
r(G)(1) # one random number generated from this distribution, e.g. 0
d(G)(1) # Density of this distribution is 0.25 for x = 1.
p(G)(1) # Probability that x<1 is 0.75.
q(G)(.1) # x = 0 is the smallest value x such that p(G)(x) >= 0.1.
prob(G) # prob of this distribution is 0.5.
prob(G) <- 0.6 # prob of this distribution is now 0.6.
as(G,"Nbinom")
G+G+G

[Package distr version 2.1 Index]