Gumbel-class(distrEx)	R Documentation

Gumbel distribution

Description

The Gumbel cumulative distribution function with location parameter loc = mu and scale parameter scale = sigma is

F(x) = exp(-exp[-(x-mu)/sigma])

for all real x, where sigma > 0; c.f. rgumbel. This distribution is also known as extreme value distribution of type I; confer Chapter~22 of Johnson et al. (1995).

Objects from the Class

Objects can be created by calls of the form new("Gumbel", loc, scale). More frequently they are created via the generating function Gumbel.

Slots

img:

Object of class "Reals".

param:

Object of class "GumbelParameter".

r:

rgumbel

d:

dgumbel

p:

pgumbel

q:

qgumbel

gaps:

(numeric) matrix or NULL

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "AbscontDistribution".

Methods

initialize

signature(.Object = "Gumbel"): initialize method.

loc

signature(object = "Gumbel"): wrapped access method for slot loc of slot param.

scale

signature(x = "Gumbel"): wrapped access method for slot scale of slot param.

loc<-

signature(object = "Gumbel"): wrapped replace method for slot loc of slot param.

scale<-

signature(x = "Gumbel"): wrapped replace method for slot scale of slot param.

+

signature(e1 = "Gumbel", e2 = "numeric"): result again of class "Gumbel"; exact.

*

signature(e1 = "Gumbel", e2 = "numeric"): result again of class "Gumbel"; exact.

E

signature(object = "Gumbel", fun = "missing", cond = "missing"): exact evaluation of expectation using explicit expressions.

var

signature(x = "Gumbel"): exact evaluation of expectation using explicit expressions.

skewness

signature(x = "Gumbel"): exact evaluation of expectation using explicit expressions.

kurtosis

signature(x = "Gumbel"): exact evaluation of expectation using explicit expressions.

median

signature(x = "Gumbel"): exact evaluation of expectation using explicit expressions.

IQR

signature(x = "Gumbel"): exact evaluation of expectation using explicit expressions.

Note

This class is based on the code provided by the package evd.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Johnson et al. (1995) Continuous Univariate Distributions. Vol. 2. 2nd ed. New York: Wiley.

See Also

rgumbel, AbscontDistribution-class

Examples


(G1 <- new("Gumbel", loc = 1, scale = 2))
plot(G1)
loc(G1)
scale(G1)
loc(G1) <- -1
scale(G1) <- 2
plot(G1)

[Package distrEx version 2.1 Index]