HyperParameter-class(distr)	R Documentation

Class 'HyperParameter'

Description

The parameter of a hypergeometric distribution, used by Hyper-class

Objects from the Class

Objects can be created by calls of the form new("HyperParameter", k, m, n). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Hyper is instantiated.

Slots

k:

Object of class "numeric": k of a hypergeometric distribution

m:

Object of class "numeric": m of a hypergeometric distribution

n:

Object of class "numeric": n of a hypergeometric distribution

name:

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "HyperParameter"): initialize method

k

signature(object = "HyperParameter"): returns the slot k of the parameter of the distribution

k<-

signature(object = "HyperParameter"): modifies the slot k of the parameter of the distribution

m

signature(object = "HyperParameter"): returns the slot m of the parameter of the distribution

m<-

signature(object = "HyperParameter"): modifies the slot m of the parameter of the distribution

n

signature(object = "HyperParameter"): returns the slot n of the parameter of the distribution

n<-

signature(object = "HyperParameter"): modifies the slot n of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Hyper-class Parameter-class

Examples


W <- new("HyperParameter",k=3, m=3, n=3)
m(W) # m of this distribution is 3.
m(W) <- 2 # m of this distribution is now 2.

[Package distr version 2.0 Index]