illustrateCLT(distrTeach)	R Documentation

Functions for Illustrating the CLT

Description

Functions for generating a sequence of plots of the density and cdf of the consecutive standardized and centered sums of iid r.v. distributed according to a prescribed discrete or absolutely continuous distribution compared to the standard normal — uses the generic function plotCLT.

Usage

illustrateCLT(Distr, len, sleep = 0)
illustrateCLT.tcl(Distr, k, Distrname)

Arguments

Distr 	 object of class "AbscontDistribution", "LatticeDistribution" or "DiscreteDistribution": distribution of the summands

len 	 integer: up to which number of summands plots are generated

k 	 integer: number of summands for which a plot is to be generated

Distrname 	 character: name of the summand distribution to be used as title in the plot

sleep 	 numeric: pause in seconds between subsequent plots

Details

illustrateCLT generates a sequence of plots, while illustrateCLT.tcl may be used with Tcl/Tk-widgets as in demo illustCLT_tcl.R.

Value

void

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

plotCLT

Examples


distroptions("DefaultNrFFTGridPointsExponent" = 13)
illustrateCLT(Distr = Unif(), len = 20)
distroptions("DefaultNrFFTGridPointsExponent" = 12)
illustrateCLT(Distr = Pois(lambda = 2), len = 20)
distroptions("DefaultNrFFTGridPointsExponent" = 13)
illustrateCLT(Distr = Pois(lambda = 2)+Unif(), len = 20)
illustrateCLT.tcl(Distr = Unif(), k = 4, "Unif()")

[Package distrTeach version 2.1 Index]