Integer-class(distr)	R Documentation

Internal Class 'Integer'

Description

For the ease of method dispatch, there is an internal S4 class Integer, which is a subclass of numeric and has a straightforward validity method.

Objects from the Class

new("Integer",

Slots

.Data:

Object of class "numeric"

Extends

Class "numeric", from data part. Class "vector", by class "numeric", distance 2.

Methods

coerce

signature(from = "numeric", to = "Integer"): create a "Integer" object from a "numeric" vector.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

numeric, vector

[Package distr version 2.0.4 Index]