InternalReturnClasses-class(distrMod)	R Documentation

Internal return classes for generating functions

Description

internal return classes for generating functions 'L2ParamFamily' and 'L2LocationFamily' (and friends); used for particular method dispatch only

Described classes

In this file we describe classes BinomFamily, PoisFamily, GammaFamily BetaFamily ``extending'' (no new slots!) class L2ParamFamily, classes NormLocationFamily and GumbelLocationFamily, ``extending'' (no new slots!) class "L2LocationFamily", classes NormScaleFamily, ExpScaleFamily, and LnormScaleFamily ``extending'' (no new slots!) class "L2ScaleFamily", and classes CauchyLocationScaleFamily and NormLocationScaleFamily, ``extending'' (no new slots!) class "L2LocationScaleFamily".

Objects from these classes

Objects are only generated internally by the mentioned generating functions.

Slots

name:

[inherited from class "ProbFamily"] object of class "character": name of the family.

distribution:

[inherited from class "ProbFamily"] object of class "Distribution": member of the family.

distrSymm:

[inherited from class "ProbFamily"] object of class "DistributionSymmetry": symmetry of distribution.

param:

[inherited from class "ParamFamily"] object of class "ParamFamParameter": parameter of the family.

fam.call:

[inherited from class "ParamFamily"] object of class "call": call by which parametric family was produced.

makeOKPar:

[inherited from class "ParamFamily"] object of class "function": has argument param — the (total) parameter, returns valid parameter; used if optim resp. optimize— try to use ``illegal'' parameter values; then makeOKPar makes a valid parameter value out of the illegal one.

startPar:

[inherited from class "ParamFamily"] object of class "function": has argument x — the data, returns starting parameter for optim resp. optimize— a starting estimator in case parameter is multivariate or a search interval in case parameter is univariate.

modifyParam:

[inherited from class "ParamFamily"] object of class "function": mapping from the parameter space (represented by "param") to the distribution space (represented by "distribution").

props:

[inherited from class "ProbFamily"] object of class "character": properties of the family.

L2deriv:

[inherited from class "L2ParamFamily"] object of class "EuclRandVariable": L2 derivative of the family.

L2deriv.fct:

[inherited from class "L2ParamFamily"] object of class "function": mapping from the parameter space (argument param of class "ParamFamParameter") to a mapping from observation x to the value of the L2derivative; L2deriv.fct is then used from observation x to value of the L2derivative; L2deriv.fct is used by modifyModel to move the L2deriv according to a change in the parameter

L2derivSymm:

[inherited from class "L2ParamFamily"] object of class "FunSymmList": symmetry of the maps included in L2deriv.

L2derivDistr:

[inherited from class "L2ParamFamily"] object of class "UnivarDistrList": list which includes the distribution of L2deriv.

L2derivDistrSymm:

[inherited from class "L2ParamFamily"] object of class "DistrSymmList": symmetry of the distributions included in L2derivDistr.

FisherInfo.fct:

[inherited from class "L2ParamFamily"] object of class "function": mapping from the parameter space (argument param of class "ParamFamParameter") to the set of positive semidefinite matrices; FisherInfo.fct is used by modifyModel to move the Fisher information according to a change in the parameter

FisherInfo:

[inherited from class "L2ParamFamily"] object of class "PosDefSymmMatrix": Fisher information of the family.

LogDeriv:

(only loc/scale classes)[inherited from class "L2GroupParamFamily"] object of class "function": has argument x; the negative logarithmic derivative of the density of the model distribution at the "standard" parameter value.

locscalename:

(only loc/scale classes)[inherited from class "L2LocationScaleUnion"] object of class "character": names of location and scale parameter

Extends

Classes BinomFamily, PoisFamily, GammaFamily BetaFamily ``extend'' (no new slots!):
Class "L2ParamFamily", directly.
Class "ParamFamily", by class "L2ParamFamily".
Class "ProbFamily", by class "ParamFamily".
Classes NormLocationFamily and GumbelLocationFamily, ``extend'' (no new slots!):
Class "L2LocationFamily", directly.
Class "L2LocationScaleUnion", by class "L2LocationFamily".
Class "L2GroupParamFamily", by class "L2LocationScaleUnion".
Class "L2ParamFamily", directly.
Class "ParamFamily", by class "L2ParamFamily".
Class "ProbFamily", by class "ParamFamily".
NormScaleFamily, ExpScaleFamily, and LnormScaleFamily ``extend'' (no new slots!):
Class "L2ScaleFamily", directly.
Class "L2LocationScaleUnion", by class "L2ScaleFamily".
Class "L2GroupParamFamily", by class "L2LocationScaleUnion".
Class "L2ParamFamily", directly.
Class "ParamFamily", by class "L2ParamFamily".
Class "ProbFamily", by class "ParamFamily".
CauchyLocationScaleFamily and NormLocationScaleFamily ``extend'' (no new slots!):
Class "L2LocationScaleFamily", directly.
Class "L2LocationScaleUnion", by class "L2LocationScaleFamily".
Class "L2GroupParamFamily", by class "L2LocationScaleUnion".
Class "L2ParamFamily", directly.
Class "ParamFamily", by class "L2ParamFamily".
Class "ProbFamily", by class "ParamFamily".

Methods

not yet done...

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

numeric-class, L2ParamFamily-class, L2GroupParamFamily-class, L2LocationFamily-class, L2ScaleFamily-class, L2LocationScaleFamily-class,

[Package distrMod version 2.2 Index]