KolmogorovDist(distrMod)	R Documentation

Methods for the computation of the Kolmogorov distance between (empirical) data and distributions

Description

Generic function for the computation of the Kolmogorov distance between (empirical) data and distributions.

Usage


## S4 method for signature 'numeric,
##   UnivariateDistribution':
KolmogorovDist(e1, e2)
## S4 method for signature 'UnivariateDistribution,
##   numeric':
KolmogorovDist(e1, e2)

Arguments

e1 	 object of class "Distribution" or class "numeric"

e2 	 object of class "Distribution" or class "numeric"

Value

Kolmogorov distance of e1 and e2.

Methods

e1 = "numeric", e2 = "UnivariateDistribution":

Kolmogorov distance between (empirical) data and a univariate distribution. The computation is based on ks.test.

e1 = "UnivariateDistribution", e2 = "numeric":

Kolmogorov distance between (empirical) data and a univariate distribution. The computation is based on ks.test.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Huber, P.J. (1981) Robust Statistics. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

See Also

KolmogorovDist, ks.test

Examples


KolmogorovDist(Norm(), rnorm(100))
KolmogorovDist((rbinom(50, size = 20, prob = 0.5)-10)/sqrt(5), Norm())
KolmogorovDist(rbinom(50, size = 20, prob = 0.5), Binom(size = 20, prob = 0.5))

[Package distrMod version 2.0 Index]