L2LocationFamily(distrMod)	R Documentation

Generating function for L2LocationFamily-class

Description

Generates an object of class "L2LocationFamily".

Usage


L2LocationFamily(loc = 0, name, centraldistribution = Norm(),
                 locname = "loc", modParam, LogDeriv,  
                 L2derivDistr.0, FisherInfo.0, distrSymm, L2derivSymm, 
                 L2derivDistrSymm, trafo, .returnClsName = NULL)

Arguments

loc 	 numeric: location parameter of the model.

name 	 character: name of the parametric family.

centraldistribution 	 object of class "AbscontDistribution"; we assume from the beginning, that centraldistribution is symmetric about its median.

modParam 	 optional function: mapping from the parameter space (represented by "param") to the distribution space (represented by "distribution").

locname 	 a character vector of length 1 containing the name of the location parameter

LogDeriv 	 function with argument x: the negative logarithmic derivative of the density of the central distribution; if missing, it is determined numerically using numeric differentiation.

L2derivDistr.0 	 object of class "UnivariateDistribution": distribution of the L2derivative at the central distribution

FisherInfo.0 	 object of class "PosSemDefSymmMatrix": Fisher information of the model at the "standard" parameter value

distrSymm 	 object of class "DistributionSymmetry": symmetry of distribution.

L2derivSymm 	 object of class "FunSymmList": symmetry of the maps contained in L2deriv

L2derivDistrSymm 	 object of class "DistrSymmList": symmetry of the distributions contained in L2derivDistr

trafo 	 matrix or function in param: transformation of the parameter

.returnClsName 	 the class name of the return value; by default this argument is NULL whereupon the return class will be L2LocationScaleFamily; but, internally, this generating function is also used to produce objects of class Classes NormLocationFamily and GumbelLocationFamily.

Details

If name is missing, the default “L2 location family” is used. The function modParam is optional. If it is missing, it is constructed from centraldistribution using the location structure of the model. Slot param is filled accordingly with the argument trafo passed to L2LocationFamily. In case L2derivDistr.0 is missing, L2derivDistr is computed via imageDistr, else L2derivDistr is assigned L2derivDistr.0, coerced to "UnivariateDistributionList". In case FisherInfo.0 is missing, Fisher information is computed from L2deriv using E. If distrSymm is missing, it is set to symmetry about loc. If L2derivSymm is missing, it is set to no symmetry, and if L2derivDistrSymm is missing, it is set to no symmetry, too.

Value

Object of class "L2LocationFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2LocationFamily-class

Examples


F1 <- L2LocationFamily()
plot(F1)

[Package distrMod version 2.0 Index]