L2LocationScaleFamily(distrMod)	R Documentation

Generating function for L2LocationScaleFamily-class

Description

Generates an object of class "L2LocationScaleFamily".

Usage


L2LocationScaleFamily(loc = 0, scale = 1, name, centraldistribution = Norm(),
                      locscalename = c("loc", "scale"), modParam, LogDeriv,  
                      L2derivDistr.0, FisherInfo.0, distrSymm, L2derivSymm, 
                      L2derivDistrSymm, trafo, ...)


Arguments

loc 	 numeric: location parameter of the model.

scale 	 positive number: scale of the model.

name 	 character: name of the parametric family.

centraldistribution 	 object of class "AbscontDistribution": central distribution; we assume by default, that centraldistribution is symmetric about 0

modParam 	 optional function: mapping from the parameter space (represented by "param") to the distribution space (represented by "distribution").

locscalename 	 a character vector of length 2 containing the names of the location and scale parameter; either unnamed, then order must be c(loc,scale), or named, then names must be "loc" and "scale"

LogDeriv 	 function with argument x: the negative logarithmic derivative of the density of the central distribution; if missing, it is determined numerically using numeric differentiation.

L2derivDistr.0 	 list of length 2 of objects of class "UnivariateDistribution": (marginal) distributions of the coordinates of the L2derivative at the central distribution

FisherInfo.0 	 object of class "PosSemDefSymmMatrix": Fisher information of the model at the "standard" parameter value

distrSymm 	 object of class "DistributionSymmetry": symmetry of distribution.

L2derivSymm 	 object of class "FunSymmList": symmetry of the maps contained in L2deriv

L2derivDistrSymm 	 object of class "DistrSymmList": symmetry of the distributions contained in L2derivDistr

trafo 	 matrix or function in param: transformation of the parameter

... 	 further arguments to be passed to E

Details

If name is missing, the default “L2 location and scale family” is used. The function modParam is optional. If it is missing, it is constructed from centraldistribution using the location and scale structure of the model. Slot param is filled accordingly with the argument trafo passed to L2LocationScaleFamily. In case L2derivDistr.0 is missing, L2derivDistr is computed via imageDistr, else L2derivDistr is assigned L2derivDistr.0, coerced to "UnivariateDistributionList". In case FisherInfo.0 is missing, Fisher information is computed from L2deriv using E. If distrSymm is missing, it is set to symmetry about loc. If L2derivSymm is missing, its location and scale components are set to no symmetry , respectively. if L2derivDistrSymm is missing, its location and scale components are set to no symmetry, respectively.

Value

Object of class "L2LocationScaleFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2LocationScaleFamily-class

Examples


F1 <- L2LocationScaleFamily()
plot(F1)

[Package distrMod version 2.0 Index]