L2ParamFamily-class(distrMod)	R Documentation

L2 differentiable parametric family

Description

Class of L2 differentiable parametric families.

Objects from the Class

Objects can be created by calls of the form new("L2ParamFamily", ...). More frequently they are created via the generating function L2ParamFamily.

Slots

name:

[inherited from class "ProbFamily"] object of class "character": name of the family.

distribution:

[inherited from class "ProbFamily"] object of class "Distribution": member of the family.

distrSymm:

[inherited from class "ProbFamily"] object of class "DistributionSymmetry": symmetry of distribution.

param:

[inherited from class "ParamFamily"] object of class "ParamFamParameter": parameter of the family.

fam.call:

[inherited from class "ParamFamily"] object of class "call": call by which parametric family was produced.

makeOKPar:

[inherited from class "ParamFamily"] object of class "function": has argument param — the (total) parameter, returns valid parameter; used if optim resp. optimize— try to use ``illegal'' parameter values; then makeOKPar makes a valid parameter value out of the illegal one.

startPar:

[inherited from class "ParamFamily"] object of class "function": has argument x — the data, returns starting parameter for optim resp. optimize— a starting estimator in case parameter is multivariate or a search interval in case parameter is univariate.

modifyParam:

[inherited from class "ParamFamily"] object of class "function": mapping from the parameter space (represented by "param") to the distribution space (represented by "distribution").

props:

[inherited from class "ProbFamily"] object of class "character": properties of the family.

L2deriv:

object of class "EuclRandVariable": L2 derivative of the family.

L2deriv.fct:

object of class "function": mapping from the parameter space (argument param of class "ParamFamParameter") to a mapping from observation x to the value of the L2derivative; L2deriv.fct is then used from observation x to value of the L2derivative; L2deriv.fct is used by modifyModel to move the L2deriv according to a change in the parameter

L2derivSymm:

[ object of class "FunSymmList": symmetry of the maps included in L2deriv.

L2derivDistr:

object of class "UnivarDistrList": list which includes the distribution of L2deriv.

L2derivDistrSymm:

object of class "DistrSymmList": symmetry of the distributions included in L2derivDistr.

FisherInfo.fct:

object of class "function": mapping from the parameter space (argument param of class "ParamFamParameter") to the set of positive semidefinite matrices; FisherInfo.fct is used by modifyModel to move the Fisher information according to a change in the parameter

FisherInfo:

object of class "PosDefSymmMatrix": Fisher information of the family.

Extends

Class "ParamFamily", directly.
Class "ProbFamily", by class "ParamFamily".

Methods

L2deriv

signature(object = "L2ParamFamily"): accessor function for L2deriv.

L2deriv

signature(object = "L2ParamFamily", param = "ParamFamParameter"): returns the L2derivative at param, i.e. evaluates slot function L2deriv.fct at param.

L2derivSymm

signature(object = "L2ParamFamily"): accessor function for L2derivSymm.

L2derivDistr

signature(object = "L2ParamFamily"): accessor function for L2derivDistr.

L2derivDistrSymm

signature(object = "L2ParamFamily"): accessor function for L2derivDistrSymm.

FisherInfo

signature(object = "L2ParamFamily"): accessor function for FisherInfo.

FisherInfo

signature(object = "L2ParamFamily", param = "ParamFamParameter"): returns the Fisher Information at param, i.e. evaluates slot function FisherInfo.fct at param.

checkL2deriv

signature(object = "L2ParamFamily"): check centering of L2deriv and compute precision of Fisher information.

E

signature(object = "L2ParamFamily", fun = "EuclRandVariable", cond = "missing"): expectation of fun under the distribution of object.

E

signature(object = "L2ParamFamily", fun = "EuclRandMatrix", cond = "missing"): expectation of fun under the distribution of object.

E

signature(object = "L2ParamFamily", fun = "EuclRandVarList", cond = "missing"): expectation of fun under the distribution of object.

plot

signature(x = "L2ParamFamily"): plot of distribution and L2deriv.

modifyModel

signature(model = "L2ParamFamily", param = "ParamFamParameter"): moves the L2-parametric Family model to parameter param

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2ParamFamily, ParamFamily-class

Examples


F1 <- new("L2ParamFamily")
plot(F1)

[Package distrMod version 2.0 Index]