L2ScaleFamily(distrMod)	R Documentation

Generating function for L2ScaleFamily-class

Description

Generates an object of class "L2ScaleFamily".

Usage


L2ScaleFamily(scale = 1, loc = 0,  name, centraldistribution = Norm(),
              locscalename = c("loc", "scale"), modParam, LogDeriv,  
              L2derivDistr.0, FisherInfo.0, distrSymm, L2derivSymm, 
              L2derivDistrSymm, trafo, .returnClsName = NULL)


Arguments

scale 	 positive number: scale parameter of the model

loc 	 numeric: location parameter of the model

name 	 character: name of the parametric family.

centraldistribution 	 object of class "AbscontDistribution": central distribution; we assume from the beginning, that centraldistribution is symmetric about 0

locscalename 	 a character vector of length 1 or 2 containing the names of the scale resp. of location and scale parameter; if length is 2, locscalename is either unnamed, then order must be c(scale,loc), or named, then names must be "loc" and "scale".

modParam 	 optional function: mapping from the parameter space (represented by "param") to the distribution space (represented by "distribution").

LogDeriv 	 function with argument x: the negative logarithmic derivative of the density of the central distribution; if missing, it is determined numerically using numeric differentiation.

L2derivDistr.0 	 object of class "UnivariateDistribution": distribution of the L2derivative at the central distribution

FisherInfo.0 	 object of class "PosSemDefSymmMatrix": Fisher information of the model at the "standard" parameter value

distrSymm 	 object of class "DistributionSymmetry": symmetry of distribution.

L2derivSymm 	 object of class "FunSymmList": symmetry of the maps contained in L2deriv

L2derivDistrSymm 	 object of class "DistrSymmList": symmetry of the distributions contained in L2derivDistr

trafo 	 matrix or function in param: transformation of the parameter

.returnClsName 	 the class name of the return value; by default this argument is NULL whereupon the return class will be L2ScaleFamily; but, internally, this generating function is also used to produce objects of class NormScaleFamily, ExpScaleFamily, and LnormScaleFamily.

Details

If name is missing, the default “L2 scale family” is used. The function modParam is optional. If it is missing, it is constructed from centraldistribution using the scale structure of the model. Slot param is filled accordingly with the argument trafo passed to L2ScaleFamily. In case L2derivDistr.0 is missing, L2derivDistr is computed via imageDistr, else L2derivDistr is assigned L2derivDistr.0, coerced to "UnivariateDistributionList". In case FisherInfo.0 is missing, Fisher information is computed from L2deriv using E. If distrSymm is missing, it is set to symmetry about loc. If L2derivSymm is missing, it is set to no symmetry, and if L2derivDistrSymm is missing, it is set to no symmetry.

Value

Object of class "L2ScaleFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2ScaleFamily-class

Examples


F1 <- L2ScaleFamily()
plot(F1)

[Package distrMod version 2.0 Index]