LMCondDistribution(distrEx)	R Documentation

Generating function for the conditional distribution of a linear regression model.

Description

Generates an object of class "AbscontCondDistribution" which is the conditional distribution of a linear regression model (given the regressor).

Usage


LMCondDistribution(Error = Norm(), theta = 0, intercept = 0, scale = 1)

Arguments

Error 	 Object of class "AbscontDistribution": error distribution.

theta 	 numeric vector: regression parameter.

intercept 	 real number: intercept parameter.

scale 	 positive real number: scale parameter.

Value

Object of class "AbscontCondDistribution"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

AbscontCondDistribution-class, E-methods

Examples


# normal error distribution
(D1 <- LMCondDistribution(theta = 1)) # corresponds to Norm(cond, 1)
plot(D1)
r(D1)
d(D1)
p(D1)
q(D1)
param(D1)
cond(D1)

d(D1)(0, cond = 1)
d(Norm(mean=1))(0)

E(D1, cond = 1)
E(D1, function(x){x^2}, cond = 2)
E(Norm(mean=2), function(x){x^2})

[Package distrEx version 2.0 Index]