LMParameter-class(distrEx)	R Documentation

Parameter of a linear regression model

Description

Parameter of a linear regression model

y = mu + t(x)theta + sigma u

with intercept mu, regression parameter theta and error scale sigma.

Objects from the Class

Objects can be created by calls of the form new("LMParameter", ...). More frequently they are created via the generating function LMParameter.

Slots

theta

numeric vector: regression parameter.

intercept

real number: intercept parameter.

scale

positive real number: scale paramter.

name

character vector: the default name is “parameter of a linear regression model”.

Extends

Class "Parameter", directly.
Class "OptionalParameter", by class "Parameter".

Methods

show

signature(object = "LMParameter")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Parameter-class,

Examples


  new("LMParameter")

[Package distrEx version 2.2 Index]