LnormParameter-class(distr)	R Documentation

Class "LnormParameter"

Description

The parameter of a log normal distribution, used by Lnorm-class

Objects from the Class

Objects can be created by calls of the form new("LnormParameter", meanlog, sdlog). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Lnorm is instantiated.

Slots

meanlog

Object of class "numeric": the mean of a log normal distribution

sdlog

Object of class "numeric": the sd of a log normal distribution

name

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "LnormParameter"): initialize method

sdlog

signature(object = "LnormParameter"): returns the slot sdlog of the parameter of the distribution

sdlog<-

signature(object = "LnormParameter"): modifies the slot sdlog of the parameter of the distribution

meanlog

signature(object = "LnormParameter"): returns the slot meanlog of the parameter of the distribution

meanlog<-

signature(object = "LnormParameter"): modifies the slot meanlog of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Examples


W <- new("LnormParameter",sdlog=1,meanlog=0)
meanlog(W) # meanlog of this distribution is 0.
meanlog(W) <- 2 # meanlog of this distribution is now 2.

[Package distr version 2.2 Index]