LogisParameter-class(distr)	R Documentation

Class 'LogisParameter'

Description

The parameter of a logistic distribution, used by Logis-class

Objects from the Class

Objects can be created by calls of the form new("LogisParameter", scale, location). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Logis is instantiated.

Slots

scale:

Object of class "numeric": the scale of a logistic distribution

location:

Object of class "numeric": the location of a logistic distribution

name:

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "LogisParameter"): initialize method

location

signature(object = "LogisParameter"): returns the slot location of the parameter of the distribution

location<-

signature(object = "LogisParameter"): modifies the slot location of the parameter of the distribution

scale

signature(object = "LogisParameter"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "LogisParameter"): modifies the slot scale of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@web.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Logis-class Parameter-class

Examples


W <- new("LogisParameter",location=0,scale=1)
scale(W) # scale of this distribution is 1.
scale(W) <- 2 # scale of this distribution is now 2.

[Package distr version 2.0 Index]