MCEstimate-class(distrMod)	R Documentation

MCEstimate-class.

Description

Class of minimum criterion estimates.

Objects from the Class

Objects can be created by calls of the form new("MCEstimate", ...). More frequently they are created via the generating functions MCEstimator, MDEstimator or MLEstimator.

Slots

name:

Object of class "character": name of the estimator.

estimate:

Object of class "ANY": estimate.

estimate.call:

Object of class "call": call by which estimate was produced.

criterion:

Object of class "numeric": minimum value of the considered criterion.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

asvar:

object of class "OptionalMatrix" which may contain the asymptotic (co)variance of the estimator.

samplesize:

object of class "numeric" — the samplesize at which the estimate was evaluated.

nuis.idx:

object of class "OptionalNumeric": indices of estimate belonging to the nuisance part

Extends

Class "Estimate", directly.

Methods

criterion

signature(object = "MCEstimate"): accessor function for slot criterion.

criterion<-

signature(object = "MCEstimate"): replacement function for slot criterion.

show

signature(object = "Estimate")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Estimate-class, MCEstimator, MDEstimator, MLEstimator

Examples


## (empirical) Data
x <- rgamma(50, scale = 0.5, shape = 3)

## parametric family of probability measures
G <- GammaFamily(scale = 1, shape = 2)

MDEstimator(x, G)
MLEstimator(x, G)

[Package distrMod version 2.0 Index]