MDEstimator(distrMod)	R Documentation

Function to compute minimum distance estimators

Description

The function MDEstimator provides a general way to compute minimum distance estimators.

Usage


MDEstimator(x, ParamFamily, distance = KolmogorovDist, dist.name, interval, par, ...)

Arguments

x 	 (empirical) data

ParamFamily 	 object of class "ParamFamily"

distance 	 (generic) function: to compute distance beetween (emprical) data and objects of class "Distribution".

dist.name 	 optional name of distance

interval 	 interval used by optimize

par 	 initial parameter values used by optim; if missing the parameter of ParamFamily are used.

... 	 further arguments to criterion or optimize or optim, respectively.

Details

The argument distance has to be a (generic) function with arguments the empirical data as well as an object of class "Distribution" and possibly ...; e.g. KolmogorovDist (default), TotalVarDist or HellingerDist.

Value

An object of S3-class "MCEstimate" which inherits from class "Estimate", a list with two components

estimate 	 maximum likelihood parameter estimate

criterion 	 minimum value of the distance

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@web.de

References

Huber, P.J. (1981) Robust Statistics. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

See Also

ParamFamily-class, ParamFamily, code{MCEstimator}, fitdistr

Examples


## (empirical) Data
x <- rgamma(50, scale = 0.5, shape = 3)

## parametric family of probability measures
theta <- c(1, 2)
names(theta) <- c("scale", "shape")
G <- ParamFamily(name = "Gamma family",
        param = ParamFamParameter(name = "scale and shape", main = theta),
        distribution = Gammad(scale = 1, shape = 2),
        modifyParam = function(theta){ 
                          Gammad(scale = theta[1], shape = theta[2]) 
                      },
        props = paste(c("The Gamma family is scale invariant via the",
                  "parametrization '(nu,shape)=(log(scale),shape)'"),
                  collapse = " "))

## Kolmogorov(-Smirnov) minimum distance estimator
MDEstimator(x = x, ParamFamily = G, distance = KolmogorovDist)

## von Mises minimum distance estimator with default mu
MDEstimator(x = x, ParamFamily = G, distance = CvMDist)

## von Mises minimum distance estimator with mu = N(0,1)
MDEstimator(x = x, ParamFamily = G, distance = CvMDist, mu = Norm())

## Total variation minimum distance estimator
## gamma distributions are discretized
MDEstimator(x = x, ParamFamily = G, distance = TotalVarDist)
## or smoothing of emprical distribution (takes some time!)
#MDEstimator(x = x, ParamFamily = G, distance = TotalVarDist, asis.smooth.discretize = "smooth")

## Hellinger minimum distance estimator
## gamma distributions are discretized
distroptions(DistrResolution = 1e-8)
MDEstimator(x = x, ParamFamily = G, distance = HellingerDist)
distroptions(DistrResolution = 1e-6) # default
## or smoothing of emprical distribution (takes some time!)
#MDEstimator(x = x, ParamFamily = G, distance = HellingerDist, asis.smooth.discretize = "smooth")

[Package distrMod version 2.0 Index]