MLEstimator(distrMod)	R Documentation

Function to compute maximum likelihood estimators

Description

The function MLEstimator provides a general way to compute maximum likelihood estimators for a given parametric family of probability measures. This is done by calling the function MCEstimator which minimizes the negative log-Likelihood.

Usage


MLEstimator(x, ParamFamily, interval, par, ...)

Arguments

x 	 (empirical) data

ParamFamily 	 object of class "ParamFamily"

interval 	 interval used by optimize

par 	 initial parameter values used by optim; if missing the parameter of ParamFamily are used.

... 	 further arguments to criterion or optimize or optim, respectively.

Details

The function calls MCEstimator using the negative log-likelihood as criterion which should be minimized.

Value

An object of S3-class "MCEstimate" which inherits from class "Estimate", a list with two components

estimate 	 maximum likelihood parameter estimate

criterion 	 minimum value of the negative log-likelihood

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@web.de

See Also

ParamFamily-class, ParamFamily, code{MCEstimator}, fitdistr

Examples


## (empirical) Data
x <- rgamma(50, scale = 0.5, shape = 3)

## parametric family of probability measures
theta <- c(1, 2)
names(theta) <- c("scale", "shape")
G <- ParamFamily(name = "Gamma family",
        param = ParamFamParameter(name = "scale and shape", main = theta),
        distribution = Gammad(scale = 1, shape = 2),
        modifyParam = function(theta){ 
                          Gammad(scale = theta[1], shape = theta[2]) 
                      },
        props = paste(c("The Gamma family is scale invariant via the",
                  "parametrization '(nu,shape)=(log(scale),shape)'"),
                  collapse = " "))

## Maximum likelihood estimator
(res <- MLEstimator(x = x, ParamFamily = G))

## implementation of ML-estimator of package MASS
library(MASS)
(res1 <- fitdistr(x, "gamma"))

## comparison
## shape
res$estimate[2]
## rate
1/res$estimate[1]
## log-likelihood
res1$loglik
## negative log-likelihood
res$criterion

[Package distrMod version 2.0 Index]