Math-methods(distr)	R Documentation

Methods for Functions from group ‘Math’ in Package ‘distr’

Description

Math-methods provide automatical generation of image distributions for random variables transformed by functions from group Math

Methods

Math

signature(x = "AbscontDistribution"): application of a mathematical function from group Math, e.g. sin or exp (including log, log10, gamma, lgamma), to this absolutely continouos distribution

Math

signature(x = "DiscreteDistribution"): application of a mathematical function, e.g. sin or exp (including log, log10, gamma, lgamma), to this discrete distribution

Math

signature(x = "UnivarLebDecDistribution"): application of a mathematical function from group Math, e.g. sin or exp (including log, log10, gamma, lgamma), to this Lebesgue decomposed distribution

Math

signature(x = "UnivarLebDecDistribution"): application of a mathematical function from group Math, e.g. sin or exp (including log, log10, gamma, lgamma), to this distribution of class "AcDcLcDistribution"

abs

signature(x = "AbscontDistribution"): application of function abs to this absolutely continouos distribution; (exactly)

abs

signature(x = "DiscreteDistribution"): application of function abs to this discrete distribution; (exactly)

sign

signature(x = "AbscontDistribution"): application of function abs to this absolutely continouos distribution; (exactly)

sign

signature(x = "DiscreteDistribution"): application of function abs to this discrete continouos distribution; (exactly)

exp

signature(x = "AbscontDistribution"): application of function exp to this absolutely continouos distribution; (exactly)

exp

signature(x = "DiscreteDistribution"): application of function exp to this discrete distribution; (exactly)

log

signature(x = "AbscontDistribution"): application of function log to this absolutely continouos distribution; (exactly for R-version >2.5.1)

log

signature(x = "DiscreteDistribution"): application of function log to this discrete distribution; (exactly for R-version >2.5.1)

[Package distr version 2.1 Index]