Max-methods(distr)	R Documentation

Methods for Function Max in Package ‘distr’

Description

Max-methods

Methods

Max

signature(object = "UnifParameter"): returns the slot Max of the parameter of the distribution

Max<-

signature(object = "UnifParameter"): modifies the slot Max of the parameter of the distribution

Max

signature(object = "Unif"): returns the slot Max of the parameter of the distribution

Max<-

signature(object = "Unif"): modifies the slot Max of the parameter of the distribution

[Package distr version 2.1 Index]