Min-methods(distr)	R Documentation

Methods for Function Min in Package ‘distr’

Description

Min-methods

Methods

Min

signature(object = "UnifParameter"): returns the slot Min of the parameter of the distribution

Min<-

signature(object = "UnifParameter"): modifies the slot Min of the parameter of the distribution

Min

signature(object = "Unif"): returns the slot Min of the parameter of the distribution

Min<-

signature(object = "Unif"): modifies the slot Min of the parameter of the distribution

[Package distr version 2.2 Index]