Minimum-methods(distr)	R Documentation

Methods for functions Minimum and Maximum in Package ‘distr’

Description

Minimum and Maximum-methods

Usage


Minimum(e1, e2, ...)
Maximum(e1, e2, ...) 
## S4 method for signature 'AbscontDistribution,
##   AbscontDistribution':
Minimum(e1,e2, ...)
## S4 method for signature 'DiscreteDistribution,
##   DiscreteDistribution':
Minimum(e1,e2, ...)
## S4 method for signature 'AbscontDistribution, Dirac':
Minimum(e1,e2, 
                   withSimplify = getdistrOption("simplifyD"))
## S4 method for signature 'AcDcLcDistribution,
##   AcDcLcDistribution':
Minimum(e1,e2, 
                   withSimplify = getdistrOption("simplifyD"))
## S4 method for signature 'AcDcLcDistribution,
##   AcDcLcDistribution':
Maximum(e1,e2, 
                   withSimplify = getdistrOption("simplifyD"))
## S4 method for signature 'AbscontDistribution, numeric':
Minimum(e1,e2, ...)
## S4 method for signature 'DiscreteDistribution, numeric':
Minimum(e1,e2, ...)
## S4 method for signature 'AcDcLcDistribution, numeric':
Minimum(e1,e2,
                   withSimplify = getdistrOption("simplifyD"))
## S4 method for signature 'AcDcLcDistribution, numeric':
Maximum(e1,e2, 
                   withSimplify = getdistrOption("simplifyD"))

Arguments

e1 	 distribution object

e2 	 distribution object or numeric

... 	 further arguments (to be able to call various methods with the same arguments

withSimplify 	 logical; is result to be piped through a call to simplifyD?

Value

the corresponding distribution of the minimum / maximum

Methods

Minimum

signature(e1 = "AbscontDistribution", e2 = "AbscontDistribution"): returns the distribution of min(X1,X2), if X1,X2 are independent and distributed according to e1 and e2 respectively; the result is again of class "AbscontDistribution"

Minimum

signature(e1 = "DiscreteDistribution", e2 = "DiscreteDistribution"): returns the distribution of min(X1,X2), if X1,X2 are independent and distributed according to e1 and e2 respectively; the result is again of class "DiscreteDistribution"

Minimum

signature(e1 = "AbscontDistribution", e2 = "Dirac"): returns the distribution of min(X1,X2), if X1,X2 are distributed according to e1 and e2 respectively; the result is of class "UnivarLebDecDistribution"

Minimum

signature(e1 = "AcDcLcDistribution", e2 = "AcDcLcDistribution"): returns the distribution of min(X1,X2), if X1,X2 are distributed according to e1 and e2 respectively; the result is of class "UnivarLebDecDistribution"

Minimum

signature(e1 = "AcDcLcDistribution", e2 = "numeric"): if e2 = n, returns the distribution of min(X1,X2,...,Xn), if X1,X2, ..., Xn are i.i.d. according to e1; the result is of class "UnivarLebDecDistribution"

Maximum

signature(e1 = "AcDcLcDistribution", e2 = "AcDcLcDistribution"): returns the distribution of max(X1,X2), if X1,X2 are distributed according to e1 and e2 respectively; translates into -Minimum(-e1,-e2); the result is of class "UnivarLebDecDistribution"

Maximum

signature(e1 = "AcDcLcDistribution", e2 = "numeric"): if e2 = n, returns the distribution of max(X1,X2,...,Xn), if X1,X2, ..., Xn are i.i.d. according to e1; translates into -Minimum(-e1,e2); the result is of class "UnivarLebDecDistribution"

See Also

Huberize, Truncate

Examples


plot(Minimum(Norm(),Pois()))
plot(Maximum(Unif(0,1), Minimum(Unif(0,1), Unif(0,1))))
plot(Minimum(Exp(4),4))

[Package distr version 2.0 Index]