MultivarDistrList(distrEllipse)	R Documentation

Generating function for MultivarDistrList-class

Description

Generates an object of class "MultivarDistrList".

Usage

MultivarDistrList(..., Dlist)

Arguments

... 	 Objects of class "MultivariateDistribution" (or subclasses)

Dlist 	 an optional list or object of class "MultivarDistrList"; if not missing it is appended to argument ...; this way MultivarMixingDistribution may also be called with a list (or "MultivarDistrList"-object) as argument as suggested in an e-mail by Krunoslav Sever (thank you!)

Value

Object of class "MVDistrList" or of class "UnivarDistrList", hence of class union "MultivarDistrList"

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

DistrList-class, MultivarDistrList-class, MultivarDistrList

Examples


(DL1 <- MultivarDistrList(Norm(), Exp(), Pois()))
(DL2 <- MultivarDistrList(MVNorm(),
                          EllipticalDistribution(radDistr=Exp(), loc=c(1,2),
                          scale=diag(c(3,1))),MVt()))

[Package distrEllipse version 0.1 Index]