MultivarDistrList-class(distrEllipse)	R Documentation

List of multivariate distributions

Description

Create a list of multivariate distributions

Objects from the Class

Objects can be created by calls of the form new("MVDistrList", ...). More frequently they are created via the generating function MultivarDistrList.

Slots

.Data:

Object of class "list". A list of multivariate distributions of the same dimension.

Extends

Class "DistrList", directly.
Class "list", by class "DistrList".
Class "vector", by class "DistrList".

Methods

coerce

signature(from = "MultivariateDistribution", to = "MultivarDistrList"): create a MultivarDistrList object from a univariate distribution

dimension

dim of the range space.

dim

synonym to dimension.

Details

In fact, class "MultivarDistrList" is an inbetween class between class "DistrList" and class "UnivarDistrList", which is a case for setIs, but we would have to modify the metadata information in package distr to realize this. So we introduce a new (sister) class "MVDistrList" which implements strictly lists of multivariate distributions, and which together with "UnivarDistrList" is a subclass of the common class union class "MultivarDistrList".

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

MultivarDistrList, DistrList-class, MultivariateDistribution-class

Examples


(DL1 <- MultivarDistrList(Norm(), Exp(), Pois()))
(DL2 <- MultivarDistrList(MVNorm(),
                          EllipticalDistribution(radDistr=Exp(), loc=c(1,2),
                          scale=diag(c(3,1))),MVt()))


[Package distrEllipse version 0.1 Index]