MultivarMixingDistribution(distrEllipse)	R Documentation

Generating function for Class 'MultivarMixingDistribution'

Description

Generates an object of class "MultivarMixingDistribution".

Usage

MultivarMixingDistribution(..., Dlist, mixCoeff, 
                                withSimplify = getdistrOption("simplifyD"))

Arguments

... 	 Objects of class "MultivariateDistribution" (or subclasses)

Dlist 	 an optional list or object of class "MultivarDistrList"; if not missing it is appended to argument ...; this way MultivarMixingDistribution may also be called with a list (or "MultivarDistrList"-object) as argument as suggested in an e-mail by Krunoslav Sever (thank you!)

mixCoeff 	 Objects of class "numeric" : a vector of probabilities for the mixing components (must be of same length as arguments in ...).

withSimplify 	 "logical": shall the return value be piped through a call to simplifyD?

Details

If mixCoeff is missing, all elements in ... are equally weighted.

Value

Object of class "MultivarMixingDistribution", or if argument withSimplify is TRUE and the resulting object would have one mixing component with probability (almost) 1, MultivarMixingDistribution will return this component.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

MultivarMixingDistribution-class, simplifyD

Examples


mylist <- MultivarMixingDistribution(Binom(3,.3), Dirac(2), Norm(), 
          mixCoeff=c(1/4,1/5,11/20))

[Package distrEllipse version 0.1 Index]