MultivarMixingDistribution-class(distrEllipse)	R Documentation

Class "MultivarMixingDistribution"

Description

MultivarMixingDistribution-class is a class to formalize multivariate mixing distributions; it is a subclass to class MultivariateDistribution.

Objects from the Class

Objects can be created by calls of the form new("MultivarMixingDistribution", ...). More frequently they are created via the generating function .

Slots

mixCoeff

Object of class "numeric": a vector of probabilities for the mixing components.

mixDistr

Object of class "MultivarDistrList": a list of multivariate distributions containing the mixing components; must be of same length as mixCoeff.

img

Object of class "Reals": the space of the image of this distribution which has dimension 1 and the name "Real Space"

param

Object of class "Parameter": the parameter of this distribution, having only the slot name "Parameter of a discrete distribution"

r

Object of class "function": generates random numbers

d

fixed to NULL

p

Object of class "OptionalFunction": if non-null cumulative distribution function

q

Object of class "OptionalFunction": if non-null quantile function

.withArith

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim

logical: used internally to issue warnings as to accuracy

.logExact

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Symmetry

object of class "DistributionSymmetry"; used internally to avoid unnecessary calculations.

Extends

Class "MultivariateDistribution" class "Distribution" by class "MultivariateDistribution".

Methods

show

signature(object = "MultivarMixingDistribution") prints the object

mixCoeff<-

signature(object = "MultivarMixingDistribution") replaces the corresponding slot

mixCoeff

signature(object = "MultivarMixingDistribution") returns the corresponding slot

mixDistr<-

signature(object = "MultivarMixingDistribution") replaces the corresponding slot

mixDistr

signature(object = "MultivarMixingDistribution") returns the corresponding slot

support

signature(object = "MultivarMixingDistribution") returns the corresponding slot

gaps

signature(object = "MultivarMixingDistribution") returns the corresponding slot

.logExact

signature(object = "Distribution"): returns slot .logExact if existing; else tries to convert the object to a newer version of its class by conv2NewVersion and returns the corresponding slot of the converted object.

.lowerExact

signature(object = "Distribution"): returns slot .lowerExact if existing; else tries to convert the object to a newer version of its class by conv2NewVersion and returns the corresponding slot of the converted object.

Symmetry

returns slot Symmetry if existing; else tries to convert the object to a newer version of its class by conv2NewVersion and returns the corresponding slot of the converted object.

plot

signature(x = "MultivarMixingDistribution", y = "missing"): plot for an spherically symmetric distribution; see .

E

corresponding expectation — see E.

dimension

dim of the range space.

dim

synonym to dimension.

show

signature(object = "MultivarMixingDistribution"): show method for spherically symmetric distributions.

showobj

signature(object = "MultivarMixingDistribution"): showobj method for spherically symmetric distributions.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Parameter-class, MultivariateDistribution-class, LatticeDistribution-class, AbscontDistribution-class, simplifyD, flat.mix

Examples


mylist <- MultivarMixingDistribution(Binom(3,.3), Dirac(2), Norm(), 
          mixCoeff=c(1/4,1/5,11/20))
mylist2 <- MultivarMixingDistribution(Binom(3,.3), mylist, 
          mixCoeff=c(.3,.7))
mylist2
p(mylist)(0.3)          
mixDistr(mylist2)

##multivariate
E1 <- diag(1,2)%*%EllipticalDistribution(radDistr=Gammad())+c(1,2)
mylistD <- MultivarMixingDistribution(MVNorm(), E1, MVt(),
          mixCoeff=c(1/4,1/5,11/20))
mylistD2 <- MultivarMixingDistribution(E1+c(-2,2), mylistD,
          mixCoeff=c(.3,.7))
mylistD2
p(mylistD)
mixDistr(mylistD2)

[Package distrEllipse version 0.2 Index]