MultivariateDistribution-class(distrEx)	R Documentation

Multivariate Distributions

Description

The class of multivariate distributions. One has at least to specify the image space of the distribution and a function generating (pseudo-)random numbers. The slot q is usually filled with NULL for dimensions > 1.

Objects from the Class

Objects can be created by calls of the form new("MultivariateDistribution", ...).

Slots

img:

Object of class "rSpace". Image space of the distribution. Usually an object of class "EuclideanSpace".

param:

Object of class "OptionalParameter". Optional parameter of the multivariate distribution.

r:

Object of class "function": generates (pseudo-)random numbers

d:

Object of class "OptionalFunction": optional density function

p:

Object of class "OptionalFunction": optional cumulative distribution function

q:

Object of class "OptionalFunction": optional quantile function

Extends

Class "Distribution", directly.

Methods

show

signature(object = "MultivariateDistribution")

plot

signature(object = "MultivariateDistribution"): not yet implemented.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Distribution-class

Examples


# Dirac-measure in (0,0)
new("MultivariateDistribution")

[Package distrEx version 1.9 Index]