MVNormDistribution-class(distrEllipse)	R Documentation

MVNorm distribution class

Description

Class MVNormDistribution implements a general multivariate distribution using code from package mvtnorm. For details to this implementation confer to the references given in this package.

Objects from the Class

Objects could in principle be created by calls to new, but more frequently you would create them via the generating function .

Slots

img

Object of class "Reals".

param

Object of class "MVtParameter".

r

function with argument n; random number generator

d

the density of this distribution, pmvnorm

p

the (vectorized) function pmvnorm.

q

the (vectorized) function qmvnorm.

radDistr

the distribution sqrt(Chisq(df=dim0))

.withArith

FALSE

.withSim

FALSE

.logExact

TRUE

.lowerExact

TRUE

Symmetry

object of class "EllipticalSymmetry" about center loc; used internally to avoid unnecessary calculations.

Extends

Class "EllipticalDistribution", directly.
Class "SphericalDistribution", by class "EllipticalDistribution".
Class "MultivariateDistribution", by class "SphericalDistribution". Class "Distribution", by class "MultivariateDistribution".

Methods

sigma

signature(object = "MVNormDistribution"): wrapped access method for slot sigma of slot param.

mean

signature(object = "MVNormDistribution"): wrapped access method for slot location of slot param.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Package mvtnorm

Examples



[Package distrEllipse version 0.2 Index]