MVNormParameter-class(distrEllipse)	R Documentation

Paramter of a multivariate normal distribution

Description

The class of the parameter of MVNorm distributions.

Objects from the Class

Objects can be created by calls of the form new("MVNormParameter", ...).

Slots

location:

numeric; center / location of the distribution.

scale:

matrix; the scale matrix; the number of rows of this matrix must be the same as the length of location.

name:

default name is “parameter of a Elliptical distribution”.

Extends

Class "EllipticalParameter", directly.
Class "Parameter", by class "EllipticalParameter".
Class "OptionalParameter", by class "Parameter".

Methods

mean

signature(object = "MVNormParameter"): access method for slot location.

sigma

signature(x = "MVNormParameter"): utility function; returns S for S=scale(x).

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

MVNormDistribution-class, Parameter-class

Examples


new("MVNormParameter")

[Package distrEllipse version 0.1 Index]