MVtDistribution(distrEllipse)	R Documentation

Generating function for MvtDistribution-class

Description

Generates an object of class "MvtDistribution".

Usage

MVt(loc = c(0,0), scale = diag(length(loc)), df = 1, ncp = 0)

Arguments

loc 	 real number: location / center of the elliptical distribution.

scale 	 a square matrix (with nrow(scale)==ncol(scale)==length(loc)) of full rank: the / a scale matrix of the elliptical distribution — unique only upto scale%*%t(scale) , i.e. if A1 and A2 are two square matrices of full rank such that A1%*%t(A1)==A2%*%t(A2) , then we obtain the same elliptical distribution for scale = A1 and for scale = A2.

df 	 integer; degrees of freedom

ncp 	 positive real number; non-centrality parameter

Value

Object of class "MvtDistribution"

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

MVtDistribution-class

Examples


E0 <- MVt()
plot(E0)
E1 <- diag(1,2)%*%E0+c(1,2)
plot(E1)
E(E1)
var(E1)

[Package distrEllipse version 0.1 Index]