MVtDistribution-class(distrEllipse)	R Documentation

MVt distribution class

Description

Class MVtDistribution implements multivariate t distributions using code from package mvtnorm. For details to this implementation confer to the references given in this package.

Objects from the Class

Objects could in principle be created by calls to new, but more frequently you would create them via the generating function .

Slots

img

Object of class "Reals".

param

Object of class "MVtParameter".

r

function with argument n; random number generator

d

the density of this distribution, dmvt

p

the (vectorized) function pmvt.

q

the (vectorized) function qmvt.

radDistr

an object of class AbscontDistribution with density

dim * choose((dim+df-1)/2, (df-1)/2) * x ^{dim-1} * df^(-dim/2) * (1+x^2/df)^{-(dim+df)/2}

.withArith

FALSE

.withSim

FALSE

.logExact

TRUE

.lowerExact

TRUE

Symmetry

object of class "EllipticalSymmetry" about center loc; used internally to avoid unnecessary calculations.

Extends

Class "EllipticalDistribution", directly.
Class "SphericalDistribution", by class "EllipticalDistribution".
Class "MultivariateDistribution", by class "SphericalDistribution". Class "Distribution", by class "MultivariateDistribution".

Methods

sigma

signature(object = "MVtDistribution"): wrapped access method for slot sigma of slot param.

ncp

signature(object = "MVtDistribution"): wrapped access method for slot ncp of slot param.

df

signature(x = "MVtDistribution"): wrapped access method for slot scale of slot param.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Package mvtnorm

Examples



[Package distrEllipse version 0.2 Index]