NbinomParameter-class(distr)	R Documentation

Class 'NbinomParameter'

Description

The parameter of a negative binomial distribution, used by Nbinom-class

Objects from the Class

Objects can be created by calls of the form new("NbinomParameter", prob, size). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Nbinom is prepared.

Slots

prob:

Object of class "numeric": the probability of a negative binomial distribution

size:

Object of class "numeric": the size of a negative binomial distribution

name:

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "NbinomParameter"): initialize method

prob

signature(object = "NbinomParameter"): returns the slot prob of the parameter of the distribution

prob<-

signature(object = "NbinomParameter"): modifies the slot prob of the parameter of the distribution

size

signature(object = "NbinomParameter"): returns the slot size of the parameter of the distribution

size<-

signature(object = "NbinomParameter"): modifies the slot size of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@web.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Nbinom-class Parameter-class

Examples


W <- new("NbinomParameter",prob=0.5,size=1)
size(W) # size of this distribution is 1.
size(W) <- 2 # size of this distribution is now 2.

[Package distr version 2.0 Index]