Norm-class(distr)	R Documentation

Class 'Norm'

Description

The normal distribution has density

f(x) = 1/(sqrt(2 pi) sigma) e^-((x - mu)^2/(2 sigma^2))

where mu is the mean of the distribution and sigma the standard deviation. C.f. rnorm

Objects from the Class

Objects can be created by calls of the form Norm(mean, sd). This object is a normal distribution.

Slots

img:

Object of class "Reals": The domain of this distribution has got dimension 1 and the name "Real Space".

param:

Object of class "UniNormParameter": the parameter of this distribution (mean and sd), declared at its instantiation

r:

Object of class "function": generates random numbers (calls function rnorm)

d:

Object of class "function": density function (calls function dnorm)

p:

Object of class "function": cumulative function (calls function pnorm)

q:

Object of class "function": inverse of the cumulative function (calls function qnorm)

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "AbscontDistribution".

Methods

-

signature(e1 = "Norm", e2 = "Norm")

+

signature(e1 = "Norm", e2 = "Norm"): For the normal distribution the exact convolution formulas are implemented thereby improving the general numerical approximation.

*

signature(e1 = "Norm", e2 = "numeric")

+

signature(e1 = "Norm", e2 = "numeric"): For the normal distribution we use its closedness under affine linear transformations.

initialize

signature(.Object = "Norm"): initialize method

mean

signature(object = "Norm"): returns the slot mean of the parameter of the distribution

mean<-

signature(object = "Norm"): modifies the slot mean of the parameter of the distribution

sd

signature(object = "Norm"): returns the slot sd of the parameter of the distribution

sd<-

signature(object = "Norm"): modifies the slot sd of the parameter of the distribution

further arithmetic methods see operators-methods

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

UniNormParameter-class AbscontDistribution-class Reals-class rnorm

Examples


N <- Norm(mean=1,sd=1) # N is a normal distribution with mean=1 and sd=1.
r(N)(1) # one random number generated from this distribution, e.g. 2.257783
d(N)(1) # Density of this distribution is  0.3989423 for x=1.
p(N)(1) # Probability that x<1 is 0.5.
q(N)(.1) # Probability that x<-0.2815516 is 0.1.
mean(N) # mean of this distribution is 1.
sd(N) <- 2 # sd of this distribution is now 2.
M <- Norm() # M is a normal distribution with mean=0 and sd=1.
O <- M+N # O is a normal distribution with mean=1 (=1+0) and sd=sqrt(5) (=sqrt(2^2+1^2)).

[Package distr version 1.9 Index]