ParamFamParameter-class(distrMod)	R Documentation

Parameter of a parametric family of probability measures

Description

Class of the parameter of parametric families of probability measures.

Objects from the Class

Objects can be created by calls of the form new("ParamFamParameter", ...). More frequently they are created via the generating function ParamFamParameter.

Slots

main:

Object of class "numeric": main parameter.

nuisance:

Object of class "OptionalNumeric": optional nuisance parameter.

trafo:

Object of class "MatrixorFunction": transformation of the parameter.

name:

Object of class "character": name of the parameter.

Extends

Class "Parameter", directly.
Class "OptionalParameter", by class "Parameter".

Methods

main

signature(object = "ParamFamParameter"): accessor function for slot main.

main<-

signature(object = "ParamFamParameter"): replacement function for slot main.

nuisance

signature(object = "ParamFamParameter"): accessor function for slot nuisance.

nuisance<-

signature(object = "ParamFamParameter"): replacement function for slot nuisance.

trafo

signature(object = "ParamFamParameter"): accessor function for slot trafo.

trafo<-

signature(object = "ParamFamParameter"): replacement function for slot trafo.

length

signature(x = "ParamFamParameter"): sum of the lengths of main and nuisance.

dimension

signature(x = "ParamFamParameter"): length of main.

show

signature(object = "ParamFamParameter")

Details for methods this-is-escaped-codenormal-bracket84bracket-normal, this-is-escaped-codenormal-bracket85bracket-normal

Detailedness of output by methods show, print is controlled by the global option show.details to be set by distrModoptions.

As method show is used when inspecting an object by typing the object's name into the console, show comes without extra arguments and hence detailedness must be controlled by global options.

Method print may be called with a (partially matched) argument show.details, and then the global option is temporarily set to this value.

More specifically, when show.detail is matched to "minimal" only class and name as well as main and nuisance part of the parameter are shown. When show.detail is matched to "medium", and if you estimate non-trivial (i.e. not the identity) transformation of the parameter of the parametric family, you will in addition be shown the derivative matrix, if the transformation is given in form of this matrix, while, if the transformation is in function form, you will only be told this. Finally, when show.detail is matched to "maximal", and you have a non-trivial transformation in function form, you will also be shown the code to this function.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Parameter-class

Examples


new("ParamFamParameter")

[Package distrMod version 2.0 Index]