Pareto-class(distrEx)	R Documentation

Pareto distribution

Description

[borrowed from actuar]:

The (Single-parameter) Pareto distribution with parameter shape = a has density:

f(x) = a b^a/x^(a + 1)

for x > b, a > 0 and b > 0.

Although there appears to be two parameters, only shape is a true parameter. The value of min = b must be set in advance.

Objects from the Class

Objects can be created by calls of the form new("Pareto", shape, Min). More frequently they are created via the generating function Pareto.

Slots

img:

Object of class "Reals".

param:

Object of class "ParetoParameter".

r:

rpareto1

d:

dpareto1

p:

ppareto1

q:

qpareto1

gaps:

(numeric) matrix or NULL

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "AbscontDistribution".

Methods

initialize

signature(.Object = "Pareto"): initialize method.

shape

signature(object = "Pareto"): wrapped access method for slot shape of slot param.

Min

signature(x = "Pareto"): wrapped access method for slot Min of slot param.

shape<-

signature(object = "Pareto"): wrapped replace method for slot shape of slot param.

Min<-

signature(x = "Pareto"): wrapped replace method for slot Min of slot param.

Note

This class is based on the code provided by the package actuar by Vincent Goulet and Mathieu Pigeon.

Author(s)

Nataliya Horbenko Nataliya.Horbenko@itwm.fraunhofer.de

References

Johnson et al. (1995) Continuous Univariate Distributions. Vol. 2. 2nd ed. New York: Wiley.
Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2004), Loss Models, From Data to Decisions, Second Edition, Wiley.

See Also

dpareto1, AbscontDistribution-class

Examples


(P1 <- new("Pareto", shape = 1, Min = 2))
plot(P1)
shape(P1)
Min(P1)
shape(P1) <- 4
Min(P1) <- 2
plot(P1)

[Package distrEx version 2.1 Index]