PoisParameter-class(distr)	R Documentation

Class 'PoisParameter'

Description

The parameter of a Poisson distribution, used by Pois-class

Objects from the Class

Objects can be created by calls of the form new("PoisParameter", lambda). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Pois is prepared.

Slots

lambda:

Object of class "numeric": the lambda of a Poisson distribution

name:

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "PoisParameter"): initialize method

lambda

signature(object = "PoisParameter"): returns the slot lambda of the parameter of the distribution

lambda<-

signature(object = "PoisParameter"): modifies the slot lambda of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Pois-class Parameter-class

Examples


W <- new("PoisParameter",lambda = 1)
lambda(W) # lambda of this distribution is 1.
lambda(W) <- 2 # lambda of this distribution is now 2.

[Package distr version 2.1 Index]