PrognCondDistribution-class(distrEx)	R Documentation

Posterior distribution in convolution

Description

The posterior distribution of X given (X+E)=y

Usage

PrognCondDistribution(Regr = Norm(), Error = Norm())

Arguments

Regr 	 an object of class "AbscontDistribution"

Error 	 an object of class "AbscontDistribution"

Details

For independent r.v.'s X,E with univariate, absolutely continuous (a.c.) distributions Regr and Error, respectively, PrognCondDistribution() returns the (factorized, conditional) posterior distribution of X given X+E=y. as an object of class PrognCondDistribution.

Value

Object of class "PrognCondDistribution"

Objects from the Class

Objects can be created by calls of the form PrognCondDistribution(Regr, Error) where Regr and error are the respective (a.c.) distributions of X and E.

Slots

cond:

Object of class "PrognCondition": condition

img:

Object of class "rSpace": the image space.

param:

Object of class "OptionalParameter": an optional parameter.

r:

Object of class "function": generates random numbers.

d:

Object of class "OptionalFunction": optional conditional density function.

p:

Object of class "OptionalFunction": optional conditional cumulative distribution function.

q:

Object of class "OptionalFunction": optional conditional quantile function.

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

Extends

Class "AbscontCondDistribution", directly.
Class "Distribution", by classes "UnivariateCondDistribution" and "AbscontCondDistribution".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

PrognCondition-class, UnivariateCondDistribution-class AbscontCondDistribution-class, Distribution-class

Examples

PrognCondDistribution()

[Package distrEx version 2.1 Index]