Reals-class(distr)	R Documentation

Class 'Reals'

Description

Particular case of a one-dimensional Euclidean Space

Usage

Reals()

Objects from the Class

Objects could theoretically be created by calls of the form new("Reals", dimension, name). Usually an object of this class is not needed on its own. It is generated automatically when a univariate absolutly continuous distribution is instantiated.

Slots

dimension:

Object of class "character": the dimension of the space, by default = 1

name:

Object of class "character": the name of the space, by default = "Real Space"

Extends

Class "EuclideanSpace", directly.
Class "rSpace", by class "EuclideanSpace".

Methods

initialize

signature(.Object = "Reals"): initialize method

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

EuclideanSpace-class Naturals-class AbscontDistribution-class

Examples


R <- Reals()
liesIn(R,c(0,0)) # FALSE

[Package distr version 1.9 Index]