RtoDPQ.d(distr)	R Documentation

Default procedure to fill slots dpq given r for discrete distributions

Description

function to do get empirical density, cumulative distribution and quantile function from random numbers

Usage


RtoDPQ.d(r, e = getdistrOption("RtoDPQ.e"))

Arguments

r 	 the random number generator

e 	 10^e numbers are generated, a higher number leads to a better result.

Details

RtoDPQ.d generates 10^e random numbers, by default e = RtoDPQ.e which are used to produce a density, cdf and quantile function. Of course, the results are usually not exact as they rely on random numbers.

Value

RtoDPQ returns a list of functions.

dfun 	 density

pfun 	 cumulative distribution function

qfun 	 quantile function

Note

Use RtoDPQ for absolutely continuous and RtoDPQ.d for discrete distributions.

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@web.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

UnivariateDistribution-class, density, approxfun, ecdf

Examples


rn2 <- function(n){rnorm(n)^2}
x <- RtoDPQ(r = rn2, e = 4, n = 512)
# returns density, cumulative distribution and quantile function of
# squared standard  normal distribution

x$dfun(4)
RtoDPQ(r = rn2, e = 5, n = 1024) # for a better result

rp2 <- function(n){rpois(n, lambda = 1)^2}
x <- RtoDPQ.d(r = rp2, e = 5)
# returns density, cumulative distribution and quantile function of
# squared Poisson distribution with parameter lambda=1

[Package distr version 2.0 Index]