SeqDataFrames-class(distrSim)	R Documentation

Class 'SeqDataFrames'

Description

An object of type "SeqDataFrames" is a list of data frames, all of which with the same numbers and names of columns (ideally with the same data-types for the columns), but with possibly varying number of rows; with correponding index operators it behaves like a three-dimensional array with dimensions sample size x observation dimension x runs.

Details

There is a validity method checking for each member of the list being a data frame and for the accordance of the column structures of the data frames.

Objects from the Class

Objects can be created by calls of the form SeqDataFrames(...), where the ... are a list of dataframes with according column structure.

Slots

data:

a list of data frames

Methods

[

signature(x = "SeqDataFrames"): returns (slices of) the data

[<-

signature(x = "SeqDataFrames"): modifies (slices of) the data

print

signature(x = "SeqDataFrames", obs0 = NULL, dims0 = NULL, runs0 = NULL, short = FALSE, ...): slices can be printed and, if argument short== TRUE only a bounded number of dimensions is shown.

show

signature(object = "SeqDataFrames"): a call to print(x)

names

signature(x = "SeqDataFrames"): returns the names of the runs

runnames

signature(x = "SeqDataFrames"): returns the names of the runs

obsdimnames

signature(x = "SeqDataFrames"): returns the names of the observation dimensions

obsDim

signature(x = "SeqDataFrames"): returns the dimension of the observations

runs

signature(x = "SeqDataFrames"): returns the number of runs

samplesize

signature(x = "SeqDataFrames"): returns the size of the samples for each run

rbind

signature(x = "SeqDataFrames"): concatenates different a list of SeqDataFrames object (with the same column structure) to a new object of class SeqDataFrames to do so we mask the rbind method from package base

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

[-methods print-methods summary-methods

[Package distrSim version 2.2 Index]