SweaveListingOptions(SweaveListingUtils)	R Documentation

Function to change the global options of the package ‘SweaveListingUtils’

Description

With SweaveListingOptions you can inspect and change the global variables of the package SweaveListingUtils.

Usage


SweaveListingOptions(...)
SweaveListingoptions(...)
getSweaveListingOption(x)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Value

SweaveListingOptions() returns a list of the global variables.
SweaveListingOptions(x) returns the global variable x.
getSweaveListingOption(x) returns the global variable x.
SweaveListingOptions(x=y) sets the value of the global variable x to y.

SweaveListingoptions

For compatibility with spelling in package distr, SweaveListingoptions is just a synonym to SweaveListingOptions.

Global Options

Rset:

default setting for \lstset-definition of R-code, default value is

list("language"="R","escapechar"="`", 
        "fancyvrb"="true","basicstyle"="\color{Rcolor}\footnotesize",
        "commentstyle"="\color{Rcomment}\ttfamily\itshape",
        "literate"="{<-}{\{\$\leftarrow\$}\}2",
        "morekeywords"="[2]{Norm,Pois,lambda,p,d,r,distroptions}"
        )

Rdset:

default setting for \lstset-definition of Rd-code, default value is

list("language"="TeX", "basicstyle"="\color{black}\tiny",
               "commentstyle"="\ttfamily\itshape"
               )

Rcolor:

default setting for color of R-code in rgb-coordinates; defaults to c(0,0.5,0.5)

Rout:

default setting for color of R-output in rgb-coordinates; defaults to c(0.461,0.039,0.102)

Rcomment:

default setting for color of R-comments in rgb-coordinates; defaults to c(0.101,0.043,0.432)

pkv:

default setting for package version as character; defaults to "2.0.2"

pkg:

default setting for package name as character; defaults to "distr"

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

options, getOption

Examples


SweaveListingOptions()
SweaveListingOptions("Rout")
SweaveListingOptions("Rout" = c(0,0,0))
# or
SweaveListingOptions(Rout = c(0,1,0))
getSweaveListingOption("Rout")

[Package SweaveListingUtils version 0.1 Index]