SweaveListingPreparations(SweaveListingUtils)	R Documentation

SweaveListingPreparations

Description

helping tool for writing the corresponding TeX preamble commands to integrate Sweave and package listings

Usage


SweaveListingPreparations(
   withOwnFileSection = FALSE,  
   withVerbatim = FALSE, 
   gin = TRUE,
   ae = TRUE,
   LineLength = 80,
   Rset = getSweaveListingOption("Rset"),
   Rdset = getSweaveListingOption("Rdset"),
   Rcolor = getSweaveListingOption("Rcolor"),
   Rbcolor = getSweaveListingOption("Rbcolor"), 
   Rout = getSweaveListingOption("Rout"),
   Rcomment = getSweaveListingOption("Rcomment"),
   pkg = getSweaveListingOption("pkg"),
   pkv = getSweaveListingOption("pkv"),
   lib.loc = NULL)

Arguments

withOwnFileSection 	 logical: Does one want to use an own definition file/ section to define Sinput, Soutput, Scode environments; if TRUE you should write some \include directive to include your own definitions / or write them in your ‘.Rnw’ file.

withVerbatim 	 logical of length 3 (filled by recycling if of shorter length) either named ("Sinput", "Soutput", "Scode") or taken in order (Sinput, Soutput, Scode): Should we use Verbatim from TeX package 'fancyvrb' (i.e. the original "\DefineVerbatimEnvironment{Sinput}{Verbatim}" by Fritz Leisch) or just TeX package 'listings' (i.e. we define our own Sinput environment based on 'listings' command 'lstnewenvironment'; this option is due to a suggestion by Andrew Ellis (thank you!).

gin 	 logical: shall we use gin

ae 	 logical: shall we use ae

LineLength 	 numeric; defaults to 80

Rset 	 list or taglist; defaults to global option "Rset"

Rdset 	 list or taglist; defaults to global option "Rdset"

Rcolor 	 numeric of length 3; defaults to global option "Rcolor"; rgb coordinates of the color in which to print R-code

Rbcolor 	 numeric of length 3; defaults to global option "Rbcolor"; rgb coordinates of the color in which to print R keywords induced by intermediate packages.

Rout 	 numeric of length 3; defaults to global option "Rout"; rgb coordinates of the color in which to print R output

Rcomment 	 numeric of length 3; defaults to global option "Rcomment"; rgb coordinates of the color in which to print comments in R-code

pkg 	 character; name of the packageto be described (e.g. in a vignette); defaults to global option "pkg"

pkv 	 character; package version to be described (e.g. in a vignette); defaults to global option "pkv"

lib.loc 	 location of a local library in which the described package resides

Details

SweaveListingPreparations writes a corresponding preamble to the ‘.Rnw’-file for the simultaneous use of Sweave and package listings; note that the lines

\usepackage{Sweave}
\SweaveOpts{keep.source=TRUE}

still have to appear in the ‘.Rnw’ file — before the corresponding SweaveListingPreparations-chunk.
The TeX code inserted by SweaveListingPreparations sets colors, requires TeX-package ‘listings’, defines listings-settings for ‘.R’- and ‘.Rd’-code. For ‘.Rd’ files it uses a new 'listings'-language definition file, ‘Rdlisting.sty’ to be found in subfolder ‘TeX’ in the SweaveListingUtils package folder. SweaveListingPreparations is to be called in an ‘.Rnw’ file and [re-]defines Sinput, Soutput, Scode environments for use with TeX-package ‘listings’, defines commands \code, \file, \pkg and sets the corresponding package version.
The default values are taken from SweaveListingOptions. The output to stdout can be captured in an ‘.Rnw’ file as

<< lstPreamble, results=tex, echo=FALSE>>=
require(SweaveListingUtils)
SweaveListingPreparations()
@

to insert the corresponding preamble parts to the produced TeX file.
If you one want to use destinct keywordstyles for packages loaded by require or library without adding extra arguments to these commands (and hence displaying the actual R code), you have to set up a global formatting matrix .tobeDefinedPkgs by setToBeDefinedPkgs.

Value

invisible()

Acknowledgement

The author wants to thank Frank E. Harrel and Andrew Ellis for very valuable suggestions to enhance this package.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

Examples


SweaveListingPreparations()
SweaveListingPreparations(pkv="2.1")

[Package SweaveListingUtils version 0.3 Index]