TParameter-class(distr)	R Documentation

Class 'TParameter'

Description

The parameter of a t distribution, used by Td-class

Objects from the Class

Objects can be created by calls of the form new("TParameter", df, ncp). Usually an object of this class is not needed on its own, it is generated automatically when an object of the class Td is instantiated.

Slots

df:

Object of class "numeric": the degrees of freedom of a T distribution

ncp:

Object of class "numeric": the noncentrality parameter of a T distribution

name:

Object of class "character": a name / comment for the parameters

Extends

Class "Parameter", directly.

Methods

initialize

signature(.Object = "TParameter"): initialize method

df

signature(object = "TParameter"): returns the slot df of the parameter of the distribution

df<-

signature(object = "TParameter"): modifies the slot df of the parameter of the distribution

ncp

signature(object = "TParameter"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "TParameter"): modifies the slot ncp of the parameter of the distribution

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Td-class Parameter-class

Examples


W <- new("TParameter",df=1, ncp = 0)
df(W) # df of this distribution is 1.
df(W) <- 2 # df of this distribution is now 2.

[Package distr version 2.0 Index]