Td-class(distr)	R Documentation

Class 'Td'

Description

The t distribution with df = n degrees of freedom has density

f(x) = Gamma((n+1)/2) / (sqrt(n pi) Gamma(n/2)) (1 + x^2/n)^-((n+1)/2)

for all real x. It has mean 0 (for n > 1) and variance n/(n-2) (for n > 2). C.f. rt

Objects from the Class

Objects can be created by calls of the form Td(df). This object is a t distribution.

Slots

img:

Object of class "Reals": The domain of this distribution has got dimension 1 and the name "Real Space".

param:

Object of class "TParameter": the parameter of this distribution (df), declared at its instantiation

r:

Object of class "function": generates random numbers (calls function rt)

d:

Object of class "function": density function (calls function dt)

p:

Object of class "function": cumulative function (calls function pt)

q:

Object of class "function": inverse of the cumulative function (calls function qt)

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "AbscontDistribution".

Methods

initialize

signature(.Object = "Td"): initialize method

df

signature(object = "Td"): returns the slot df of the parameter of the distribution

df<-

signature(object = "Td"): modifies the slot df of the parameter of the distribution

ncp

signature(object = "Td"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "Td"): modifies the slot ncp of the parameter of the distribution

Ad hoc methods

For R Version <2.3.0 ad hoc methods are provided for slots q, r if ncp!=0; for R Version >=2.3.0 the methods from package stats are used.

Note

The general non-central t with parameters (df,Del) = (df, ncp) is defined as a the distribution of T(df,Del) := (U + Del) / (Chi(df) / sqrt(df)) where U and Chi(df) are independent random variables, U ~ N(0,1), and Chi(df)^2 is chi-squared, see rchisq.

The most used applications are power calculations for t-tests:
Let T= (mX - m0) / (S/sqrt(n)) where mX is the mean and S the sample standard deviation (sd) of X_1,X_2,...,X_n which are i.i.d. N(mu,sigma^2). Then T is distributed as non-centrally t with df= n-1 degrees of freedom and non-centrality parameter ncp= (mu - m0) * sqrt(n)/sigma.

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

TParameter-class, AbscontDistribution-class, Reals-class, rt

Examples


T <- Td(df = 1) # T is a t distribution with df = 1.
r(T)(1) # one random number generated from this distribution, e.g. -0.09697573
d(T)(1) # Density of this distribution is 0.1591549 for x = 1.
p(T)(1) # Probability that x < 1 is 0.75.
q(T)(.1) # Probability that x < -3.077684 is 0.1.
df(T) # df of this distribution is 1.
df(T) <- 2 # df of this distribution is now 2.
Tn <- Td(df = 1, ncp = 5) 
  # T is a noncentral t distribution with df = 1 and ncp = 5.
d(Tn)(1) ## from R 2.3.0 on ncp no longer ignored...

[Package distr version 2.0 Index]