UnivarDistrList-class(distr)	R Documentation

List of univariate distributions

Description

Create a list of univariate distributions

Objects from the Class

Objects can be created by calls of the form new("UnivarDistrList", ...). More frequently they are created via the generating function DistrList.

Slots

.Data:

Object of class "list". A list of univariate distributions.

Extends

Class "DistrList", directly.
Class "list", by class "DistrList".
Class "vector", by class "DistrList".

Methods

coerce

signature(from = "UnivariateDistribution", to = "UnivarDistrList"): create a UnivarDistrList object from a univariate distribution

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

UnivarDistrList, DistrList-class, UnivariateDistribution-class

Examples


(DL <- new("UnivarDistrList", list(Norm(), Exp())))
plot(DL)
as(Norm(), "UnivarDistrList")

[Package distr version 2.0 Index]