UnivarLebDecDistribution(distr)	R Documentation

Generating function for Class 'UnivarLebDecDistribution'

Description

Generates an object of class "UnivarLebDecDistribution".

Usage


UnivarLebDecDistribution(acPart, discretePart, acWeight, discreteWeight,
                                     r = NULL, e = NULL, n = NULL)

Arguments

acPart 	 Object of class "AbscontDistribution" (or subclasses); a.c. part of the distribution

discretePart 	 Object of class "AbscontDistribution" (or subclasses); discrete part of the distribution

acWeight 	 Object of class "numeric"; weight of the a.c. part of the distribution

discreteWeight 	 Object of class "numeric"; weight of the discrete part of the distribution

r 	 optional argument; if given, this is a random number generator as function r <- function(n){....} to produce r.v.'s distributed according to the distribution; used in a call to RtoDPQ.LC if acPart and discretePart are missing.

e 	 optional argument; if argument r is given, this is the number of r.v.'s drawn to fill the empty slots of this object; if missing filled with getdistrOption("RtoDPQ.e").

n 	 optional argument; if argument r is given, this is the number gridpoints used in filling the empty p,d,q slots of this object; if missing filled with getdistrOption("DefaultNrGridPoints").

Details

At least one of arguments discretePart, acPart, or r must be given; if the first two are missing, slots are filled by a call to RtoDPQ.LV. For this purpose argument r is used together with arguments e and n. If the latter are missing they are filled with getdistrOption("RtoDPQ.e") and getdistrOption("DefaultNrGridPoints"), respectively. If argument discretePart is missing but acPart is not, discreteWeight is set to 0 and discretePart is set to Dirac(0). If argument acPart is missing but discretePart is not, acWeight is set to 0 and discretePart is set to Norm(). If both arguments acPart and discretePart are given, at least one of arguments discreteWeight and acWeight must be given and lie in [0,1], else an error is thrown. If only one argument acWeight or discreteWeight is given the other one is gotten as 1-[ac/discrete]Weight. Else if both are given, they must sum up to 1. If a weight is smaller than getdistrOption("TruncQuantile"), it is set to 0.

Value

Object of class "UnivarLebDecDistribution".

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

UnivarLebDecDistribution-class, simplifyD

Examples


mylist <- UnivarLebDecDistribution(discretePart=Binom(3,.3), acPart=Norm(2,2),
          acWeight=11/20)
mylist          

[Package distr version 2.2 Index]