UnivarMixingDistribution(distr)	R Documentation

Generating function for Class 'UnivarMixingDistribution'

Description

Generates an object of class "UnivarMixingDistribution".

Usage

UnivarMixingDistribution(..., mixCoeff, 
                                withSimplify = getdistrOption("simplifyD"))

Arguments

... 	 Objects of class "UnivariateDistribution" (or subclasses)

mixCoeff 	 Objects of class "numeric" : a vector of probabilities for the mixing components (must be of same length as arguments in ...).

withSimplify 	 "logical": shall the return value be piped through a call to simplifyD?

Details

If mixCoeff is missing, all elements in ... are equally weighted.

Value

Object of class "UnivarMixingDistribution", or if argument withSimplify is TRUE and the resulting object would have one mixing component with probability (almost) 1, UnivarMixingDistribution will return this component.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

UnivarMixingDistribution-class, simplifyD

Examples


mylist <- UnivarMixingDistribution(Binom(3,.3), Dirac(2), Norm(), 
          mixCoeff=c(1/4,1/5,11/20))

[Package distr version 2.0 Index]