UnivarMixingDistribution-class(distr)	R Documentation

Class 'UnivarMixingDistribution'

Description

UnivarMixingDistribution-class is a class to formalize univariate mixing distributions; it is a subclass to class UnivariateDistribution.

Objects from the Class

Objects can be created by calls of the form new("UnivarMixingDistribution", ...). More frequently they are created via the generating function UnivarMixingDistribution.

Slots

mixCoeff:

Object of class "numeric": a vector of probabilities for the mixing components.

mixDistr:

Object of class "UnivarDistrList": a list of univariate distributions containing the mixing components; must be of same length as mixCoeff.

img:

Object of class "Reals": the space of the image of this distribution which has dimension 1 and the name "Real Space"

param:

Object of class "Parameter": the parameter of this distribution, having only the slot name "Parameter of a discrete distribution"

r:

Object of class "function": generates random numbers

d:

fixed to NULL

p:

Object of class "function": cumulative distribution function

q:

Object of class "function": quantile function

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

Extends

Class "UnivariateDistribution" class "Distribution" by class "UnivariateDistribution".

Methods

show

signature(object = "UnivarMixingDistribution") prints the object

mixCoeff<-

signature(object = "UnivarMixingDistribution") replaces the corresponding slot

mixCoeff

signature(object = "UnivarMixingDistribution") returns the corresponding slot

mixDistr<-

signature(object = "UnivarMixingDistribution") replaces the corresponding slot

mixDistr

signature(object = "UnivarMixingDistribution") returns the corresponding slot

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Parameter-class, UnivariateDistribution-class, LatticeDistribution-class, AbscontDistribution-class, simplifyD, flat.mix

Examples


mylist <- UnivarMixingDistribution(Binom(3,.3), Dirac(2), Norm(), 
          mixCoeff=c(1/4,1/5,11/20))
mylist2 <- UnivarMixingDistribution(Binom(3,.3), mylist, 
          mixCoeff=c(.3,.7))
mylist2
p(mylist)(0.3)          
mixDistr(mylist2)

[Package distr version 2.0 Index]