asSemivar-class(distrMod)	R Documentation

Semivariance Risk Type

Description

Class for semi-variance risk.

Objects from the Class

Objects can be created by calls of the form new("asSemivar", ...). More frequently they are created via the generating function asSemivar.

Slots

sign:

Object of class "numeric" – accounts for whether we only consider positive (sign = 1) or negative bias (sign = -1).

Methods

sign

signature(object = "asSemivar"): accessor function for slot sign.

sign<-

signature(object = "asSemivar", value = "numeric"): replacement function for slot sign.

Extends

Class "asGRisk", directly.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

BiasType-class

Examples


asSemivar()

[Package distrMod version 2.2 Index]