cload(distrSim)	R Documentation

cload

Description

loads the comment file from a saved Dataclass object

Usage


cload(filename)

Arguments

filename 	 the filename which was declared at the instantiaton of the Dataclass

Details

Uses function load to load the comment file from a saved Dataclass object.

Value

no value is returned

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Dataclass-class load savedata-methods

Examples


# see Dataclass and Evaluation for examples
## The function is currently defined as
function(filename){
  eval.parent(parse(text=paste("load(\"",filename,".comment\")", sep = "")))
  }

[Package distrSim version 2.2 Index]