df-methods(distr)	R Documentation

Methods for Function df in Package ‘distr’

Description

df-methods

Methods

df

signature(object = "TParameter"): returns the slot df of the parameter of the distribution

df<-

signature(object = "TParameter"): modifies the slot df of the parameter of the distribution

df

signature(object = "Td"): returns the slot df of the parameter of the distribution

df<-

signature(object = "Td"): modifies the slot df of the parameter of the distribution

df

signature(object = "ChisqParameter"): returns the slot df of the parameter of the distribution

df<-

signature(object = "ChisqParameter"): modifies the slot df of the parameter of the distribution

df

signature(object = "Chisq"): returns the slot df of the parameter of the distribution

df<-

signature(object = "Chisq"): modifies the slot df of the parameter of the distribution

[Package distr version 2.2 Index]