distrExOptions(distrEx)	R Documentation

Function to change the global variables of the package ‘distrEx’

Description

With distrExOptions you can inspect and change the global variables of the package distrEx.

Usage


distrExOptions(...)
distrExoptions(...)
getdistrExOption(x)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Value

distrExOptions() returns a list of the global variables.newline distrExOptions(x) returns the global variable x.newline getdistrExOption(x) returns the global variable x.newline distrExOptions(x=y) sets the value of the global variable x to y.

distrExoptions

For compatibility with spelling in package distr, distrExoptions is just a synonym to distrExOptions.

Global Options

MCIterations:

number of Monte-Carlo iterations used for crude Monte-Carlo integration.

GLIntegrateTruncQuantile:

If integrate fails and there are infinite integration limits, the function GLIntegrate is called inside of distrExIntegrate with the corresponding quantiles GLIntegrateTruncQuantile respectively, 1 - GLIntegrateTruncQuantile as finite integration limits.

GLIntegrateOrder:

The order used for the Gauss-Legendre integration inside of distrExIntegrate.

ElowerTruncQuantile:

The lower limit of integration used inside of E which corresponds to the ElowerTruncQuantile-quantile.

EupperTruncQuantile:

The upper limit of integration used inside of E which corresponds to the (1-ElowerTruncQuantile)-quantile.

ErelativeTolerance:

The relative tolerance used inside of E when calling distrExIntegrate.

m1dfLowerTruncQuantile:

The lower limit of integration used inside of m1df which corresponds to the m1dfLowerTruncQuantile-quantile.

m1dfRelativeTolerance:

The relative tolerance used inside of m1df when calling distrExIntegrate.

m2dfLowerTruncQuantile:

The lower limit of integration used inside of m2df which corresponds to the m2dfLowerTruncQuantile-quantile.

m2dfRelativeTolerance:

The relative tolerance used inside of m2df when calling distrExIntegrate.

nDiscretize:

number of support values used for the discretization of objects of class "AbscontDistribution".

hSmooth:

smoothing parameter to smooth objects of class "DiscreteDistribution". This is done via convolution with the normal distribution Norm(mean = 0, sd = hSmooth).

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

options, getOption

Examples


distrExOptions()
distrExOptions("ElowerTruncQuantile")
distrExOptions("ElowerTruncQuantile" = 1e-6)
# or
distrExOptions(ElowerTruncQuantile = 1e-6)
getdistrExOption("ElowerTruncQuantile")

[Package distrEx version 2.1 Index]