distrModMASK(distrMod)	R Documentation

Masking of/by other functions in package 'distrMod'

Description

Provides information on the (intended) masking of and (non-intended) masking by other other functions in package distrMod

Usage


distrModMASK(library = NULL)

Arguments

library 	 a character vector with path names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. If the default is used, the loaded packages are searched before the libraries

Value

no value is returned

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

Examples


distrModMASK()

[Package distrMod version 2.2 Index]