distrModOptions(distrMod)	R Documentation

Function to change the global variables of the package ‘distrMod’

Description

With distrModOptions you can inspect and change the global variables of the package distrMod.

Usage


distrModOptions(...)
getdistrModOption(x)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Value

distrModOptions() returns a list of the global variables.
distrModOptions(x) returns the global variable x.
getdistrModOption(x) returns the global variable x.
distrModOptions(x=y) sets the value of the global variable x to y.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@web.de

See Also

options, getOption

[Package distrMod version 2.0 Index]