distrSimoptions(distrSim)	R Documentation

functions to change the global variables of the package ‘distrSim’

Description

With distrSimoptions and getdistrSimOption you may inspect and change the global variables used by package distrSim.

Usage


distrSimoptions(...)
getdistrSimOption(x)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Details

Invoking distrSimoptions() with no arguments returns a list with the current values of the options. To access the value of a single option, one should use getdistrSimOption("MaxNumberofSummarizedRuns"), e.g., rather than distroptions("MaxNumberofSummarizedRuns") which is a list of length one.

Value

distrSimoptions() returns a list of the global options of distrSim.
distrSimoptions("MaxNumberofSummarizedRuns") returns the global option MaxNumberofSummarizedRuns as a list of length 1.
distrSimoptions("MaxNumberofSummarizedRuns" = 3) sets the value of the global option MaxNumberofSummarizedRuns to 3. getdistrSimOption("MaxNumberofSummarizedRuns") the current value set for option MaxNumberofSummarizedRuns.

Currently available options

MaxNumberofPlottedObs

maximal number of observation plotted; defaults to 4000

MaxNumberofPlottedObsDims

maximal number of observation dimensions plotted in parallel; defaults to 6

MaxNumberofPlottedRuns

maximal number of runs plotted in parallel; defaults to 6

MaxNumberofSummarizedObsDims

maximal number of observation dimensions summarized in parallel; defaults to 6

MaxNumberofSummarizedRuns

maximal number of runs summarized in parallel; defaults to 6

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@web.de

See Also

options, getOption, distroptions, getdistrOption

Examples


distrSimoptions()
distrSimoptions("MaxNumberofPlottedObsDims")
distrSimoptions("MaxNumberofPlottedObsDims" = 5)
# or
getdistrSimOption("MaxNumberofPlottedObsDims")

[Package distrSim version 2.0 Index]