distrTEstoptions(distrTEst)	R Documentation

functions to change the global variables of the package ‘distrTEst’

Description

With distrTEstoptions and getdistrTEstOption you may inspect and change the global variables used by package distrTEst.

Usage


distrTEstoptions(...)
getdistrTEstOption(x)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Details

Invoking distrTEstoptions() with no arguments returns a list with the current values of the options. To access the value of a single option, one should use getdistrTEstOption("MaxNumberofSummarizedEvaluations"), e.g., rather than distroptions("MaxNumberofSummarizedEvaluations") which is a list of length one.

Value

distrTEstoptions() returns a list of the global options of distrTEst.
distrTEstoptions("MaxNumberofSummarizedEvaluations") returns the global option MaxNumberofSummarizedEvaluations as a list of length 1.
distrTEstoptions("MaxNumberofSummarizedEvaluations" = 3) sets the value of the global option MaxNumberofSummarizedEvaluations to 3. getdistrTEstOption("MaxNumberofSummarizedEvaluations") the current value set for option MaxNumberofSummarizedEvaluations.

Currently available options

MaxNumberofPlottedEvaluations

maximal number of evaluations plotted; defaults to 6

MaxNumberofPlottedEvaluationDims

maximal number of evaluation dimensions plotted in parallel; defaults to 6

MaxNumberofSummarizedEvaluations

maximal number of evaluations summarized in parallel; defaults to 15

MaxNumberofPrintedEvaluations

maximal number of evaluations printed in parallel; defaults to 15

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

See Also

options, getOption, distroptions, getdistrOption

Examples


distrTEstoptions()
distrTEstoptions("MaxNumberofPlottedEvaluationDims")
distrTEstoptions("MaxNumberofPlottedEvaluationDims" = 5)
# or
getdistrTEstOption("MaxNumberofPlottedEvaluationDims")

[Package distrTEst version 2.0 Index]