evaluate-methods(distrTEst)	R Documentation

Methods for Function evaluate in Package 'distrTEst'

Description

evaluate-methods to produce objects of class "Evaluation"

Arguments

object 	 the data set / simulation on which the evaluation takes place

estimator 	 the estimation function used; should be able to deal with data in matrix form samplesize x obsDim, and, should return either a univariate result or a vector (with named coordinates, if possible).

resname

name

filename 	 character; the filename for the Evaluation object (where it is to be saved to); by default the filename of the Data set object which is concatenated with the name of the estimator in savedata.

Details


besides the arguments determining the method dispatch, we have:

evaluate(object, estimator, ..., resname = "res", 
                       name = as.character(substitute(object)),
                       filename = filename(object))


Methods

evaluate

signature(object = "Dataclass", estimator = "function"): creates an object of class "Evaluation", see there for further information

evaluate

signature(object = "Contsimulation", estimator = "function"): creates an object of class "Evaluation", see there for further information

See Also

Evaluation-class

[Package distrTEst version 1.9 Index]