existsPIC-methods(distrMod)	R Documentation

Methods for Function existsPIC in Package ‘distrMod’

Description

existsPIC-methods to check whether in a given L2 differentiable model at parameter value theta there exist (partial) influence curves to Trafo D_theta.

Usage

existsPIC(object, ...)
## S4 method for signature 'L2ParamFamily':
existsPIC(object, warning = TRUE, tol = .Machine$double.eps)

Arguments

object 	 L2ParamFamily

... 	 further arguments used by specific methods.

warning 	 logical: should a warning be issued if there exist no (partial) influence curves?

tol 	 the tolerance the linear algebraic operations. Default is .Machine$double.eps.

Details

To check the existence of (partial) influence curves and, simultaneously, for bounded (partial) influence curves, by Lemma 1.1.3 in Kohl(2005) [resp. the fact that ker I = ker J for J= E (Lambda',1)' (Lambda',1) w and w=min(1, b/|(Lambda',1)|)], it suffices to check that ker I is a subset of ker D_theta. This is done by a call to isKerAinKerB.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@web.de

References

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

isKerAinKerB

[Package distrMod version 2.0 Index]