flat.LCD(distr)	R Documentation

Flattening a list of Lebesgue decomposed distributions

Description

flattens a list of Lebesgue decomposed distributions endowed with weights to give one Lebesgue decomposed distribution

Usage


flat.LCD(..., mixCoeff = NULL, withgaps = getdistrOption("withgaps"))

Arguments

... 	 list of Lebesgue decomposed distributions

mixCoeff 	 Object of class "numeric" of the same length as ...: a vector of probabilities for the mixing components.

withgaps 	 logical; shall gaps be detected empirically?

Details

flat.LCD flattens a list of Lebesgue decomposed distributions given through ..., i.e., it takes all list elements and mixing coefficients and builds up the mixed distribution (forgetting about the components); the result will be one distribution of class UnivarLebDecDistribution. If mixCoeff is missing, all list elements are equally weighted. It is used internally in our methods for "*", "/", "^" (see operators-methods), Minimum, and convpow, as well in method flat.mix.

Value

flat.LCD returns an object of class UnivarLebDecDistribution.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

UnivarLebDecDistribution-class, operators-methods

Examples


D1 <- as(Norm(),"UnivarLebDecDistribution")
D2 <- as(Pois(1),"UnivarLebDecDistribution")
D3 <- as(Binom(1,.4),"UnivarLebDecDistribution")
flat.LCD(D1,D2,D3, mixCoeff = c(0.4,0.5,0.1))

[Package distr version 2.2 Index]