gaps-methods(distr)	R Documentation

Methods for Functions gaps and setgaps in Package ‘distr’

Description

[set]gaps-methods

Usage

gaps(object)
gaps(object)
gaps(object) <- value
setgaps(object, ...)
## S4 method for signature 'AbscontDistribution':
gaps(object)
## S4 method for signature 'AbscontDistribution':
setgaps(object, exactq = 6, 
           ngrid = 50000, ...)

Arguments

object 	 object of class "AbscontDistribution" (or subclasses)

... 	 further arguments to be passed to setgaps; not yet used.

value 	 n x 2 matrix m of numerics where c(t(m)) is an ordered vector; value to be assigned to slot gaps

exactq 	 density values smaller than 10^-exactq are considered as 0.

ngrid 	 number of gridpoints at which the density is evaluated.

Methods

gaps

signature(object = "AbscontDistribution"): returns slot gaps of an absolutely continuous distribution

setgaps

signature(object = "AbscontDistribution"): tries to find out the gaps (where d(object) is approximately 0) and fills slot gaps of object correspondingly

gaps<-

signature(object = "AbscontDistribution"): modifies slot gaps of an absolutely continuous distribution

[Package distr version 2.1 Index]