getLabel(distr)	R Documentation

Labels for distribution objects

Description

a help function to get reasonable labels for distribution objects

Usage


getLabel(x, withnames = TRUE)

Arguments

x 	 a distribution object

withnames 	 logical: are the parameters (if any) of x to be displayed with names?

Remark

The need for this helper function (external to our plot methods) was brought to our attention in a mail by Kouros Owzar owzar001@mc.duke.edu.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@web.de

See Also

plot-methods

Examples


## example due to Kouros Owzar:
foo<- function(law,n, withnames = TRUE)
  {
    data.frame(muhat=mean(r(law)(n)),n=n,law= getLabel(law,withnames))
  } 
### a function that groups certain informations on 
##  created with distribution objects
do.call("rbind",lapply(list(Exp(1),Norm(0,1),Weibull(1,1)),foo,n=100))
do.call("rbind",lapply(list(Exp(1),Norm(0,1),Weibull(1,1)),foo,n=100,FALSE))

[Package distr version 2.0 Index]