getLow,getUp(distr)	R Documentation

getLow getUp functions of package distr

Description

getLow, getUp return lower and upper endpoint of a distribution — truncated to lower/upper TruncQuantile if infinite; in case of an object of class "LatticeDistribution" with infinite lattice length, we search for the smallest/largest point in the lattice which is returned by succesive halving of x=0.5 in q(object)(x, lower.tail) for lower.tail TRUE resp. false.

Usage


## S4 method for signature 'AbscontDistribution':
getUp(object,
                                      eps = getdistrOption("TruncQuantile"))
## S4 method for signature 'DiscreteDistribution':
getUp(object, ...)
## S4 method for signature 'LatticeDistribution':
getUp(object, ...)
## S4 method for signature 'UnivarLebDecDistribution':
getUp(object,
                                      eps = getdistrOption("TruncQuantile"))
## S4 method for signature 'UnivarMixingDistribution':
getUp(object,
                                      eps = getdistrOption("TruncQuantile"))
## S4 method for signature 'AbscontDistribution':
getLow(object,
                                       eps = getdistrOption("TruncQuantile"))
## S4 method for signature 'DiscreteDistribution':
getLow(object, ...)
## S4 method for signature 'LatticeDistribution':
getLow(object, ...)
## S4 method for signature 'UnivarLebDecDistribution':
getLow(object,
                                      eps = getdistrOption("TruncQuantile"))
## S4 method for signature 'UnivarMixingDistribution':
getLow(object,
                                      eps = getdistrOption("TruncQuantile"))

Arguments

object 	 a distribution object

eps 	 truncation point (numeric)

... 	 for convenience only; makes it possible to call getLow, getUp with argument eps no matter of the class of object; is ignored in these functions.

Value

getLow,getUp 	 a numeric of length 1

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

[Package distr version 2.1 Index]