InternalClassUnions(distr)	R Documentation

Class unions in 'distr'

Description

Class unions in package distr defined for internal purposes; these are OptionalNumeric,

Details

These classes are used internally to make available methods or to allow slots of classes to be filled with varying types. In particular

"OptionalNumeric"

may contain objects of class "numeric" or "NULL"; it is used e.g. for slot nuisance of class "ParamFamParameter", as it may or may not be present but if so it has to be numeric.

Objects from the Class

All of these classes are virtual: No objects may be created from them.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

numeric-class,

[Package distr version 2.2 Index]