internal_mlehelpers_for_distrMod(distrMod)	R Documentation

Internal helper functions for treating MCEstimators in package distrMod

Description

These functions are used internally by functions MCEstimator and MLEstimator in package ``distrMod''.

Usage


.negLoglikelihood(x, Distribution, ...)
.process.meCalcRes(res, PFam, trafo, res.name, call, asvar.fct, ...)

Arguments

x 	 numeric; the data at which to evaluate the negative log likelihood

Distribution 	 an object of class Distribution; the distribution at which to evaluate the negative log likelihood

... 	 in case of .negLoglikelihood simply to be able to use the general calling for the criterion function; in case of .process.meCalcRes further arguments to be passed to function asvar.fct

res 	 a list; the return value of mleCalc or mceCalc

PFam 	 an object of class ParamFamily; the parametric family at which to evaluate the MCE

trafo 	 the trafo matrix or a function returning it or NULL

res.name 	 the name of the corresponding estimator

call 	 the original call with which the estimate was produced

asvar.fct 	 a function producing an asymptotic covariance matrix or NULL.

Details

.negLoglikelihood uses the log -argument of the corresponding d-slot of the distribution if available; else produces log(d(Distribution)(x)).

.get.criterion.fct produces a function criterion.fct to fill slot minuslogl when an object of class MCEstimate is coerced to class mle (from package stats4); this way we may use profiling methods introduced there also for objects of our classes. More specifically, we produce a function where all coordinates/components of theta appear as separate named arguments, which then calls fun with these separate arguments again stacked to one (named) vector argument;

.process.meCalcRes processes the resulting return value list of methods mceCalc and mleCalc to give a corresponding object of class MCEstimate.

Value

negLoglikelihood 	 the (numeric) value of the negative log likelihood evaluated at the data and the given distribution

.get.criterion.fct 	 a function as described in section ``details''

.process.meCalcRes 	 an object of class MCEstimate

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

MCEstimate-class, mle-class,

[Package distrMod version 2.2 Index]