lambda-methods(distr)	R Documentation

Methods for Function lambda in Package ‘distr’

Description

lambda-methods

Methods

lambda

signature(object = "PoisParameter"): returns the slot lambda of the parameter of the distribution

lambda<-

signature(object = "PoisParameter"): modifies the slot lambda of the parameter of the distribution

lambda

signature(object = "Pois"): returns the slot lambda of the parameter of the distribution

lambda<-

signature(object = "Pois"): modifies the slot lambda of the parameter of the distribution

[Package distr version 2.0 Index]