location-methods(distr)	R Documentation

Methods for Function location in Package ‘distr’

Description

location-methods

Methods

location

signature(object = "LogisParameter"): returns the slot location of the parameter of the distribution

location<-

signature(object = "LogisParameter"): modifies the slot location of the parameter of the distribution

location

signature(object = "Logis"): returns the slot location of the parameter of the distribution

location<-

signature(object = "Logis"): modifies the slot location of the parameter of the distribution

location

signature(object = "CauchyParameter"): returns the slot location of the parameter of the distribution

location<-

signature(object = "CauchyParameter"): modifies the slot location of the parameter of the distribution

location

signature(object = "Cauchy"): returns the slot location of the parameter of the distribution

location<-

signature(object = "Cauchy"): modifies the slot location of the parameter of the distribution

location

signature(object = "DiracParameter"): returns the slot location of the parameter of the distribution

location<-

signature(object = "DiracParameter"): modifies the slot location of the parameter of the distribution

location

signature(object = "Dirac"): returns the slot location of the parameter of the distribution

location<-

signature(object = "Dirac"): modifies the slot location of the parameter of the distribution

[Package distr version 2.2 Index]