m1df(distrEx)	R Documentation

Generic Function for the Computation of Clipped First Moments

Description

Generic function for the computation of clipped first moments. The moments are clipped at upper.

Usage


m1df(object, upper)

Arguments

object 	 object of class "Distribution"

upper 	 clipping bound

Details

The precision of the computations can be controlled via certain global options; cf. distrExOptions.

Value

The first moment of object clipped at upper is computed.

Methods

object = "UnivariateDistribution":

clipped first moment for univariate distributions which is computed using crude Monte-Carlo integration.

object = "AbscontDistribution":

clipped first moment for absolutely continuous univariate distributions which is computed using distrExIntegrate.

object = "DiscreteDistribution":

clipped first moment for discrete univariate distributions which is computed using support and sum.

object = "Binom":

clipped first moment for Binomial distributions which is computed using pbinom.

object = "Pois":

clipped first moment for Poisson distributions which is computed using ppois.

object = "Norm":

clipped first moment for normal distributions which is computed using dnorm and pnorm.

object = "Exp":

clipped first moment for exponential distributions which is computed using pexp.

object = "Chisq":

clipped first moment for Chi^2 distributions which is computed using pchisq.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

distrExIntegrate, m2df, E

Examples


# standard normal distribution
N1 <- Norm()
m1df(N1, 0)

# Poisson distribution
P1 <- Pois(lambda=2)
m1df(P1, 3)

# absolutely continuous distribution
D1 <- Norm() + Exp() # convolution
m1df(D1, 2)
m1df(D1, Inf)
E(D1)

[Package distrEx version 2.0.2 Index]