mean-methods(distr)	R Documentation

Methods for Function mean in Package ‘distr’

Description

mean-methods

Methods

mean

signature(object = "NormParameter"): returns the slot mean of the parameter of the distribution

mean<-

signature(object = "NormParameter"): modifies the slot mean of the parameter of the distribution

mean

signature(object = "Norm"): returns the slot mean of the parameter of the distribution

mean<-

signature(object = "Norm"): modifies the slot mean of the parameter of the distribution

[Package distr version 2.0 Index]