n-methods(distr)	R Documentation

Methods for Function n in Package ‘distr’

Description

n-methods

Methods

n

signature(object = "HyperParameter"): returns the slot n of the parameter of the distribution

n<-

signature(object = "HyperParameter"): modifies the slot n of the parameter of the distribution

n

signature(object = "Hyper"): returns the slot n of the parameter of the distribution

n<-

signature(object = "Hyper"): modifies the slot n of the parameter of the distribution

[Package distr version 2.1 Index]