ncp-methods(distr)	R Documentation

Methods for Function ncp in Package 'distr'

Description

ncp-methods

Methods

ncp

signature(object = "BetaParameter"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "BetaParameter"): modifies the slot ncp of the parameter of the distribution

ncp

signature(object = "Beta"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "Beta"): modifies the slot ncp of the parameter of the distribution

ncp

signature(object = "ChisqParameter"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "ChisqParameter"): modifies the slot ncp of the parameter of the distribution

ncp

signature(object = "Chisq"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "Chisq"): modifies the slot ncp of the parameter of the distribution

ncp

signature(object = "FParameter"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "FParameter"): modifies the slot ncp of the parameter of the distribution

ncp

signature(object = "Fd"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "Fd"): modifies the slot ncp of the parameter of the distribution

ncp

signature(object = "TParameter"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "TParameter"): modifies the slot ncp of the parameter of the distribution

ncp

signature(object = "Td"): returns the slot ncp of the parameter of the distribution

ncp<-

signature(object = "Td"): modifies the slot ncp of the parameter of the distribution

[Package distr version 1.9 Index]