options(distr)	R Documentation

additional options in package ‘distr’

Description

In package distr, we add an extra option "newDevice"; it is inspected and manipulated as usual.

Details

We do not change the behaviour of options or getOption; for the general documentation to these two functions, confer options, getOption. Here we only document added options.

Additionally available options in package 'distr'

"newDevice"

logical; controls behaviour when generating several plots within one function; if TRUE a call to devNew is inserted to hook function plot.new; if FALSE, we reproduce the usual behaviour in graphics, i.e.; we do not call devNew. Defaults to FALSE.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

options, getOption

Examples


getOption("newDevice") 
options("newDevice"=TRUE) 

[Package distr version 2.0.3 Index]