pivot-methods(distr)	R Documentation

Methods for Function pivot in Package ‘distr’

Description

pivot-methods

Methods

pivot

signature(object = "Lattice"): returns the slot pivot of the lattice

pivot<-

signature(object = "Lattice"): modifies the slot pivot of the lattice

pivot

signature(object = "LatticeDistribution"): returns the slot pivot of the lattice slot of the distribution

pivot<-

signature(object = "LatticeDistribution"): modifies the slot pivot of the lattice slot of the distribution

[Package distr version 2.2 Index]