plot-methods(distrEllipse)	R Documentation

Methods for Function plot in Package ‘distrEllipse’

Description

plot-methods

Usage

plot(x, y, ...)
## S4 method for signature 'SphericalDistribution,
##   missing':
plot(x, Nsim = 2000, ...,
                withED = TRUE, lwd.Ed = 2, col.Ed = c(3,4),
                withMean = TRUE, cex.mean = 2, pch.mean = 20, col.mean = 2)

Arguments

x 	 object of class "SphericalDistribution" distribution to be plotted

y 	 missing

Nsim 	 number of (simulated) points to be plotted.

withED 	 logical; shall principal axes of the contour ellipsoid be plot in (for each panel)?

lwd.Ed 	 line width of principal axes (for each panel).

col.Ed 	 color of principal axes (for each panel).

withMean 	 logical; shall mean be plot in (for each panel)?

cex.mean 	 size of the mean symbol (for each panel).

pch.mean 	 mean symbol (for each panel).

col.mean 	 color of the mean symbol (for each panel).

... 	 addtional arguments for plot — see plot, plot.default, plot.stepfun

Details

Using pairs, plots all pairs of coordinates of the object, using simulated values. Any parameters of pairs may be passed on to this particular plot method.

See Also

pairs, plot plot.default, plot.stepfun, par

[Package distrEllipse version 0.1 Index]