print-methods(distr)	R Documentation

Methods for Functions print/show in Package ‘distr’

Description

print/show-methods

Methods

print

signature(x = "UnivariateDistribution"): returns the class of the object and its parameters

show

signature(x = "UnivariateDistribution"): returns the class of the object and its parameters

[Package distr version 2.0 Index]