prob-methods(distr)	R Documentation

Methods for Function prob in Package ‘distr’

Description

prob-methods

Methods

prob

signature(object = "BinomParameter"): returns the slot prop of the parameter of the distribution

prob<-

signature(object = "BinomParameter"): modifies the slot prob of the parameter of the distribution

prob

signature(object = "Binom"): returns the slot prop of the parameter of the distribution

prob<-

signature(object = "Binom"): modifies the slot prob of the parameter of the distribution

prob

signature(object = "NbinomParameter"): returns the slot prop of the parameter of the distribution

prob<-

signature(object = "NbinomParameter"): modifies the slot prob of the parameter of the distribution

prob

signature(object = "Nbinom"): returns the slot prop of the parameter of the distribution

prob<-

signature(object = "Nbinom"): modifies the slot prob of the parameter of the distribution

prob

signature(object = "GeomParameter"): returns the slot prop of the parameter of the distribution (deprecated from 1.9 on)

prob<-

signature(object = "GeomParameter"): modifies the slot prob of the parameter of the distribution (deprecated from 1.9 on)

prob

signature(object = "Geom"): returns the slot prop of the parameter of the distribution

prob<-

signature(object = "Geom"): modifies the slot prob of the parameter of the distribution

prob

signature(object = "DiscreteDistribution"): returns the (named) vector of probabilities for the support points of the distribution.

prob<-

signature(object = "DiscreteDistribution"): generates a new object of class "DiscreteDistribution" with the same support as object as well as the same .withSim, .withArith, .lowerExact, .logExact slots.

prob

signature(object = "UnivarLebDecDistribution"): returns a 2 x n matrix where n is the length of the support of the discrete part of the distribution; the first row named "cond" gives the vector of probabilities for the support points of the discrete part of the distribution (i.e.; conditional on being in the discrete part), the second row named "abs" is like the first one but multiplied with discreteWeight of the distribution, hence gives the absolute probabilities of the support points; the columns are named by the support values.

[Package distr version 2.1 Index]