rate-methods(distr)	R Documentation

Methods for Function rate in Package ‘distr’

Description

rate-methods

Methods

rate

signature(object = "ExpParameter"): returns the slot rate of the parameter of the distribution

rate<-

signature(object = "ExpParameter"): modifies the slot rate of the parameter of the distribution

rate

signature(object = "Exp"): returns the slot rate of the parameter of the distribution

rate<-

signature(object = "Exp"): modifies the slot rate of the parameter of the distribution

rate

signature(object = "DExp"): returns the slot rate of the parameter of the distribution

rate<-

signature(object = "DExp"): modifies the slot rate of the parameter of the distribution

[Package distr version 2.2 Index]