savedata-methods(distrTEst)	R Documentation

Methods for Function savedata in Package 'distrTEst'

Description

savedata-methods

Methods

savedata

signature(object = "Evaluation"): saves the object in two files in the directory of R - one with data —filename = <filename>—, one without as comment file —filename = <filename>.comment—(see example); <filename>the filename can be specified in the optional argument fileN; by default it is concatenated from the filename of the Dataclass object and the estimatorname, which you may either pass as argument estimatorName or it is taken as the R-name of the corresponding R-function.

Note

For an example, see Simulation-class and Contsimulation-class

See Also

Dataclass-class Simulation-class Contsimulation-class Evaluation-class

[Package distrTEst version 1.9 Index]