scale-methods(distr)	R Documentation

Methods for Function scale in Package ‘distr’

Description

scale-methods

Methods

scale

signature(object = "GammaParameter"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "GammaParameter"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "Gammad"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "Gammad"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "LogisParameter"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "LogisParameter"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "Logis"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "Logis"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "WeibullParameter"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "WeibullParameter"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "Weibull"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "Weibull"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "CauchyParameter"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "CauchyParameter"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "Cauchy"): returns the slot scale of the parameter of the distribution

scale<-

signature(object = "Cauchy"): modifies the slot scale of the parameter of the distribution

scale

signature(object = "Chisq"): if ncp(object) is 0, a Chi-squared distribution is also a Gamma distribution; in this case, scale returns 2 else an error;

[Package distr version 2.1 Index]