sd-methods(distr)	R Documentation

Methods for Function sd in Package ‘distr’

Description

sd-methods

Methods

sd

signature(x = "Any"): see sd

sd

signature(x = "NormParameter"): returns the slot sd of the parameter of the distribution

sd<-

signature(object = "NormParameter"): modifies the slot sd of the parameter of the distribution

sd

signature(x = "Norm"): returns the slot sd of the parameter of the distribution

sd<-

signature(object = "Norm"): modifies the slot sd of the parameter of the distribution

See Also

sd

[Package distr version 2.1 Index]