sdlog-methods(distr)	R Documentation

Methods for Function sdlog in Package ‘distr’

Description

sdlog-methods

Methods

sdlog

signature(object = "LnormParameter"): returns the slot sdlog of the parameter of the distribution

sdlog<-

signature(object = "LnormParameter"): modifies the slot sdlog of the parameter of the distribution

sdlog

signature(object = "Lnorm"): returns the slot sdlog of the parameter of the distribution

sdlog<-

signature(object = "Lnorm"): modifies the slot sdlog of the parameter of the distribution

[Package distr version 2.2 Index]