seed-methods(distrSim)	R Documentation

Methods for Function seed in Package ‘distrSim’

Description

seed-methods

Methods

seed

signature(object = "Simulation"): returns the slot seed of an object of class "Simulation"

seed<-

signature(object = "Simulation"): changes the slot seed of an object of class "Simulation"

seed

signature(object = "Contsimulation"): returns the slot seed of an object of class "Contsimulation"

seed<-

signature(object = "Contsimulation"): changes the slot seed of an object of class "Contsimulation"

Note

The value to which the seed is set has to be consistent with the setRNG-package; to draw a ``new'' simulation, use something like seed(X)<-setRNG();simulate(X); cf. manual to this package, p.~9

[Package distrSim version 2.1 Index]