shape-methods(distr)	R Documentation

Methods for Function shape in Package ‘distr’

Description

shape-methods

Methods

shape

signature(object = "GammaParameter"): returns the slot shape of a parameter of a Gamma distribution

shape<-

signature(object = "GammaParameter"): modifies the slot shape of a parameter of a Gamma distribution

shape

signature(object = "Gammad"): returns the slot shape of the parameter slot of a Gamma distribution

shape<-

signature(object = "Gammad"): modifies the slot shape of the parameter slot of a Gamma distribution

shape

signature(object = "WeibullParameter"): returns the slot shape of a parameter of a Weibull distribution

shape<-

signature(object = "WeibullParameter"): modifies the slot shape of a parameter of a Weibull distribution

shape

signature(object = "Weibull"): returns the slot shape of the parameter slot of the distribution

shape<-

signature(object = "Weibull"): modifies the slot shape of the parameter slot of the distribution

shape

signature(object = "Chisq"): if ncp(object) is 0, a Chi-squared distribution is also a Gamma distribution; in this case, shape returns df(object)/2 else an error;

shape

signature(object = "Exp"): returns the slot shape of the parameter slot of the Exp distribution (=1)

[Package distr version 2.2 Index]