showobj-methods(distr)	R Documentation

Helping Methods for Functions print/show in Package ‘distr’

Description

showobj-methods

Usage


showobj(object, ...)
## S4 method for signature 'UnivariateDistribution':
showobj(object, className = class(object)[1])
## S4 method for signature 'DistrList':
showobj(object, class = class(object))
## S4 method for signature 'UnivarMixingDistribution':
showobj(object, class = class(object))
## S4 method for signature 'UnivarLebDecDistribution':
showobj(object, class = class(object),
objs)

Arguments

... 	 further parameters for showobj; not yet used.

object 	 distribution [list] object

className 	 name of the class of the object

objs 	 a character capturing the input from show, print

Value

a character vector, which is then shown by method show, but can also be indented if necessary — like in the possibly nested list classes

See Also

print-methods

[Package distr version 2.0 Index]