size-methods(distr)	R Documentation

Methods for Function size in Package ‘distr’

Description

size-methods

Methods

size

signature(object = "BinomParameter"): returns the slot size of the parameter of the distribution

size<-

signature(object = "BinomParameter"): modifies the slot size of the parameter of the distribution

size

signature(object = "Binom"): returns the slot size of the parameter of the distribution

size<-

signature(object = "Binom"): modifies the slot size of the parameter of the distribution

size

signature(object = "NbinomParameter"): returns the slot size of the parameter of the distribution

size<-

signature(object = "NbinomParameter"): modifies the slot size of the parameter of the distribution

size

signature(object = "Nbinom"): returns the slot size of the parameter of the distribution

size<-

signature(object = "Nbinom"): modifies the slot size of the parameter of the distribution

size

signature(object = "Geom"): returns the slot size of the parameter of the distribution

[Package distr version 2.2 Index]