standardMethods(distr)	R Documentation

Utility to automatically generate accessor and replacement functions

Description

Creates definitions for accessor and replacement functions of an given class.

Usage


standardMethods(class, writetofile = FALSE, directory)

Arguments

class 	 the class for which accessor and replacement functions are to be produced, given as a string

writetofile 	 logical value, indicating wheter output is to be written to a file

directory 	 if writetofile = TRUE, the output is written to a file in the given directory, the name of the file starting with "classname" and ending with "StandardMethods.txt"

Value

no value is returned

Author(s)

Thomas Stabla statho@web.de

Examples


setClass("testclass", representation(a = "numeric", b = "character"))
standardMethods("testclass")

[Package distr version 2.2 Index]