Subsetting-methods(distrSim)	R Documentation

Subsetting/Indexing methods for SeqDataFrames objects in Package ‘distrSim’

Description

Operators acting on SeqDataFrames objects to extract or replace parts.

Usage

 "["(x,i,j,k, drop = FALSE) # is: #  x[i,j, k, drop = FALSE]
        "[<-"(x,i,j,k, value)      # is: #  x[i,j, k] <- value

Arguments

x 	 object of class SeqDataFrame from which to extract element(s) or in which to replace element(s).

i 	 observation index; may be missing

j 	 observation dimension index; may be missing

k 	 run index; may be missing

drop

value

{modication to be assigned to}

Value

again an object of class SeqDataFrame with the prescribed indices / values

Methods

"["

signature(x = "SeqDataFrames"): access method for indices for objects of class SeqDataFrames

"[<-"

signature(x = "SeqDataFrames"): replacement method for indices for objects of class SeqDataFrames

See Also

help("[",package="base")

Examples


s0 <- matrix(1:6,3,2)
d0 <- data.frame(s0)
d1 <- data.frame(s0 + 3)
SF <- seqDataFrames(d0, d1)
SF[1,2,1]


[Package distrSim version 2.0 Index]