summary-methods(distrTEst)	R Documentation

Methods for Function summary in Package ‘distrTEst’

Description

summary-methods

Methods

summary

signature(object = "Evaluation"): returns the name of the data object, its filename, the estimator used and a statistical summary of the result;optional arguments:

runs0

the indices of runs to be summarized;

dims0

the indices of result dimensions to be summarized;

internal argument:

inList

decides if name of Dataobject and Datafile are printed out (which is done if inLIST==FALSE); defaults to FALSE but is TRUE when summary is called from summary-method for signature(object = "EvaluationList")

summary

signature(object = "EvaluationList"): returns, for each member of the list a summary of the corresponding Evaluation object; optional arguments:

eval0

the indices of evaluations to be summarized;— of this vector eval0 maximally MaxNumberofSummarizedEvaluations evaluations are summarized where MaxNumberofSummarizedEvaluations is a global option, see distrTEstoptions

runs0

the indices of runs to be summarized;

dims0

the indices of observation dimensions to be summarized;

[Package distrTEst version 2.1 Index]