validParameter-methods(distrMod)	R Documentation

Methods for function validParameter in Package ‘distrMod’

Description

Methods for function validParameter in package distrMod to check whether a new parameter (e.g. "proposed" by an optimization) is valid.

Usage


validParameter(object, ...)
## S4 method for signature 'ParamFamily':
validParameter(object, param)
## S4 method for signature 'L2ScaleFamily':
validParameter(object, param, tol=.Machine$double.eps)
## S4 method for signature 'L2LocationFamily':
validParameter(object, param)
## S4 method for signature 'L2LocationScaleFamily':
validParameter(object, param, tol=.Machine$double.eps)
## S4 method for signature 'BinomFamily':
validParameter(object, param, tol=.Machine$double.eps)
## S4 method for signature 'PoisFamily':
validParameter(object, param, tol=.Machine$double.eps)
## S4 method for signature 'GammaFamily':
validParameter(object, param, tol=.Machine$double.eps)

Arguments

object 	 an object of class ParamFamily

param 	 either a numeric vector or an object of class ParamFamParameter

tol 	 accuracy upto which the conditions have to be fulfilled

... 	 additional argument(s) for methods.

Details

method for signature

ParamFamily

checks if all parameters are finite by is.finite if their length is between 1 and the joint length of main and nuisance parameter of object, and finally, if a call to modifyParam(object) with argument param would throw an error.

L2ScaleFamily

checks if the parameter is finite by is.finite, if its length is 1, and if it is strictly larger than 0 (upto argument tol)

L2LocationFamily

checks if the parameter is finite by is.finite, if its length is 1

L2LocationScaleFamily

checks if the scale parameter is finite by is.finite, if its length is 1 or 2 (e.g. if one features as nuisance parameter), and if it is strictly larger than 0 (upto argument tol)

BinomFamily

checks if the parameter is finite by is.finite, if its length is 1, and if it is strictly larger than 0 and strictly smaller than 1 (upto argument tol)

PoisFamily

checks if the parameter is finite by is.finite, if its length is 1, and if it is strictly larger than 0 (upto argument tol)

GammaFamily

checks if both parameters are finite by is.finite, if their length is 1 or 2 (e.g. if one features as nuisance parameter), and if both are strictly larger than 0 (upto argument tol)

Value

logical of length 1 — valid or not

Examples


 NS <- NormLocationScaleFamily()
 validParameter(NS, c(scale=0.1, loc=2))
 validParameter(NS, c(scale=-0.1, loc=2))
 validParameter(NS, c(scale=0, loc=2))
 validParameter(NS, c(mean=2, sd=2))

[Package distrMod version 2.1.1 Index]