Version Management(distrSim)	R Documentation

Methods for Version Management in Package 'distrSim'

Description

Version-Management-methods

Usage


## S4 method for signature 'Dataclass':
getVersion(object)
## S4 method for signature 'Dataclass':
conv2NewVersion(object)

Arguments

object 	 object of class "Dataclass" (or subclasses)

Far-reaching Change in Design

From version 1.8 of this package on, we have changed the format how data / simulations are stored: In order to be able to cope with multivariate distributions, regression distributions and (later) time series distributions, we have switched to the common array format: samplesize x obsDim x runs; for saved objects from earlier versions, we provide the functions isOldVersion (from package distr) and conv2NewVersion to check whether the object was generated by an older version of this package and to convert such an object to the new format, respectively.

Methods

getVersion

signature(object = "Dataclass"): returns slot version of an object of class "Dataclass".

conv2NewVersion

signature(object = "Dataclass"): changes an object with format runs x samplesize (old format) to samplesize x obsDim x runs (new format)

See Also

isOldVersion, conv2NewVersion

[Package distrSim version 1.9 Index]