width-methods(distr)	R Documentation

Methods for Function width in Package 'distr'

Description

width-methods

Methods

width

signature(object = "Lattice"): returns the slot width of the lattice

width<-

signature(object = "Lattice"): modifies the slot width of the lattice

width

signature(object = "LatticeDistribution"): returns the slot width of the lattice slot of the distribution

width<-

signature(object = "LatticeDistribution"): modifies the slot width of the lattice slot of the distribution

[Package distr version 1.9 Index]