distrMod-package(distrMod)	R Documentation

distrMod – probability models

Description

Based on the packages distr and distrEx package distrMod provides a flexible framework which allows computation of estimators like maximum likelihood or minimum distance estimators for probability models.

Details

Package: 	 distrMod

Version: 	 2.0

Date: 	 2008-05-08

Depends: 	 R(>= 2.6.0), methods, startupmsg, distr(>= 2.0), distrEx(>= 2.0), RandVar(>= 0.6.2)

LazyLoad: 	 yes

License: 	 GPL (version 2 or later)

URL: 	 http://distr.r-forge.r-project.org/

Classes



ProbFamily classes
slots: [<name>(<class>)] 
name(character), distribution(Distribution),
distrSym(DistributionSymmetry), props(character)

"ProbFamily"
|>"ParamFamily"
additional slots:
param(ParamFamParameter), modifyParam(function)

"ParamFamParameter" is subclass of class "Parameter" of package "distr".
Additional slots:
main(numeric), nuisance(OptionalNumeric), trafo(matrix)

"OptionalNumeric" = union("numeric", "NULL")

Symmetry classes
slots:
type(character), SymmCenter(ANY)

"Symmetry"
|>"NoSymmetry"
|>"EllipticalSymmetry"
|>|>"SphericalSymmetry"

Methods

not specified yet...

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

As for general packageStartupMessage's, you may also suppress all the start-up banner by wrapping the library or require call into suppressPackageStartupMessages() from startupmsg-version 0.5 on.

Demos

none so far ...

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

Maintainer: Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

P. Ruckdeschel, M. Kohl, T. Stabla, F. Camphausen (2006): S4 Classes for Distributions, {em R News}, {em 6}(2), 2-6. http://CRAN.R-project.org/doc/Rnews/Rnews_2006-2.pdf

a vignette for packages distr, distrSim, distrTEst, and distrEx is included into the mere documentation package distrDoc and may be called by require("distrDoc");vignette("distr")

[Package distrMod version 2.0 Index]