Confint-class(distrMod)	R Documentation

Confint-class

Description

Return value S4 classes for method ``confint''.

Objects from the Class

Objects could in principle be created by calls of the form new("Confint", ...). The preferred form is to have them created via a call to confint.

Slots

type:

Object of class "character": type of the confidence interval (asymptotic, bootstrap,...).

confint:

Object of class "array": the confidence interval(s).

estimate.call:

Object of class "call": the estimate(s) for which the confidence intervals are produced.

name.estimate:

Object of class "character": the name of the estimate(s) for which the confidence intervals are produced.

trafo.estimate:

Object of class "matrix": the trafo/derivative matrix of the estimate(s) for which the confidence intervals are produced.

nuisance.estimate:

Object of class "OptionalNumeric": the nuisance parameter (if any) at which the confidence intervals are produced.

Methods

type

signature(object = "Confint"): accessor function for slot type.

confint

signature(object = "Confint"): accessor function for slot type.

estimate.call

signature(object = "Confint"): accessor function for slot estimate.call.

name.estimate

signature(object = "Confint"): accessor function for slot name.estimate.

trafo.estimate

signature(object = "Confint"): accessor function for slot trafo.estimate.

nuisance.estimate

signature(object = "Confint"): accessor function for slot nuisance.estimate.

show

signature(object = "Confint"): shows a detailed view of the object; slot nuisance.estimate is only shown if non-null, and slot trafo.estimate only if different from a unit matrix.

print

signature(object = "Confint"): just as show, but with additional arguments digits.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

Estimator, confint, Estimate-class, trafo-methods

[Package distrMod version 2.0 Index]