Distribution-class(distr)	R Documentation

Class 'Distribution'

Description

The Distribution-class is the mother-class of class UnivariateDistribution.

Objects from the Class

Objects can be created by calls of the form new("Distribution").

Slots

img:

Object of class "rSpace": the space of the image

param:

Object of class "OptionalParameter": the parameter

r:

Object of class "function": generates random numbers

d:

Object of class "OptionalFunction": density function

p:

Object of class "OptionalFunction": cumulative distribution function

q:

Object of class "OptionalFunction": quantile function

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Symmetry:

object of class "DistributionSymmetry"; used internally to avoid unnecessary calculations.

Methods

img

signature(object = "Distribution"): returns the space of the image

param

signature(object = "Distribution"): returns the parameter

r

signature(object = "Distribution"): returns the random number generator

d

signature(object = "Distribution"): returns the density function

p

signature(object = "Distribution"): returns the cumulative distribution function

q

signature(object = "Distribution"): returns the quantile function

.logExact

signature(object = "Distribution"): returns slot .logExact if existing; else tries to convert the object to a newer version of its class by conv2NewVersion and returns the corresponding slot of the converted object.

.lowerExact

signature(object = "Distribution"): returns slot .lowerExact if existing; else tries to convert the object to a newer version of its class by conv2NewVersion and returns the corresponding slot of the converted object.

Symmetry:

returns slot Symmetry if existing; else tries to convert the object to a newer version of its class by conv2NewVersion and returns the corresponding slot of the converted object.

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

UnivariateDistribution-class Parameter-class

[Package distr version 2.2 Index]