KolmogorovDist(distrEx)	R Documentation

Generic function for the computation of the Kolmogorov distance of two distributions

Description

Generic function for the computation of the Kolmogorov distance d_k of two distributions P and Q where the distributions are defined on a finite-dimensional Euclidean space (R^m, B^m) with B^m the Borel-sigma-algebra on R^m. The Kolmogorov distance is defined as

d_k(P,Q)=sup{|P({y in R^m | y <= x})-Q({y in R^m | y <= x})| | x in R^m}

where <= is coordinatewise on R^m.

Usage


KolmogorovDist(e1, e2, ...)
## S4 method for signature 'AbscontDistribution,
##   AbscontDistribution':
KolmogorovDist(e1,e2)
## S4 method for signature 'AbscontDistribution,
##   DiscreteDistribution':
KolmogorovDist(e1,e2)
## S4 method for signature 'DiscreteDistribution,
##   AbscontDistribution':
KolmogorovDist(e1,e2)
## S4 method for signature 'DiscreteDistribution,
##   DiscreteDistribution':
KolmogorovDist(e1,e2)

Arguments

e1 	 object of class "Distribution"

e2 	 object of class "Distribution"

... 	 further arguments to be used in particular methods (not in package distrEx)

Value

A list containing the following components:

e1 	 object of class "Distribution"; distribution 1

e2 	 object of class "Distribution"; distribution 2

Kolmogorov.distance 	 Kolmogorov distance of e1 and e2

Methods

e1 = "AbscontDistribution", e2 = "AbscontDistribution":

Kolmogorov distance of two absolutely continuous univariate distributions which is computed using a union of a (pseudo-)random and a deterministic grid.

e1 = "DiscreteDistribution", e2 = "DiscreteDistribution":

Kolmogorov distance of two discrete univariate distributions. The distance is attained at some point of the union of the supports of e1 and e2.

e1 = "AbscontDistribution", e2 = "DiscreteDistribution":

Kolmogorov distance of absolutely continuous and discrete univariate distributions. It is computed using a union of a (pseudo-)random and a deterministic grid in combination with the support of e2.

e1 = "DiscreteDistribution", e2 = "AbscontDistribution":

Kolmogorov distance of discrete and absolutely continuous univariate distributions. It is computed using a union of a (pseudo-)random and a deterministic grid in combination with the support of e1.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1981) Robust Statistics. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

See Also

ContaminationSize, TotalVarDist, HellingerDist, Distribution-class

Examples


KolmogorovDist(Norm(), Gumbel())
KolmogorovDist(Norm(), Td(10))
KolmogorovDist(Norm(mean = 50, sd = sqrt(25)), Binom(size = 100))
KolmogorovDist(Pois(10), Binom(size = 20)) 

[Package distrEx version 1.9 Index]