L2ParamFamily-class(distrMod)	R Documentation

L2 differentiable parametric family

Description

Class of L2 differentiable parametric families.

Objects from the Class

Objects can be created by calls of the form new("L2ParamFamily", ...). More frequently they are created via the generating function L2ParamFamily.

Slots

name:

[inherited from class "ProbFamily"] object of class "character": name of the family.

distribution:

[inherited from class "ProbFamily"] object of class "Distribution": member of the family.

distrSymm:

[inherited from class "ProbFamily"] object of class "DistributionSymmetry": symmetry of distribution.

param:

[inherited from class "ParamFamily"] object of class "ParamFamParameter": parameter of the family.

fam.call:

[inherited from class "ParamFamily"] object of class "call": call by which parametric family was produced.

makeOKPar:

[inherited from class "ParamFamily"] object of class "function": has argument param — the (total) parameter, returns valid parameter; used if optim resp. optimize— try to use ``illegal'' parameter values; then makeOKPar makes a valid parameter value out of the illegal one.

startPar:

[inherited from class "ParamFamily"] object of class "function": has argument x — the data, returns starting parameter for optim resp. optimize— a starting estimator in case parameter is multivariate or a search interval in case parameter is univariate.

modifyParam:

[inherited from class "ParamFamily"] object of class "function": mapping from the parameter space (represented by "param") to the distribution space (represented by "distribution").

props:

[inherited from class "ProbFamily"] object of class "character": properties of the family.

L2deriv:

object of class "EuclRandVariable": L2 derivative of the family.

L2deriv.fct:

object of class "function": mapping from the parameter space (argument param of class "ParamFamParameter") to a mapping from observation x to the value of the L2derivative; L2deriv.fct is then used from observation x to value of the L2derivative; L2deriv.fct is used by modifyModel to move the L2deriv according to a change in the parameter

L2derivSymm:

[ object of class "FunSymmList": symmetry of the maps included in L2deriv.

L2derivDistr:

object of class "UnivarDistrList": list which includes the distribution of L2deriv.

L2derivDistrSymm:

object of class "DistrSymmList": symmetry of the distributions included in L2derivDistr.

FisherInfo.fct:

object of class "function": mapping from the parameter space (argument param of class "ParamFamParameter") to the set of positive semidefinite matrices; FisherInfo.fct is used by modifyModel to move the Fisher information according to a change in the parameter

FisherInfo:

object of class "PosDefSymmMatrix": Fisher information of the family.

Extends

Class "ParamFamily", directly.
Class "ProbFamily", by class "ParamFamily".

Methods

L2deriv

signature(object = "L2ParamFamily"): accessor function for L2deriv.

L2deriv

signature(object = "L2ParamFamily", param = "ParamFamParameter"): returns the L2derivative at param, i.e. evaluates slot function L2deriv.fct at param.

L2derivSymm

signature(object = "L2ParamFamily"): accessor function for L2derivSymm.

L2derivDistr

signature(object = "L2ParamFamily"): accessor function for L2derivDistr.

L2derivDistrSymm

signature(object = "L2ParamFamily"): accessor function for L2derivDistrSymm.

FisherInfo

signature(object = "L2ParamFamily"): accessor function for FisherInfo.

FisherInfo

signature(object = "L2ParamFamily", param = "ParamFamParameter"): returns the Fisher Information at param, i.e. evaluates slot function FisherInfo.fct at param.

checkL2deriv

signature(object = "L2ParamFamily"): check centering of L2deriv and compute precision of Fisher information.

E

signature(object = "L2ParamFamily", fun = "EuclRandVariable", cond = "missing"): expectation of fun under the distribution of object.

E

signature(object = "L2ParamFamily", fun = "EuclRandMatrix", cond = "missing"): expectation of fun under the distribution of object.

E

signature(object = "L2ParamFamily", fun = "EuclRandVarList", cond = "missing"): expectation of fun under the distribution of object.

plot

signature(x = "L2ParamFamily"): plot of distribution and L2deriv. More precisely, this method has arguments plot(x, withSweave = getdistrOption("withSweave"), main = FALSE, inner = TRUE, sub = FALSE, col.inner = par("col.main"), cex.inner = 0.8, bmar = par("mar")[1], tmar = par("mar")[3], ..., mfColRow = TRUE, to.draw.arg = NULL) where

x

object of class "L2ParamFamily"

withSweave

logical: if TRUE (for working with Sweave) no extra device is opened and height/width are not set

main

logical: is a main title to be used? or
just as argument main in plot.default.

inner

logical: do panels for density/probability function - cdf - quantile function have their own titles? or
list which is filled to length 4 if necessary (possibly using recycling rules): titles for density/probability function - cdf - quantile function and L2-derivative (each of the same form as argument main in plot.default)

inner

logical: do panels have their own titles? or
character vector of / cast to length 3+'number of plotted dimensions'; if argument to.draw.arg is used, this refers to a vector of length length(to.draw.arg), the actually plotted dimensions. For further information, see also description of argument main in plot.default.

sub

logical: is a sub-title to be used? or
just as argument sub in plot.default.

tmar

top margin – useful for non-standard main title sizes

bmar

bottom margin – useful for non-standard sub title sizes

cex.inner

magnification to be used for inner titles relative to the current setting of cex; as in par

col.inner

character or integer code; color for the inner title

mfColRow

shall default partition in panels be used — defaults to TRUE

to.draw.arg

if mfColRow==FALSE, either NULL (default; everything is plotted) or a vector of either integers (the indices of the subplots to be drawn) or characters — the names of the subplots to be drawn: these names are to be chosen among c("d","p","q", dimnms) where dimnms is either the row names of the trafo matrix rownames(trafo(x@param)) or if the last expression is NULL a vector "dim<dimnr>", dimnr running through the number of rows of the trafo matrix.

...

addtional arguments for plot — see plot, plot.default, plot.stepfun

if ... contains argument ylim, this may either be as in plot.default (i.e. a vector of length 2) or a vector of length 4, where the first two elements are the values for ylim in panels "d.c" and "d.d", and the last two elements are the values for ylim resp. xlim in panels "p", "p.c", "p.d" and "q", "q.c", "q.d".

modifyModel

signature(model = "L2ParamFamily", param = "ParamFamParameter"): moves the L2-parametric Family model to parameter param

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2ParamFamily, ParamFamily-class

Examples


F1 <- new("L2ParamFamily")
plot(F1)

## selection of subpanels for plotting
F2 <- L2LocationScaleFamily()
layout(matrix(c(1,2,3,3), nrow=2, byrow=TRUE))
plot(F2,mfColRow = FALSE,
     to.draw.arg=c("p","q","loc"))

[Package distrMod version 2.1 Index]