Length-methods(distr)	R Documentation

Methods for Function Length in Package ‘distr’

Description

Length-methods

Methods

Length

signature(object = "Lattice"): returns the slot Length of the lattice

Length<-

signature(object = "Lattice"): modifies the slot Length of the lattice

Length

signature(object = "LatticeDistribution"): returns the slot Length of the lattice slot of the distribution

Length<-

signature(object = "LatticeDistribution"): modifies the slot Length of the lattice slot of the distribution

[Package distr version 2.1 Index]