MultivarDistrList-class(distrEllipse)	R Documentation

List of multivariate distributions

Description

Create a list of multivariate distributions

Objects from the Class

Objects can be created by calls of the form new("MultivarDistrList", ...). More frequently they are created via the generating function MultivarDistrList.

Slots

.Data:

Object of class "list". A list of multivariate distributions of the same dimension.

Extends

Class "DistrList", directly.
Class "list", by class "DistrList".
Class "vector", by class "DistrList".

Methods

coerce

signature(from = "MultivariateDistribution", to = "MultivarDistrList"): create a MultivarDistrList object from a univariate distribution

dim

returns the dimension of the range space.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

MultivarDistrList, DistrList-class, MultivariateDistribution-class

Examples


(DL <- new("MultivarDistrList", list(Norm(), Exp())))
plot(DL)
as(Norm(), "MultivarDistrList")

[Package distrEllipse version 0.1 Index]