ParamFamily-class(distrMod)	R Documentation

Parametric family of probability measures.

Description

Class of parametric families of probability measures.

Objects from the Class

Objects can be created by calls of the form new("ParamFamily", ...). More frequently they are created via the generating function ParamFamily.

Slots

name:

[inherited from class "ProbFamily"] object of class "character": name of the family.

distribution:

[inherited from class "ProbFamily"] object of class "Distribution": member of the family.

distrSymm:

[inherited from class "ProbFamily"] object of class "DistributionSymmetry": symmetry of distribution.

param:

object of class "ParamFamParameter": parameter of the family.

fam.call:

object of class "call": call by which parametric family was produced.

makeOKPar:

object of class "function": has argument param — the (total) parameter, returns valid parameter; used if optim resp. optimize— try to use ``illegal'' parameter values; then makeOKPar makes a valid parameter value out of the illegal one.

startPar:

object of class "function": has argument x — the data, returns starting parameter for optim resp. optimize— a starting estimator in case parameter is multivariate or a search interval in case parameter is univariate.

modifyParam:

object of class "function": mapping from the parameter space (represented by "param") to the distribution space (represented by "distribution").

props:

[inherited from class "ProbFamily"] object of class "character": properties of the family.

Extends

Class "ProbFamily", directly.

Methods

main

signature(object = "ParamFamily"): wrapped accessor function for slot main of slot param.

nuisance

signature(object = "ParamFamily"): wrapped accessor function for slot nuisance of slot param.

trafo

signature(object = "ParamFamily", param = "missing"): wrapped accessor function for slot trafo of slot param.

param

signature(object = "ParamFamily"): accessor function for slot param.

modifyParam

signature(object = "ParamFamily"): accessor function for slot modifyParam.

fam.call

signature(object = "ParamFamily"): accessor function for slot fam.call.

plot

signature(x = "ParamFamily"): plot of slot distribution.

show

signature(object = "ParamFamily")

Details for methods this-is-escaped-codenormal-bracket116bracket-normal, this-is-escaped-codenormal-bracket117bracket-normal

Detailedness of output by methods show, print is controlled by the global option show.details to be set by distrModoptions.

As method show is used when inspecting an object by typing the object's name into the console, show comes without extra arguments and hence detailedness must be controlled by global options.

Method print may be called with a (partially matched) argument show.details, and then the global option is temporarily set to this value.

For class ParamFamily, this becomes relevant for slot param. For details therefore confer to ParamFamParameter-class.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Distribution-class

Examples


F1 <- new("ParamFamily") # prototype
plot(F1)

[Package distrMod version 2.0 Index]