SweaveListingOptions(SweaveListingUtils)	R Documentation

Function to change the global options of the package ‘SweaveListingUtils’

Description

With SweaveListingOptions you can inspect and change the global variables of the package SweaveListingUtils.

Usage


SweaveListingOptions(...)
SweaveListingoptions(...)
getSweaveListingOption(x)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Value

SweaveListingOptions() returns a list of the global variables.
SweaveListingOptions(x) returns the global variable x.
getSweaveListingOption(x) returns the global variable x.
SweaveListingOptions(x=y) sets the value of the global variable x to y.

SweaveListingoptions

For compatibility with spelling in package distr, SweaveListingoptions is just a synonym to SweaveListingOptions.

Global Options

Rset:

default setting for \lstset-definition of R-code, default value is
list("fancyvrb" = "true", "language" = "R", "escapechar" = "`",
"basicstyle" = "{\\color{Rcolor}\\footnotesize}",
"keywordstyle" = "{\\bf\\color{Rcolor}}",
"commentstyle" = "{\\color{Rcomment}\\ttfamily\\itshape}",
"literate" = "{<-}{{$\\leftarrow$}}2{<<-}{{$\\twoheadleftarrow$}}2",
"alsoother" = "{$}","alsoletter" = "{.<-}"
"otherkeywords" = "{!,!=,~,$,*,\\&,\\ %/\\ %,\\ %*\\ %,\\ %\\ %,<-,<<-,/}")
(CAVEAT: the spaces between the double backslashes and the percent sign in the last line are not intended; I simply did not find another work-around)

Rdset:

default setting for \lstset-definition of Rd-code, default value is
list("fancyvrb" = "true", "language" = "Rd", "basicstyle" = "{\\color{black}\\tiny}",
"keywordstyle" = "{\\bf}",
"commentstyle" = "\\ttfamily\\itshape",
"alsolanguage" = "R")

Rcolor:

default setting for color of R-code in rgb-coordinates; defaults to c(0,0.5,0.5)

Rbcolor:

default setting for color of R-code symbols of intermediate packages in rgb-coordinates; defaults to c(0,0.4,0.4)

Rout:

default setting for color of R-output in rgb-coordinates; defaults to c(0.461,0.039,0.102)

Rcomment:

default setting for color of R-comments in rgb-coordinates; defaults to c(0.101,0.043,0.432)

pkv:

default setting for package version as character; defaults to "2.0.2"

inSweave:

logical; default setting for masked functions library and require; shall they show Sweave-behaviour? defaults to FALSE

Keywordstyle:

default setting for R-symbols as character; defaults to "{\\bf}"

interm.Keywordstyle:

default setting for package name as character; defaults to "{\\bf\\color{Rbcolor}}"

intermediate:

default setting: shall symbols from automatically loaded intermediate packages also be registered for listings-printing? defaults to TRUE

overwrite:

default setting: shall R symbols already defined in the original R language definition in ‘lstlang3.sty’ be overwritten if they reappear in a required package (to be printed in a different keywordstyle)? defaults to FALSE

fromRForge:

default setting: logical: shall code sniplets be drawn from "r-forge.r-project.org" (or a mirror) or not; defaults to TRUE.

base.url:

default setting for the base url to download code sniplets from (character); defaults to "http://r-forge.r-project.org/plugins/scmsvn/viewcvs.php/*checkout*/pkg/"

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

options, getOption

Examples


SweaveListingOptions()
SweaveListingOptions("Rout")
SweaveListingOptions("Rout" = c(0,0,0))
# or
SweaveListingOptions(Rout = c(0,1,0))
getSweaveListingOption("Rout")

[Package SweaveListingUtils version 0.1.1 Index]