UnivariateDistribution-class(distr)	R Documentation

Class 'UnivariateDistribution'

Description

The UnivariateDistribution-class is the mother-class of the classes AbscontDistribution and DiscreteDistribution.

Objects from the Class

Objects can be created by calls of the form new("UnivariateDistribution").

Slots

img:

Object of class "Reals": the space of the image of this distribution which has dimension 1 and the name "Real Space"

param:

Object of class "Parameter": the parameter of this distribution

r:

Object of class "function": generates random numbers

d:

Object of class "function": density function

p:

Object of class "function": cumulative distribution function

q:

Object of class "function": quantile function

Extends

Class "Distribution", directly.

Methods

initialize

signature(.Object = "UnivariateDistribution"):
initialize method

dim

signature(x = "UnivariateDistribution"):
returns the dimension of the support of the distribution

-

signature(e1 = "UnivariateDistribution"):
application of `-' to this univariate distribution

*

signature(e1 = "UnivariateDistribution", e2 = "numeric"):
multiplication of this univariate distribution by an object of class `numeric'

/

signature(e1 = "UnivariateDistribution", e2 = "numeric"):
division of this univariate distribution by an object of class `numeric'

+

signature(e1 = "UnivariateDistribution", e2 = "numeric"):
addition of this univariate distribution to an object of class `numeric'

-

signature(e1 = "UnivariateDistribution", e2 = "numeric"):
subtraction of an object of class `numeric' from this univariate distribution

*

signature(e1 = "numeric", e2 = "UnivariateDistribution"):
multiplication of this univariate distribution by an object of class `numeric'

+

signature(e1 = "numeric", e2 = "UnivariateDistribution"):
addition of this univariate distribution to an object of class `numeric'

-

signature(e1 = "numeric", e2 = "UnivariateDistribution"):
subtraction of this univariate distribution from an object of class `numeric'

+

signature(e1 = "UnivariateDistribution", e2 = "UnivariateDistribution"):
Convolution of two univariate distributions. The slots p, d and q are approximated by grids.

-

signature(e1 = "UnivariateDistribution", e2 = "UnivariateDistribution"):
Convolution of two univariate distributions. The slots p, d and q are approximated by grids.

simplifyr

signature(object = "UnivariateDistribution"):
simplifies the r-method of a distribution, see there for further information

print

signature(object = "UnivariateDistribution"):
returns the class of the object and its parameters

show

signature(object = "UnivariateDistribution"):
as print

Author(s)

Thomas Stabla statho3@web.de,
Florian Camphausen fcampi@gmx.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,
Matthias Kohl Matthias.Kohl@stamats.de

See Also

Parameter-class Distribution-class AbscontDistribution-class
DiscreteDistribution-class Reals-class RtoDPQ simplifyr-methods

[Package distr version 2.0 Index]