distrModOptions(distrMod)	R Documentation

Function to change the global variables of the package ‘distrMod’

Description

With distrModOptions you can inspect and change the global variables of the package distrMod.

Usage


distrModOptions(...)
getdistrModOption(x)
distrModoptions(...)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Details

Invoking distrModoptions() with no arguments returns a list with the current values of the options. To access the value of a single option, one should use getdistrModOption("show.details"), e.g., rather than distrModoptions("show.details") which is a list of length one.

Value

distrModoptions() returns a list of the global options of distrMod.
distrModoptions("show.details") returns the global option show.details as a list of length 1.
distrModoptions("show.details" = "minimal") sets the value of the global option show.details to "minimal". getdistrModOption("show.details") the current value set for option show.details.

distrModoptions

For compatibility with spelling in package distr, distrModoptions is just a synonym to distrModoptions.

Currently available options

show.details

degree of detailedness for method show for objects of classes of the distrXXX family of packages. Possible values are

"maximal"

all information is shown

"minimal"

only the most important information is shown

"medium"

somewhere in the middle; see actual show-methods for details.

The default value is "maximal".

use.generalized.inverse.by.default

logical; decides whether by default (i.e., if argument generalized of solve is not explicitely set), solve is to use generalized inverses if the original solve-method from package base fails; if the option is FALSE, in case of failure, and unless argument generalized is not explicitely set to TRUE, solve will throw an error as is the base-method behavior. The default value is TRUE.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

options, getOption, distroptions, getdistrOption

Examples


distrModoptions()
distrModoptions("show.details")
distrModoptions("show.details" = "maximal")
distrModOptions("show.details" = "minimal")
# or
getdistrModOption("show.details")

[Package distrMod version 2.0 Index]