InternalClassUnions(distrMod)	R Documentation

Class unions in 'distrMod'

Description

Class unions in package distrMod defined for internal purposes; these are OptionalNumeric, OptionalMatrix, MatrixorFunction, ShowDetails

Details

These classes are used internally to make available methods or to allow slots of classes to be filled with varying types. In particular

"OptionalNumeric"

may contain objects of class "numeric" or "NULL"; it is used e.g. for slot nuisance of class "ParamFamParameter", as it may or may not be present but if so it has to be numeric.

"OptionalNumericOrMatrix"

may contain objects of class "OptionalNumeric" or "matrix"; it is used e.g. for slot asvar of class "Estimate", as it may or may not be present but if so it has to be a number (numeric) or a matrix.

"MatrixorFunction"

may contain objects of class "OptionalFunction" or "matrix"; it is used e.g. for slot trafo of class "ParamFamParameter", as it may or may not be present and if it is present, it has to either be a function or a matrix, see link{trafo-methods}.

"ShowDetails"

may contain objects of class "Estimate", "MCEstimate", "Confint", "PosSemDefSymmMatrix" "ParamFamily", or "ParamFamParameter"; used to provide sort of a ``show with extra arguments'', in form of a common print method for these S4 classes, which essentially just temporarily sets the global options according to the optional arguments digits and show.details, calls show and then re-sets the options to their global settings.

"L2LocationScaleUnion"

is a proper, but virtual in-between class between class "L2GroupParamFamily" and "L2LocationFamily", "L2ScaleFamily", and "L2LocationScaleFamily"; in addition to class "L2GroupParamFamily" it has a slot locscalename (with corresponding accessor and replacement method) to capture the fact that location and scale parameter may carry names other than "loc" resp. "scale"; used to provide a common class for (parts of) methods modifyModel locscalename and locscalename<-.

Objects from the Class

All of these classes are virtual: No objects may be created from them.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

numeric-class, matrix-class, function-class, MCEstimate-class, Estimate-class, Confint-class, PosSemDefSymmMatrix-class, L2GroupParamFamily-class, L2LocationFamily-class, L2ScaleFamily-class, L2LocationScaleFamily-class, ParamFamParameter-class ParamFamily-class

[Package distrMod version 2.1 Index]