internals_for_distrMod(distrMod)	R Documentation

Internal functions of package distrMod

Description

These functions are used internally by package ``distrMod''.

Usage


.inArgs(arg, fct)
.isUnitMatrix(m)
.csimpsum(fx)
.validTrafo(trafo, dimension)
.CvMMDCovariance(L2Fam, param, mu = distribution(L2Fam),  
                 withplot = FALSE, withpreIC = FALSE,
                 N = getdistrOption("DefaultNrGridPoints")+1,
                 rel.tol=.Machine$double.eps^0.3, 
                 TruncQuantile = getdistrOption("TruncQuantile"), 
                 IQR.fac = 15, ...)
.show.with.sd(est, s)
.getLogDeriv(distr, 
             lowerTruncQuantile = getdistrExOption("ElowerTruncQuantile"), 
             upperTruncQuantile = getdistrExOption("EupperTruncQuantile"), 
                         IQR.fac = getdistrExOption("IQR.fac"))

Arguments

arg 	 a formal argument as character

fct 	 a function

m 	 a matrix

est 	 an estimator; usually a vector

s 	 a standard deviation

trafo 	 an object of class MatrixorFunction

dimension 	 a numeric

L2Fam 	 an object of class L2ParamFamily — for which we want to determine the IC resp. the as. [co]variance of the corresponding Minimum CvM estimator

param 	 an object of class ParamFamParameter, the parameter value at which we want to determine the IC resp. the as. [co]variance of the corresponding Minimum CvM estimator

mu 	 an object of class UnivariateDistribution: integration measure (resp. distribution) for CvM distance

rel.tol 	 relative tolerance for distrExIntegrate.

TruncQuantile 	 quantile for quantile based integration range.

lowerTruncQuantile 	 lower quantile for quantile based integration range.

upperTruncQuantile 	 upper quantile for quantile based integration range.

IQR.fac 	 factor for scale based integration range (i.e.; median of the distribution +-IQR.fac*IQR).

withplot 	 logical: shall we plot corresponding ICs?

withpreIC 	 logical: shall we return a list with components preIC and var or just var; here var is the corresponding asymptotic variance and preIC the corresponding EuclRandVarList featuring as argument Curve in ICs of package RobAStBase

N 	 a numeric: the number of gridpoints for constructing the mu- resp. P_theta-``primitive'' function

fx 	 a vector of function evaluations multiplied by the gridwidth

distr 	 an object of class AbscontDistribution

... 	 further argument to be passed through — so .CvMMDCovariance can digest more arguments

Details

.inArgs (borrowed from package distr) checks whether an argument arg is a formal argument of fct — not vectorized.

.csimpsum (borrowed from package distr) produces a primitimive function out of function evaluations by means of vectorized Simpson quadrature method, returning already the function values of the prime function on a grid; it is to mimick the behaviour of cumsum.

.isUnitMatrix checks whether the argument is a unit matrix.

.validTrafo checks whether the argument is a valid transformation.

.CvMMDCovariance determines the IC resp. the as. [co]variance of the corresponding Minimum CvM estimator. Still some checking / optimization / improvement needed.

.show.with.sd is code borrowed from print.fitdistr in package MASS by B.D. Ripley. It pretty-prints estimates with corresponding sd's below.

.getLogDeriv determines numerically the negative logarithmic derivative of the density of distribution distr; to this end uses D1ss, D2ss from Martin Maechler's package sfsmisc.

Value

.getLogderiv 	 a function in one argument x — the negative logarithmic derivative of the density

.inArgs 	 logical (length 1)

.csimpsum 	 numeric (of length half the input length)

.isUnitMatrix 	 logical (length 1)

.validTrafo 	 logical (length 1)

.CvMMDCovariance 	 corresponding as. [co]variance of the corresponding Minimum CvM estimator or list withcomponents preIC and var —see above

.show.with.sd 	 invisible()

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de Matthias Kohl Matthias.Kohl@stamats.de

See Also

MLEstimator, Estimate-class, MCEstimate-class, Confint-class, ParamFamParameter-class

[Package distrMod version 2.1 Index]