internals_for_distrMod(distrMod)	R Documentation

Internal functions of package distrMod

Description

These functions are used internally by package ``distrMod''.

Usage


.inArgs(arg, fct)
.isUnitMatrix(m)
.validTrafo(trafo, dimension)
.CvMMDCovariance(L2Fam, param, mu = distribution(L2Fam), expon = 3, 
                 withplot = FALSE, N = 200,
                 ...)

Arguments

arg 	 a formal argument as character

fct 	 a function

m 	 a matrix

trafo 	 an object of class MatrixorFunction

dimension 	 a numeric

L2Fam 	 an object of class L2ParamFamily — for which we want to determine the IC resp. the as. [co]variance of the corresponding Minimum CvM estimator

param 	 an object of class ParamFamParameter, the parameter value at which we want to determine the IC resp. the as. [co]variance of the corresponding Minimum CvM estimator

mu 	 an object of class UnivariateDistribution: integration measure (resp. distribution) for CvM distance

expon 	 a numeric: number to exponentiate getdistrOption("TruncQuantile") to get upper and lower mu- resp. P_theta-quantiles for the gridpoints (see below)

withplot 	 logical: shall we plot corresponding ICs?

N 	 a numeric: the number of gridpoints for constructing the mu- resp. P_theta-``primitive'' function

... 	 further argument to be passed through — so .CvMMDCovariance can digest more arguments

Details

.inArgs checks whether an argument arg is a formal argument of fct — not vectorized.

.isUnitMatrix checks whether the argument is a unit matrix.

.validTrafo checks whether the argument is a valid transformation.

.CvMMDCovariance determines the IC resp. the as. [co]variance of the corresponding Minimum CvM estimator. Still some checking / optimization / improvement needed.

Value

.inArgs 	 logical (length 1)

.isUnitMatrix 	 logical (length 1)

.validTrafo 	 logical (length 1)

.CvMMDCovariance 	 corresponding as. [co]variance of the corresponding Minimum CvM estimator

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de Matthias Kohl Matthias.Kohl@stamats.de

See Also

MLEstimator, Estimate-class, MCEstimate-class, Confint-class, ParamFamParameter-class

[Package distrMod version 2.0 Index]