liesInSupport(distrEx)	R Documentation

Generic Function for Testing the Support of a Distribution

Description

The function tests if x lies in the support of the distribution object.

Usage


## S4 method for signature 'DiscreteMVDistribution,
##   numeric':
liesInSupport(object, x)
## S4 method for signature 'DiscreteMVDistribution,
##   matrix':
liesInSupport(object, x)

Arguments

object 	 object of class "Distribution"

x 	 numeric vector or matrix

Value

logical vector

Methods

object = "DiscreteMVDistribution", x = "numeric":

does x lie in the support of object.

object = "DiscreteMVDistribution", x = "matrix":

does x lie in the support of object.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Distribution-class

Examples


M <- matrix(rpois(30, lambda = 10), ncol = 3)
D1 <- DiscreteMVDistribution(M)
M1 <- rbind(r(D1)(10), matrix(rpois(30, lam = 10), ncol = 3))
liesInSupport(D1, M1)

[Package distrEx version 2.2 Index]