lstset(SweaveListingUtils)	R Documentation

lstset and friends

Description

Functions for defining how listings prints R and Rd source code

Usage


lstset(taglist, LineLength = 80, startS = "\\lstset{")
lstsetR(Rset = NULL, LineLength = 80, add = TRUE, startS = "\\lstset{")
lstsetRd(Rdset = NULL, LineLength = 80, add = TRUE, startS = "\\lstset{")

Arguments

LineLength 	 numeric number of characters per line for lstset and friends; defaults to 80

taglist 	 S3-object of class taglist; arguments for \lstset of TeX package listings.

Rset 	 object of S3-class taglist or named list of characters; the listings settings for R-code (for \lstset); defaults to NULL.

Rdset 	 object of S3-class taglist or named list of characters; the listings settings for Rd-code (for \lstset); defaults to NULL.

add 	 boolean; defaults to TRUE; if TRUE, argument list Rset resp. Rdset will be appended to default value lists getSweaveListingOption("Rset") resp. getSweaveListingOption("Rdset") (see below), overwriting respective entries of the default value lists.

startS 	 character; defaults to "\\lstset{"; what to do by default we use \lstset; an alternative is to use \lstdefinestyle which amounts to "\\lstdefinestyle{".

Details

lstset writes out to stdout a call to TeX command \lstset{arg1 = val1, arg2 = val2, .....} and doing so respects a maximal number of characters per line and does not break arg=val tags.

lstsetR and lstsetRd expect either objects of S3 class taglist, or lists of named characters as first arguments, which in the latter case are then converted to taglist; both lstsetR and lstsetRd use particular default values to define R resp. Rd output format. More specifically for R code, it uses getSweaveListingOption("Rset"), and for Rd code, it getSweaveListingOption("Rdset"),

The output to stdout can be captured in an ‘.Rnw’ file as

<< lstsetR, results=tex, echo=FALSE>>=
lstsetR()
@

to insert the corresponding \lstset command to the produced TeX file.

Value

invisible()

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

Examples


lstset(taglist(A="H", b=2, 3),30)
lstset(taglist(A="H", b=2, 3),30, startS = "\\lstdefinestyle{Rstyle}{")
lstsetR()
lstsetRd()

[Package SweaveListingUtils version 0.2 Index]