meanlog-methods(distr)	R Documentation

Methods for Function meanlog in Package 'distr'

Description

meanlog-methods

Methods

meanlog

signature(object = "LnormParameter"): returns the slot meanlog of the parameter of the distribution

meanlog<-

signature(object = "LnormParameter"): modifies the slot meanlog of the parameter of the distribution

meanlog

signature(object = "Lnorm"): returns the slot meanlog of the parameter of the distribution

meanlog<-

signature(object = "Lnorm"): modifies the slot meanlog of the parameter of the distribution

[Package distr version 1.9 Index]