plot-methods(distrTEst)	R Documentation

Methods for Function plot in Package ‘distrTEst’

Description

plot-methods

Methods

plot

signature(x = "Evaluation", y="missing"): returns a boxplot of the result

plot

signature(x = "EvaluationList", y="missing"): regroups the list according to the different columns/coordinates of the result of the evaluation; for each such coordinate a boxplot is generated containing possibly several procedures and if evaluated at a Contsimulation object also grouped into evaluations on ideal and real data. The plot-arguments main and ylim may be transmitted coordinatewise (i.e.; a vector of length (result-dimension), respectively a 2 x (result-dimension) matrix) or globally, using the usual recycling rules.

[Package distrTEst version 2.1 Index]