print-methods(distrSim)	R Documentation

Methods for Function print/show in Package ‘distrSim’

Description

print and show - methods

Methods

print

signature(x = "SeqDataFrames"): extra argument short = FALSE; if TRUE the output is cut according to distrSimoptions.

print

signature(x = "Dataclass"): returns name, filename, the observation dimension, the number of runs and the size of the sample, and, if from a version > 1.8, also the package version under which the object was generated

print

signature(x = "Simulation"): as for signature(x = "Dataclass") and, additionally, seed and the distribution

print

signature(x = "Contsimulation"): as for signature(x = "Dataclass") and, additionally, seed, the contamination rate and the distributions

show

signature(x = "Dataclass"), signature(x = "Simulation"), signature(x = "Contsimulation"), \code{signature(x = "SeqDataFrames")}: as corresponding print method

[Package distrSim version 2.1 Index]