print-methods(distrTEst)	R Documentation

Methods for Functions print and show in Package ‘distrTEst’

Description

print/show-methods

Methods

print

signature(x = "Evaluation"): returns the name of the data object, its filename, the estimator used and the result; optional arguments:

runs0

the indices of runs to be summarized;

dims0

the indices of result dimensions to be summarized;

internal argument:

inList

decides if name of Dataobject and Datafile are printed out (which is done if inLIST==FALSE); defaults to FALSE but is TRUE when summary is called from summary-method for signature(object = "EvaluationList")

print

signature(object = "EvaluationList"): after printing the name of the list, for each member of the list print is executed; optional arguments:

eval0

the indices of evaluations to be summarized;— of this vector eval0 maximally MaxNumberofSummarizedEvaluations evaluations are summarized where MaxNumberofPrintedEvaluations is a global option, see distrTEstoptions

runs0

the indices of runs to be summarized;

dims0

the indices of observation dimensions to be summarized;

show

signature(x = "Evaluation"): the same as print (without optional arguments)

show

signature(x = "EvaluationList"): the same as print (without optional arguments)

[Package distrTEst version 2.0 Index]