savedata-methods(distrSim)	R Documentation

Methods for Function savedata in Package ‘distrSim’

Description

savedata-methods

Methods

savedata

signature(object = "Dataclass"): saves the object (with the data) in the directory of R

savedata

signature(object = "Simulation"): saves the object without the data in the directory of R (After loading the data can be reproduced by using simulate.)

savedata

signature(object = "Contsimulation"): saves the object without the data in the directory of R (After loading the data can be reproduced by using simulate.)

Note

For an example, see Simulation-class and Contsimulation-class

See Also

Dataclass-class Simulation-class Contsimulation-class Evaluation-class

[Package distrSim version 2.1 Index]