summary-methods(distrSim)	R Documentation

Methods for Function summary in Package ‘distrSim’

Description

summary-methods

Methods

summary

signature(object = "Dataclass"): returns the same information as print, moreover a statistical summary for each run; optional arguments:

runs0

the indices of runs to be summarized; — of this vector runs0 maximally MaxNumberofSummarizedRuns runs are summarized where MaxNumberofSummarizedRuns is a global option, see distrSimoptions

dims0

the indices of observation dimensions to be summarized; — of this vector dims0 maximally MaxNumberofSummarizedObsDims dimensions are summarized where MaxNumberofSummarizedObsDims is a global option, see distrSimoptions

summary

signature(object = "Simulation"): returns name, filename, seed, number of runs, the size of the sample and a statistical summary for each run; optional arguments: as with signature(object = "Dataclass")

summary

signature(object = "Contsimulation"): returns name, filename, seed, number of runs, the size of the sample, the rate and a statistical summary for each run of the real data; optional arguments: as with signature(object = "Dataclass")

[Package distrSim version 2.1 Index]